Make functions for getting the right svgs
Previously, there was a very big switch statement for finding the right SVG. This worked, but it is unsustainable as we add more different kinds of palettes and button kinds and locales. Now, we moved the logic for finding the SVG to the SVG struct itself.

I have also moved SVG sources to new file. Before, all the SVG sources were in the SVG.swift file. This was too much for that file. I want to separate out the long, bland SVG source from the other logic in the SVG struct. This will be more important as we add more logic to the SVG struct.
