+++ Sources/Afterpay/Resources/SVG.swift
@@ -14,7 +14,6 @@ import Macaw
 #endif
 import UIKit
 
-// swiftlint:disable line_length file_length type_body_length
 struct SVG: Equatable {
 
   let size: CGSize
@@ -30,831 +29,76 @@ struct SVG: Equatable {
     self.svgString = svgString
   }
 
-  static let afterpayBadgeBlackOnMint = SVG(
-    size: CGSize(width: 1582, height: 551),
-    minimumWidth: 64,
-    svgString: """
-    <svg width="1582" height="551" viewBox="0 0 1582 551" fill="none" xmlns="http://www.w3.org/2000/svg">
-    <g clip-path="url(#clip0)">
-    <path d="M1306.3 550.5H275.2C123.2 550.5 0 427.3 0 275.3C0 123.3 123.2 0.100006 275.2 0.100006H1306.3C1458.3 0.100006 1581.5 123.3 1581.5 275.3C1581.6 427.2 1458.3 550.5 1306.3 550.5Z" fill="#B2FCE4"/>
-    <path d="M1348 209.5L1313.4 189.7L1278.3 169.6C1255.1 156.3 1226.1 173 1226.1 199.8V204.3C1226.1 206.8 1227.4 209.1 1229.6 210.3L1245.9 219.6C1250.4 222.2 1256 218.9 1256 213.7V203C1256 197.7 1261.7 194.4 1266.3 197L1298.3 215.4L1330.2 233.7C1334.8 236.3 1334.8 243 1330.2 245.6L1298.3 263.9L1266.3 282.3C1261.7 284.9 1256 281.6 1256 276.3V271C1256 244.2 1227 227.4 1203.8 240.8L1168.7 260.9L1134.1 280.7C1110.8 294.1 1110.8 327.8 1134.1 341.2L1168.7 361L1203.8 381.1C1227 394.4 1256 377.7 1256 350.9V346.4C1256 343.9 1254.7 341.6 1252.5 340.4L1236.2 331.1C1231.7 328.5 1226.1 331.8 1226.1 337V347.7C1226.1 353 1220.4 356.3 1215.8 353.7L1183.8 335.3L1151.9 317C1147.3 314.4 1147.3 307.7 1151.9 305.1L1183.8 286.8L1215.8 268.4C1220.4 265.8 1226.1 269.1 1226.1 274.4V279.7C1226.1 306.5 1255.1 323.3 1278.3 309.9L1313.4 289.8L1348 270C1371.3 256.5 1371.3 222.9 1348 209.5Z" fill="black"/>
-    <path d="M1121 216.1L1040 383.4H1006.4L1036.7 320.9L989 216.1H1023.5L1054.1 286.3L1087.5 216.1H1121Z" fill="black"/>
-    <path d="M311.1 275.5C311.1 255.5 296.6 241.5 278.8 241.5C261 241.5 246.5 255.8 246.5 275.5C246.5 295 261 309.5 278.8 309.5C296.6 309.5 311.1 295.5 311.1 275.5ZM311.4 334.9V319.5C302.6 330.2 289.5 336.8 273.9 336.8C241.3 336.8 216.6 310.7 216.6 275.5C216.6 240.6 242.3 214 274.6 214C289.8 214 302.6 220.7 311.4 231.1V216.1H340.6V334.9H311.4V334.9Z" fill="black"/>
-    <path d="M482.6 308.5C472.4 308.5 469.5 304.7 469.5 294.7V242H488.3V216.1H469.5V187.1H439.6V216.1H401V204.3C401 194.3 404.8 190.5 415.3 190.5H421.9V167.5H407.5C382.8 167.5 371.1 175.6 371.1 200.3V216.2H354.5V242H371.1V334.9H401V242H439.6V300.2C439.6 324.4 448.9 334.9 473.1 334.9H488.5V308.5H482.6Z" fill="black"/>
-    <path d="M590 264.8C587.9 249.4 575.3 240.1 560.5 240.1C545.8 240.1 533.6 249.1 530.6 264.8H590ZM530.3 283.3C532.4 300.9 545 310.9 561 310.9C573.6 310.9 583.3 305 589 295.5H619.7C612.6 320.7 590 336.8 560.3 336.8C524.4 336.8 499.2 311.6 499.2 275.7C499.2 239.8 525.8 213.9 561 213.9C596.4 213.9 622.1 240 622.1 275.7C622.1 278.3 621.9 280.9 621.4 283.3H530.3V283.3Z" fill="black"/>
-    <path d="M812.5 275.5C812.5 256.3 798 241.5 780.2 241.5C762.4 241.5 747.9 255.8 747.9 275.5C747.9 295 762.4 309.5 780.2 309.5C798 309.5 812.5 294.8 812.5 275.5ZM718.4 383.4V216.1H747.6V231.5C756.4 220.6 769.5 213.9 785.1 213.9C817.2 213.9 842.4 240.3 842.4 275.2C842.4 310.1 816.7 336.7 784.4 336.7C769.4 336.7 757.1 330.8 748.5 320.8V383.3H718.4V383.4Z" fill="black"/>
-    <path d="M947.7 275.5C947.7 255.5 933.2 241.5 915.4 241.5C897.6 241.5 883.1 255.8 883.1 275.5C883.1 295 897.6 309.5 915.4 309.5C933.2 309.5 947.7 295.5 947.7 275.5ZM948 334.9V319.5C939.2 330.2 926.1 336.8 910.5 336.8C877.9 336.8 853.2 310.7 853.2 275.5C853.2 240.6 878.9 214 911.2 214C926.4 214 939.2 220.7 948 231.1V216.1H977.2V334.9H948V334.9Z" fill="black"/>
-    <path d="M665.7 227.7C665.7 227.7 673.1 213.9 691.4 213.9C699.2 213.9 704.2 216.6 704.2 216.6V246.9C704.2 246.9 693.2 240.1 683.1 241.5C673 242.9 666.6 252.1 666.6 264.5V334.8H636.4V216.1H665.6V227.7H665.7Z" fill="black"/>
-    </g>
-    <defs>
-    <clipPath id="clip0">
-    <rect width="1582" height="551" fill="white"/>
-    </clipPath>
-    </defs>
-    </svg>
-    """
-  )
+  static func badge(palette: ColorPalette, locale: Locale) -> SVG {
+    switch (palette, locale) {
+    case (.blackOnMint, Locales.greatBritain):
+      return .clearpayBadgeBlackOnMint
+    case (.mintOnBlack, Locales.greatBritain):
+      return .clearpayBadgeMintOnBlack
+    case (.whiteOnBlack, Locales.greatBritain):
+      return .clearpayBadgeWhiteOnBlack
+    case (.blackOnWhite, Locales.greatBritain):
+      return .clearpayBadgeBlackOnWhite
+    case (.blackOnMint, _):
+      return .afterpayBadgeBlackOnMint
+    case (.mintOnBlack, _):
+      return .afterpayBadgeMintOnBlack
+    case (.whiteOnBlack, _):
+      return .afterpayBadgeWhiteOnBlack
+    case (.blackOnWhite, _):
+      return .afterpayBadgeBlackOnWhite
+    }
+  }
 
-  static let clearpayBadgeBlackOnMint = SVG(
-    size: CGSize(width: 1620, height: 551),
-    minimumWidth: 64,
-    svgString: """
-    <svg width="1620" height="551" viewBox="0 0 1620 551" fill="none" xmlns="http://www.w3.org/2000/svg">
-    <path d="M1344 550.5H275.5C123.5 550.5 0.300049 427.3 0.300049 275.3C0.300049 123.3 123.5 0.100006 275.5 0.100006H1344C1496 0.100006 1619.2 123.3 1619.2 275.3C1619.3 427.2 1496 550.5 1344 550.5Z" fill="#B2FCE4"/>
-    <path d="M1158.7 216.2L1077.6 383.5H1044L1074.3 321L1026.7 216.2H1061.2L1091.8 286.4L1125.2 216.2H1158.7Z" fill="black"/>
-    <path d="M850.1 275.6C850.1 256.4 835.6 241.6 817.8 241.6C800 241.6 785.5 255.9 785.5 275.6C785.5 295.1 800 309.6 817.8 309.6C835.6 309.6 850.1 294.8 850.1 275.6ZM756 383.5V216.2H785.2V231.6C794 220.7 807.1 214 822.7 214C854.8 214 880 240.4 880 275.3C880 310.2 854.3 336.8 822 336.8C807 336.8 794.7 330.9 786.1 320.9V383.4H756V383.5Z" fill="black"/>
-    <path d="M985.4 275.6C985.4 255.6 970.9 241.6 953.1 241.6C935.3 241.6 920.8 255.9 920.8 275.6C920.8 295.1 935.3 309.6 953.1 309.6C970.9 309.6 985.4 295.5 985.4 275.6ZM985.6 335V319.6C976.8 330.3 963.7 336.9 948.1 336.9C915.5 336.9 890.8 310.8 890.8 275.6C890.8 240.7 916.5 214.1 948.8 214.1C964 214.1 976.8 220.8 985.6 231.2V216.2H1014.8V335H985.6Z" fill="black"/>
-    <path d="M703.3 227.8C703.3 227.8 710.7 214 729 214C736.8 214 741.8 216.7 741.8 216.7V247C741.8 247 730.8 240.2 720.7 241.6C710.6 243 704.2 252.2 704.2 264.6V335H674V216.2H703.2V227.8H703.3Z" fill="black"/>
-    <path d="M1385.5 209.5L1350.9 189.7L1315.8 169.6C1292.6 156.3 1263.6 173 1263.6 199.8V204.3C1263.6 206.8 1264.9 209.1 1267.1 210.3L1283.4 219.6C1287.9 222.2 1293.5 218.9 1293.5 213.7V203C1293.5 197.7 1299.2 194.4 1303.8 197L1335.8 215.4L1367.7 233.7C1372.3 236.3 1372.3 243 1367.7 245.6L1335.8 263.9L1303.8 282.3C1299.2 284.9 1293.5 281.6 1293.5 276.3V271C1293.5 244.2 1264.5 227.4 1241.3 240.8L1206.2 260.9L1171.6 280.7C1148.3 294.1 1148.3 327.8 1171.6 341.2L1206.2 361L1241.3 381.1C1264.5 394.4 1293.5 377.7 1293.5 350.9V346.4C1293.5 343.9 1292.2 341.6 1290 340.4L1273.7 331.1C1269.2 328.5 1263.6 331.8 1263.6 337V347.7C1263.6 353 1257.9 356.3 1253.3 353.7L1221.3 335.3L1189.4 317C1184.8 314.4 1184.8 307.7 1189.4 305.1L1221.3 286.8L1253.3 268.4C1257.9 265.8 1263.6 269.1 1263.6 274.4V279.7C1263.6 306.5 1292.6 323.3 1315.8 309.9L1350.9 289.8L1385.5 270C1408.8 256.5 1408.8 222.9 1385.5 209.5Z" fill="black"/>
-    <path d="M338.3 290.7C331.9 318 308.4 336.8 278 336.8C242.9 336.8 216.5 310.4 216.5 275.5C216.5 240.6 243.3 214 278.5 214C308.2 214 331.4 232.8 338.1 259.6H307.2C301.7 248.9 290.8 241.8 278.7 241.8C261.1 241.8 246.4 256.3 246.4 275.5C246.4 294.7 261.1 309.2 278.7 309.2C291.3 309.2 302.2 302.1 307.4 290.7H338.3Z" fill="black"/>
-    <path d="M352.2 334.9V167.3H382.1V334.9H352.2Z" fill="black"/>
-    <path d="M427.1 283.4C429.2 301 441.8 310.9 457.7 310.9C470.3 310.9 480 305 485.7 295.5H516.3C509.2 320.7 486.6 336.8 456.9 336.8C421 336.8 395.9 311.6 395.9 275.8C395.9 240 422.5 214.1 457.6 214.1C493 214.1 518.6 240.2 518.6 275.8C518.6 278.4 518.4 281 517.9 283.4H427.1V283.4ZM486.7 264.8C484.6 249.4 472 240.1 457.3 240.1C442.6 240.1 430.5 249.1 427.4 264.8H486.7Z" fill="black"/>
-    <path d="M625.4 334.9V319.5C616.6 330.2 603.6 336.8 587.9 336.8C555.4 336.8 530.7 310.7 530.7 275.5C530.7 240.6 556.3 214 588.6 214C603.8 214 616.6 220.6 625.4 231.1V216.1H654.6V334.8H625.4V334.9ZM625.1 275.5C625.1 255.6 610.6 241.5 592.8 241.5C575 241.5 560.5 255.7 560.5 275.5C560.5 295 575 309.5 592.8 309.5C610.6 309.5 625.1 295.5 625.1 275.5Z" fill="black"/>
-    </svg>
-    """
-  )
+  static func buyNow(palette: ColorPalette, locale: Locale) -> SVG {
+    switch (palette, locale) {
+    case (.blackOnMint, Locales.greatBritain), (.blackOnWhite, Locales.greatBritain):
+      return .clearpayBuyNowBlackOnMint
+    case (.whiteOnBlack, Locales.greatBritain), (.mintOnBlack, Locales.greatBritain):
+      return .clearpayBuyNowWhiteOnBlack
+    case (.blackOnMint, _), (.blackOnWhite, _):
+      return .afterpayBuyNowBlackOnMint
+    case (.whiteOnBlack, _), (.mintOnBlack, _):
+      return .afterpayBuyNowWhiteOnBlack
+    }
+  }
 
-  static let afterpayBadgeMintOnBlack = SVG(
-    size: CGSize(width: 1582, height: 551),
-    minimumWidth: 64,
-    svgString: """
-    <svg width="1582" height="551" viewBox="0 0 1582 551" fill="none" xmlns="http://www.w3.org/2000/svg">
-    <g clip-path="url(#clip0)">
-    <path d="M1306.3 550.5H275.2C123.2 550.5 0 427.3 0 275.3C0 123.3 123.2 0.100006 275.2 0.100006H1306.3C1458.3 0.100006 1581.5 123.3 1581.5 275.3C1581.6 427.2 1458.3 550.5 1306.3 550.5Z" fill="black"/>
-    <path d="M1348 209.5L1313.4 189.7L1278.3 169.6C1255.1 156.3 1226.1 173 1226.1 199.8V204.3C1226.1 206.8 1227.4 209.1 1229.6 210.3L1245.9 219.6C1250.4 222.2 1256 218.9 1256 213.7V203C1256 197.7 1261.7 194.4 1266.3 197L1298.3 215.4L1330.2 233.7C1334.8 236.3 1334.8 243 1330.2 245.6L1298.3 263.9L1266.3 282.3C1261.7 284.9 1256 281.6 1256 276.3V271C1256 244.2 1227 227.4 1203.8 240.8L1168.7 260.9L1134.1 280.7C1110.8 294.1 1110.8 327.8 1134.1 341.2L1168.7 361L1203.8 381.1C1227 394.4 1256 377.7 1256 350.9V346.4C1256 343.9 1254.7 341.6 1252.5 340.4L1236.2 331.1C1231.7 328.5 1226.1 331.8 1226.1 337V347.7C1226.1 353 1220.4 356.3 1215.8 353.7L1183.8 335.3L1151.9 317C1147.3 314.4 1147.3 307.7 1151.9 305.1L1183.8 286.8L1215.8 268.4C1220.4 265.8 1226.1 269.1 1226.1 274.4V279.7C1226.1 306.5 1255.1 323.3 1278.3 309.9L1313.4 289.8L1348 270C1371.3 256.5 1371.3 222.9 1348 209.5Z" fill="#B2FCE4"/>
-    <path d="M1121 216.1L1040 383.4H1006.4L1036.7 320.9L989 216.1H1023.5L1054.1 286.3L1087.5 216.1H1121Z" fill="#B2FCE4"/>
-    <path d="M311.1 275.5C311.1 255.5 296.6 241.5 278.8 241.5C261 241.5 246.5 255.8 246.5 275.5C246.5 295 261 309.5 278.8 309.5C296.6 309.5 311.1 295.5 311.1 275.5ZM311.4 334.9V319.5C302.6 330.2 289.5 336.8 273.9 336.8C241.3 336.8 216.6 310.7 216.6 275.5C216.6 240.6 242.3 214 274.6 214C289.8 214 302.6 220.7 311.4 231.1V216.1H340.6V334.9H311.4V334.9Z" fill="#B2FCE4"/>
-    <path d="M482.6 308.5C472.4 308.5 469.5 304.7 469.5 294.7V242H488.3V216.1H469.5V187.1H439.6V216.1H401V204.3C401 194.3 404.8 190.5 415.3 190.5H421.9V167.5H407.5C382.8 167.5 371.1 175.6 371.1 200.3V216.2H354.5V242H371.1V334.9H401V242H439.6V300.2C439.6 324.4 448.9 334.9 473.1 334.9H488.5V308.5H482.6Z" fill="#B2FCE4"/>
-    <path d="M590 264.8C587.9 249.4 575.3 240.1 560.5 240.1C545.8 240.1 533.6 249.1 530.6 264.8H590ZM530.3 283.3C532.4 300.9 545 310.9 561 310.9C573.6 310.9 583.3 305 589 295.5H619.7C612.6 320.7 590 336.8 560.3 336.8C524.4 336.8 499.2 311.6 499.2 275.7C499.2 239.8 525.8 213.9 561 213.9C596.4 213.9 622.1 240 622.1 275.7C622.1 278.3 621.9 280.9 621.4 283.3H530.3V283.3Z" fill="#B2FCE4"/>
-    <path d="M812.5 275.5C812.5 256.3 798 241.5 780.2 241.5C762.4 241.5 747.9 255.8 747.9 275.5C747.9 295 762.4 309.5 780.2 309.5C798 309.5 812.5 294.8 812.5 275.5ZM718.4 383.4V216.1H747.6V231.5C756.4 220.6 769.5 213.9 785.1 213.9C817.2 213.9 842.4 240.3 842.4 275.2C842.4 310.1 816.7 336.7 784.4 336.7C769.4 336.7 757.1 330.8 748.5 320.8V383.3H718.4V383.4Z" fill="#B2FCE4"/>
-    <path d="M947.7 275.5C947.7 255.5 933.2 241.5 915.4 241.5C897.6 241.5 883.1 255.8 883.1 275.5C883.1 295 897.6 309.5 915.4 309.5C933.2 309.5 947.7 295.5 947.7 275.5ZM948 334.9V319.5C939.2 330.2 926.1 336.8 910.5 336.8C877.9 336.8 853.2 310.7 853.2 275.5C853.2 240.6 878.9 214 911.2 214C926.4 214 939.2 220.7 948 231.1V216.1H977.2V334.9H948V334.9Z" fill="#B2FCE4"/>
-    <path d="M665.7 227.7C665.7 227.7 673.1 213.9 691.4 213.9C699.2 213.9 704.2 216.6 704.2 216.6V246.9C704.2 246.9 693.2 240.1 683.1 241.5C673 242.9 666.6 252.1 666.6 264.5V334.8H636.4V216.1H665.6V227.7H665.7Z" fill="#B2FCE4"/>
-    </g>
-    <defs>
-    <clipPath id="clip0">
-    <rect width="1582" height="551" fill="white"/>
-    </clipPath>
-    </defs>
-    </svg>
-    """
-  )
+  static func checkout(palette: ColorPalette, locale: Locale) -> SVG {
+    switch (palette, locale) {
+    case (.blackOnMint, Locales.greatBritain), (.blackOnWhite, Locales.greatBritain):
+      return .clearpayCheckoutBlackOnMint
+    case (.whiteOnBlack, Locales.greatBritain), (.mintOnBlack, Locales.greatBritain):
+      return .clearpayCheckoutWhiteOnBlack
+    case (.blackOnMint, _), (.blackOnWhite, _):
+      return .afterpayCheckoutBlackOnMint
+    case (.whiteOnBlack, _), (.mintOnBlack, _):
+      return .afterpayCheckoutWhiteOnBlack
+    }
+  }
 
-  static let clearpayBadgeMintOnBlack = SVG(
-    size: CGSize(width: 1620, height: 551),
-    minimumWidth: 64,
-    svgString: """
-    <svg width="1620" height="551" viewBox="0 0 1620 551" fill="none" xmlns="http://www.w3.org/2000/svg">
-    <path d="M1344 550.5H275.5C123.5 550.5 0.300003 427.3 0.300003 275.3C0.300003 123.3 123.5 0.100006 275.5 0.100006H1344C1496 0.100006 1619.2 123.3 1619.2 275.3C1619.3 427.2 1496 550.5 1344 550.5Z" fill="black"/>
-    <path d="M1158.7 216.2L1077.6 383.5H1044L1074.3 321L1026.7 216.2H1061.2L1091.8 286.4L1125.2 216.2H1158.7Z" fill="#B2FCE4"/>
-    <path d="M850.1 275.6C850.1 256.4 835.6 241.6 817.8 241.6C800 241.6 785.5 255.9 785.5 275.6C785.5 295.1 800 309.6 817.8 309.6C835.6 309.6 850.1 294.8 850.1 275.6ZM756 383.5V216.2H785.2V231.6C794 220.7 807.1 214 822.7 214C854.8 214 880 240.4 880 275.3C880 310.2 854.3 336.8 822 336.8C807 336.8 794.7 330.9 786.1 320.9V383.4H756V383.5Z" fill="#B2FCE4"/>
-    <path d="M985.4 275.6C985.4 255.6 970.9 241.6 953.1 241.6C935.3 241.6 920.8 255.9 920.8 275.6C920.8 295.1 935.3 309.6 953.1 309.6C970.9 309.6 985.4 295.5 985.4 275.6ZM985.6 335V319.6C976.8 330.3 963.7 336.9 948.1 336.9C915.5 336.9 890.8 310.8 890.8 275.6C890.8 240.7 916.5 214.1 948.8 214.1C964 214.1 976.8 220.8 985.6 231.2V216.2H1014.8V335H985.6Z" fill="#B2FCE4"/>
-    <path d="M703.3 227.8C703.3 227.8 710.7 214 729 214C736.8 214 741.8 216.7 741.8 216.7V247C741.8 247 730.8 240.2 720.7 241.6C710.6 243 704.2 252.2 704.2 264.6V335H674V216.2H703.2V227.8H703.3Z" fill="#B2FCE4"/>
-    <path d="M1385.5 209.5L1350.9 189.7L1315.8 169.6C1292.6 156.3 1263.6 173 1263.6 199.8V204.3C1263.6 206.8 1264.9 209.1 1267.1 210.3L1283.4 219.6C1287.9 222.2 1293.5 218.9 1293.5 213.7V203C1293.5 197.7 1299.2 194.4 1303.8 197L1335.8 215.4L1367.7 233.7C1372.3 236.3 1372.3 243 1367.7 245.6L1335.8 263.9L1303.8 282.3C1299.2 284.9 1293.5 281.6 1293.5 276.3V271C1293.5 244.2 1264.5 227.4 1241.3 240.8L1206.2 260.9L1171.6 280.7C1148.3 294.1 1148.3 327.8 1171.6 341.2L1206.2 361L1241.3 381.1C1264.5 394.4 1293.5 377.7 1293.5 350.9V346.4C1293.5 343.9 1292.2 341.6 1290 340.4L1273.7 331.1C1269.2 328.5 1263.6 331.8 1263.6 337V347.7C1263.6 353 1257.9 356.3 1253.3 353.7L1221.3 335.3L1189.4 317C1184.8 314.4 1184.8 307.7 1189.4 305.1L1221.3 286.8L1253.3 268.4C1257.9 265.8 1263.6 269.1 1263.6 274.4V279.7C1263.6 306.5 1292.6 323.3 1315.8 309.9L1350.9 289.8L1385.5 270C1408.8 256.5 1408.8 222.9 1385.5 209.5Z" fill="#B2FCE4"/>
-    <path d="M338.3 290.7C331.9 318 308.4 336.8 278 336.8C242.9 336.8 216.5 310.4 216.5 275.5C216.5 240.6 243.3 214 278.5 214C308.2 214 331.4 232.8 338.1 259.6H307.2C301.7 248.9 290.8 241.8 278.7 241.8C261.1 241.8 246.4 256.3 246.4 275.5C246.4 294.7 261.1 309.2 278.7 309.2C291.3 309.2 302.2 302.1 307.4 290.7H338.3Z" fill="#B2FCE4"/>
-    <path d="M352.2 334.9V167.3H382.1V334.9H352.2Z" fill="#B2FCE4"/>
-    <path d="M427.1 283.4C429.2 301 441.8 310.9 457.7 310.9C470.3 310.9 480 305 485.7 295.5H516.3C509.2 320.7 486.6 336.8 456.9 336.8C421 336.8 395.9 311.6 395.9 275.8C395.9 240 422.5 214.1 457.6 214.1C493 214.1 518.6 240.2 518.6 275.8C518.6 278.4 518.4 281 517.9 283.4H427.1V283.4ZM486.7 264.8C484.6 249.4 472 240.1 457.3 240.1C442.6 240.1 430.5 249.1 427.4 264.8H486.7Z" fill="#B2FCE4"/>
-    <path d="M625.4 334.9V319.5C616.6 330.2 603.6 336.8 587.9 336.8C555.4 336.8 530.7 310.7 530.7 275.5C530.7 240.6 556.3 214 588.6 214C603.8 214 616.6 220.6 625.4 231.1V216.1H654.6V334.8H625.4V334.9ZM625.1 275.5C625.1 255.6 610.6 241.5 592.8 241.5C575 241.5 560.5 255.7 560.5 275.5C560.5 295 575 309.5 592.8 309.5C610.6 309.5 625.1 295.5 625.1 275.5Z" fill="#B2FCE4"/>
-    </svg>
-    """
-  )
+  static func payNow(palette: ColorPalette, locale: Locale) -> SVG {
+    switch (palette, locale) {
+    case (.blackOnMint, Locales.greatBritain), (.blackOnWhite, Locales.greatBritain):
+      return .clearpayPayNowBlackOnMint
+    case (.whiteOnBlack, Locales.greatBritain), (.mintOnBlack, Locales.greatBritain):
+      return .clearpayPayNowWhiteOnBlack
+    case (.blackOnMint, _), (.blackOnWhite, _):
+      return .afterpayPayNowBlackOnMint
+    case (.whiteOnBlack, _), (.mintOnBlack, _):
+      return .afterpayPayNowWhiteOnBlack
+    }
+  }
 
-  static let afterpayBadgeWhiteOnBlack: SVG = SVG(
-    size: CGSize(width: 1582, height: 551),
-    minimumWidth: 64,
-    svgString: """
-    <svg width="1582" height="551" viewBox="0 0 1582 551" fill="none" xmlns="http://www.w3.org/2000/svg">
-    <g clip-path="url(#clip0)">
-    <path d="M1306.3 550.5H275.2C123.2 550.5 0 427.3 0 275.3C0 123.3 123.2 0.100006 275.2 0.100006H1306.3C1458.3 0.100006 1581.5 123.3 1581.5 275.3C1581.6 427.2 1458.3 550.5 1306.3 550.5Z" fill="black"/>
-    <path d="M1348 209.5L1313.4 189.7L1278.3 169.6C1255.1 156.3 1226.1 173 1226.1 199.8V204.3C1226.1 206.8 1227.4 209.1 1229.6 210.3L1245.9 219.6C1250.4 222.2 1256 218.9 1256 213.7V203C1256 197.7 1261.7 194.4 1266.3 197L1298.3 215.4L1330.2 233.7C1334.8 236.3 1334.8 243 1330.2 245.6L1298.3 263.9L1266.3 282.3C1261.7 284.9 1256 281.6 1256 276.3V271C1256 244.2 1227 227.4 1203.8 240.8L1168.7 260.9L1134.1 280.7C1110.8 294.1 1110.8 327.8 1134.1 341.2L1168.7 361L1203.8 381.1C1227 394.4 1256 377.7 1256 350.9V346.4C1256 343.9 1254.7 341.6 1252.5 340.4L1236.2 331.1C1231.7 328.5 1226.1 331.8 1226.1 337V347.7C1226.1 353 1220.4 356.3 1215.8 353.7L1183.8 335.3L1151.9 317C1147.3 314.4 1147.3 307.7 1151.9 305.1L1183.8 286.8L1215.8 268.4C1220.4 265.8 1226.1 269.1 1226.1 274.4V279.7C1226.1 306.5 1255.1 323.3 1278.3 309.9L1313.4 289.8L1348 270C1371.3 256.5 1371.3 222.9 1348 209.5Z" fill="white"/>
-    <path d="M1121 216.1L1040 383.4H1006.4L1036.7 320.9L989 216.1H1023.5L1054.1 286.3L1087.5 216.1H1121Z" fill="white"/>
-    <path d="M311.1 275.5C311.1 255.5 296.6 241.5 278.8 241.5C261 241.5 246.5 255.8 246.5 275.5C246.5 295 261 309.5 278.8 309.5C296.6 309.5 311.1 295.5 311.1 275.5ZM311.4 334.9V319.5C302.6 330.2 289.5 336.8 273.9 336.8C241.3 336.8 216.6 310.7 216.6 275.5C216.6 240.6 242.3 214 274.6 214C289.8 214 302.6 220.7 311.4 231.1V216.1H340.6V334.9H311.4V334.9Z" fill="white"/>
-    <path d="M482.6 308.5C472.4 308.5 469.5 304.7 469.5 294.7V242H488.3V216.1H469.5V187.1H439.6V216.1H401V204.3C401 194.3 404.8 190.5 415.3 190.5H421.9V167.5H407.5C382.8 167.5 371.1 175.6 371.1 200.3V216.2H354.5V242H371.1V334.9H401V242H439.6V300.2C439.6 324.4 448.9 334.9 473.1 334.9H488.5V308.5H482.6Z" fill="white"/>
-    <path d="M590 264.8C587.9 249.4 575.3 240.1 560.5 240.1C545.8 240.1 533.6 249.1 530.6 264.8H590ZM530.3 283.3C532.4 300.9 545 310.9 561 310.9C573.6 310.9 583.3 305 589 295.5H619.7C612.6 320.7 590 336.8 560.3 336.8C524.4 336.8 499.2 311.6 499.2 275.7C499.2 239.8 525.8 213.9 561 213.9C596.4 213.9 622.1 240 622.1 275.7C622.1 278.3 621.9 280.9 621.4 283.3H530.3V283.3Z" fill="white"/>
-    <path d="M812.5 275.5C812.5 256.3 798 241.5 780.2 241.5C762.4 241.5 747.9 255.8 747.9 275.5C747.9 295 762.4 309.5 780.2 309.5C798 309.5 812.5 294.8 812.5 275.5ZM718.4 383.4V216.1H747.6V231.5C756.4 220.6 769.5 213.9 785.1 213.9C817.2 213.9 842.4 240.3 842.4 275.2C842.4 310.1 816.7 336.7 784.4 336.7C769.4 336.7 757.1 330.8 748.5 320.8V383.3H718.4V383.4Z" fill="white"/>
-    <path d="M947.7 275.5C947.7 255.5 933.2 241.5 915.4 241.5C897.6 241.5 883.1 255.8 883.1 275.5C883.1 295 897.6 309.5 915.4 309.5C933.2 309.5 947.7 295.5 947.7 275.5ZM948 334.9V319.5C939.2 330.2 926.1 336.8 910.5 336.8C877.9 336.8 853.2 310.7 853.2 275.5C853.2 240.6 878.9 214 911.2 214C926.4 214 939.2 220.7 948 231.1V216.1H977.2V334.9H948V334.9Z" fill="white"/>
-    <path d="M665.7 227.7C665.7 227.7 673.1 213.9 691.4 213.9C699.2 213.9 704.2 216.6 704.2 216.6V246.9C704.2 246.9 693.2 240.1 683.1 241.5C673 242.9 666.6 252.1 666.6 264.5V334.8H636.4V216.1H665.6V227.7H665.7Z" fill="white"/>
-    </g>
-    <defs>
-    <clipPath id="clip0">
-    <rect width="1582" height="551" fill="white"/>
-    </clipPath>
-    </defs>
-    </svg>
-    """
-  )
-
-  static let clearpayBadgeWhiteOnBlack = SVG(
-    size: CGSize(width: 1620, height: 551),
-    minimumWidth: 64,
-    svgString: """
-    <svg width="1620" height="551" viewBox="0 0 1620 551" fill="none" xmlns="http://www.w3.org/2000/svg">
-    <path d="M1344 550.5H275.5C123.5 550.5 0.300049 427.3 0.300049 275.3C0.300049 123.3 123.5 0.0999756 275.5 0.0999756H1344C1496 0.0999756 1619.2 123.3 1619.2 275.3C1619.3 427.2 1496 550.5 1344 550.5Z" fill="black"/>
-    <path d="M1158.7 216.2L1077.6 383.5H1044L1074.3 321L1026.7 216.2H1061.2L1091.8 286.4L1125.2 216.2H1158.7Z" fill="white"/>
-    <path d="M850.1 275.6C850.1 256.4 835.6 241.6 817.8 241.6C800 241.6 785.5 255.9 785.5 275.6C785.5 295.1 800 309.6 817.8 309.6C835.6 309.6 850.1 294.8 850.1 275.6ZM756 383.5V216.2H785.2V231.6C794 220.7 807.1 214 822.7 214C854.8 214 880 240.4 880 275.3C880 310.2 854.3 336.8 822 336.8C807 336.8 794.7 330.9 786.1 320.9V383.4H756V383.5Z" fill="white"/>
-    <path d="M985.4 275.6C985.4 255.6 970.9 241.6 953.1 241.6C935.3 241.6 920.8 255.9 920.8 275.6C920.8 295.1 935.3 309.6 953.1 309.6C970.9 309.6 985.4 295.5 985.4 275.6ZM985.6 335V319.6C976.8 330.3 963.7 336.9 948.1 336.9C915.5 336.9 890.8 310.8 890.8 275.6C890.8 240.7 916.5 214.1 948.8 214.1C964 214.1 976.8 220.8 985.6 231.2V216.2H1014.8V335H985.6Z" fill="white"/>
-    <path d="M703.3 227.8C703.3 227.8 710.7 214 729 214C736.8 214 741.8 216.7 741.8 216.7V247C741.8 247 730.8 240.2 720.7 241.6C710.6 243 704.2 252.2 704.2 264.6V335H674V216.2H703.2V227.8H703.3Z" fill="white"/>
-    <path d="M1385.5 209.5L1350.9 189.7L1315.8 169.6C1292.6 156.3 1263.6 173 1263.6 199.8V204.3C1263.6 206.8 1264.9 209.1 1267.1 210.3L1283.4 219.6C1287.9 222.2 1293.5 218.9 1293.5 213.7V203C1293.5 197.7 1299.2 194.4 1303.8 197L1335.8 215.4L1367.7 233.7C1372.3 236.3 1372.3 243 1367.7 245.6L1335.8 263.9L1303.8 282.3C1299.2 284.9 1293.5 281.6 1293.5 276.3V271C1293.5 244.2 1264.5 227.4 1241.3 240.8L1206.2 260.9L1171.6 280.7C1148.3 294.1 1148.3 327.8 1171.6 341.2L1206.2 361L1241.3 381.1C1264.5 394.4 1293.5 377.7 1293.5 350.9V346.4C1293.5 343.9 1292.2 341.6 1290 340.4L1273.7 331.1C1269.2 328.5 1263.6 331.8 1263.6 337V347.7C1263.6 353 1257.9 356.3 1253.3 353.7L1221.3 335.3L1189.4 317C1184.8 314.4 1184.8 307.7 1189.4 305.1L1221.3 286.8L1253.3 268.4C1257.9 265.8 1263.6 269.1 1263.6 274.4V279.7C1263.6 306.5 1292.6 323.3 1315.8 309.9L1350.9 289.8L1385.5 270C1408.8 256.5 1408.8 222.9 1385.5 209.5Z" fill="white"/>
-    <path d="M338.3 290.7C331.9 318 308.4 336.8 278 336.8C242.9 336.8 216.5 310.4 216.5 275.5C216.5 240.6 243.3 214 278.5 214C308.2 214 331.4 232.8 338.1 259.6H307.2C301.7 248.9 290.8 241.8 278.7 241.8C261.1 241.8 246.4 256.3 246.4 275.5C246.4 294.7 261.1 309.2 278.7 309.2C291.3 309.2 302.2 302.1 307.4 290.7H338.3Z" fill="white"/>
-    <path d="M352.2 334.9V167.3H382.1V334.9H352.2Z" fill="white"/>
-    <path d="M427.1 283.4C429.2 301 441.8 310.9 457.7 310.9C470.3 310.9 480 305 485.7 295.5H516.3C509.2 320.7 486.6 336.8 456.9 336.8C421 336.8 395.9 311.6 395.9 275.8C395.9 240 422.5 214.1 457.6 214.1C493 214.1 518.6 240.2 518.6 275.8C518.6 278.4 518.4 281 517.9 283.4H427.1V283.4ZM486.7 264.8C484.6 249.4 472 240.1 457.3 240.1C442.6 240.1 430.5 249.1 427.4 264.8H486.7Z" fill="white"/>
-    <path d="M625.4 334.9V319.5C616.6 330.2 603.6 336.8 587.9 336.8C555.4 336.8 530.7 310.7 530.7 275.5C530.7 240.6 556.3 214 588.6 214C603.8 214 616.6 220.6 625.4 231.1V216.1H654.6V334.8H625.4V334.9ZM625.1 275.5C625.1 255.6 610.6 241.5 592.8 241.5C575 241.5 560.5 255.7 560.5 275.5C560.5 295 575 309.5 592.8 309.5C610.6 309.5 625.1 295.5 625.1 275.5Z" fill="white"/>
-    </svg>
-    """
-  )
-
-  static let afterpayBadgeBlackOnWhite: SVG = SVG(
-    size: CGSize(width: 1582, height: 551),
-    minimumWidth: 64,
-    svgString: """
-    <svg width="1582" height="551" viewBox="0 0 1582 551" fill="none" xmlns="http://www.w3.org/2000/svg">
-    <g clip-path="url(#clip0)">
-    <path d="M1306.3 550.5H275.2C123.2 550.5 0 427.3 0 275.3C0 123.3 123.2 0.100006 275.2 0.100006H1306.3C1458.3 0.100006 1581.5 123.3 1581.5 275.3C1581.6 427.2 1458.3 550.5 1306.3 550.5Z" fill="white"/>
-    <path d="M1348 209.5L1313.4 189.7L1278.3 169.6C1255.1 156.3 1226.1 173 1226.1 199.8V204.3C1226.1 206.8 1227.4 209.1 1229.6 210.3L1245.9 219.6C1250.4 222.2 1256 218.9 1256 213.7V203C1256 197.7 1261.7 194.4 1266.3 197L1298.3 215.4L1330.2 233.7C1334.8 236.3 1334.8 243 1330.2 245.6L1298.3 263.9L1266.3 282.3C1261.7 284.9 1256 281.6 1256 276.3V271C1256 244.2 1227 227.4 1203.8 240.8L1168.7 260.9L1134.1 280.7C1110.8 294.1 1110.8 327.8 1134.1 341.2L1168.7 361L1203.8 381.1C1227 394.4 1256 377.7 1256 350.9V346.4C1256 343.9 1254.7 341.6 1252.5 340.4L1236.2 331.1C1231.7 328.5 1226.1 331.8 1226.1 337V347.7C1226.1 353 1220.4 356.3 1215.8 353.7L1183.8 335.3L1151.9 317C1147.3 314.4 1147.3 307.7 1151.9 305.1L1183.8 286.8L1215.8 268.4C1220.4 265.8 1226.1 269.1 1226.1 274.4V279.7C1226.1 306.5 1255.1 323.3 1278.3 309.9L1313.4 289.8L1348 270C1371.3 256.5 1371.3 222.9 1348 209.5Z" fill="black"/>
-    <path d="M1121 216.1L1040 383.4H1006.4L1036.7 320.9L989 216.1H1023.5L1054.1 286.3L1087.5 216.1H1121Z" fill="black"/>
-    <path d="M311.1 275.5C311.1 255.5 296.6 241.5 278.8 241.5C261 241.5 246.5 255.8 246.5 275.5C246.5 295 261 309.5 278.8 309.5C296.6 309.5 311.1 295.5 311.1 275.5ZM311.4 334.9V319.5C302.6 330.2 289.5 336.8 273.9 336.8C241.3 336.8 216.6 310.7 216.6 275.5C216.6 240.6 242.3 214 274.6 214C289.8 214 302.6 220.7 311.4 231.1V216.1H340.6V334.9H311.4V334.9Z" fill="black"/>
-    <path d="M482.6 308.5C472.4 308.5 469.5 304.7 469.5 294.7V242H488.3V216.1H469.5V187.1H439.6V216.1H401V204.3C401 194.3 404.8 190.5 415.3 190.5H421.9V167.5H407.5C382.8 167.5 371.1 175.6 371.1 200.3V216.2H354.5V242H371.1V334.9H401V242H439.6V300.2C439.6 324.4 448.9 334.9 473.1 334.9H488.5V308.5H482.6Z" fill="black"/>
-    <path d="M590 264.8C587.9 249.4 575.3 240.1 560.5 240.1C545.8 240.1 533.6 249.1 530.6 264.8H590ZM530.3 283.3C532.4 300.9 545 310.9 561 310.9C573.6 310.9 583.3 305 589 295.5H619.7C612.6 320.7 590 336.8 560.3 336.8C524.4 336.8 499.2 311.6 499.2 275.7C499.2 239.8 525.8 213.9 561 213.9C596.4 213.9 622.1 240 622.1 275.7C622.1 278.3 621.9 280.9 621.4 283.3H530.3V283.3Z" fill="black"/>
-    <path d="M812.5 275.5C812.5 256.3 798 241.5 780.2 241.5C762.4 241.5 747.9 255.8 747.9 275.5C747.9 295 762.4 309.5 780.2 309.5C798 309.5 812.5 294.8 812.5 275.5ZM718.4 383.4V216.1H747.6V231.5C756.4 220.6 769.5 213.9 785.1 213.9C817.2 213.9 842.4 240.3 842.4 275.2C842.4 310.1 816.7 336.7 784.4 336.7C769.4 336.7 757.1 330.8 748.5 320.8V383.3H718.4V383.4Z" fill="black"/>
-    <path d="M947.7 275.5C947.7 255.5 933.2 241.5 915.4 241.5C897.6 241.5 883.1 255.8 883.1 275.5C883.1 295 897.6 309.5 915.4 309.5C933.2 309.5 947.7 295.5 947.7 275.5ZM948 334.9V319.5C939.2 330.2 926.1 336.8 910.5 336.8C877.9 336.8 853.2 310.7 853.2 275.5C853.2 240.6 878.9 214 911.2 214C926.4 214 939.2 220.7 948 231.1V216.1H977.2V334.9H948V334.9Z" fill="black"/>
-    <path d="M665.7 227.7C665.7 227.7 673.1 213.9 691.4 213.9C699.2 213.9 704.2 216.6 704.2 216.6V246.9C704.2 246.9 693.2 240.1 683.1 241.5C673 242.9 666.6 252.1 666.6 264.5V334.8H636.4V216.1H665.6V227.7H665.7Z" fill="black"/>
-    </g>
-    <defs>
-    <clipPath id="clip0">
-    <rect width="1582" height="551" fill="white"/>
-    </clipPath>
-    </defs>
-    </svg>
-    """
-  )
-
-  static let clearpayBadgeBlackOnWhite = SVG(
-    size: CGSize(width: 1620, height: 551),
-    minimumWidth: 64,
-    svgString: """
-    <svg width="1620" height="551" viewBox="0 0 1620 551" fill="none" xmlns="http://www.w3.org/2000/svg">
-    <path d="M1344 550.5H275.5C123.5 550.5 0.300003 427.3 0.300003 275.3C0.300003 123.3 123.5 0.0999756 275.5 0.0999756H1344C1496 0.0999756 1619.2 123.3 1619.2 275.3C1619.3 427.2 1496 550.5 1344 550.5Z" fill="white"/>
-    <path d="M1158.7 216.2L1077.6 383.5H1044L1074.3 321L1026.7 216.2H1061.2L1091.8 286.4L1125.2 216.2H1158.7Z" fill="black"/>
-    <path d="M850.1 275.6C850.1 256.4 835.6 241.6 817.8 241.6C800 241.6 785.5 255.9 785.5 275.6C785.5 295.1 800 309.6 817.8 309.6C835.6 309.6 850.1 294.8 850.1 275.6ZM756 383.5V216.2H785.2V231.6C794 220.7 807.1 214 822.7 214C854.8 214 880 240.4 880 275.3C880 310.2 854.3 336.8 822 336.8C807 336.8 794.7 330.9 786.1 320.9V383.4H756V383.5Z" fill="black"/>
-    <path d="M985.4 275.6C985.4 255.6 970.9 241.6 953.1 241.6C935.3 241.6 920.8 255.9 920.8 275.6C920.8 295.1 935.3 309.6 953.1 309.6C970.9 309.6 985.4 295.5 985.4 275.6ZM985.6 335V319.6C976.8 330.3 963.7 336.9 948.1 336.9C915.5 336.9 890.8 310.8 890.8 275.6C890.8 240.7 916.5 214.1 948.8 214.1C964 214.1 976.8 220.8 985.6 231.2V216.2H1014.8V335H985.6Z" fill="black"/>
-    <path d="M703.3 227.8C703.3 227.8 710.7 214 729 214C736.8 214 741.8 216.7 741.8 216.7V247C741.8 247 730.8 240.2 720.7 241.6C710.6 243 704.2 252.2 704.2 264.6V335H674V216.2H703.2V227.8H703.3Z" fill="black"/>
-    <path d="M1385.5 209.5L1350.9 189.7L1315.8 169.6C1292.6 156.3 1263.6 173 1263.6 199.8V204.3C1263.6 206.8 1264.9 209.1 1267.1 210.3L1283.4 219.6C1287.9 222.2 1293.5 218.9 1293.5 213.7V203C1293.5 197.7 1299.2 194.4 1303.8 197L1335.8 215.4L1367.7 233.7C1372.3 236.3 1372.3 243 1367.7 245.6L1335.8 263.9L1303.8 282.3C1299.2 284.9 1293.5 281.6 1293.5 276.3V271C1293.5 244.2 1264.5 227.4 1241.3 240.8L1206.2 260.9L1171.6 280.7C1148.3 294.1 1148.3 327.8 1171.6 341.2L1206.2 361L1241.3 381.1C1264.5 394.4 1293.5 377.7 1293.5 350.9V346.4C1293.5 343.9 1292.2 341.6 1290 340.4L1273.7 331.1C1269.2 328.5 1263.6 331.8 1263.6 337V347.7C1263.6 353 1257.9 356.3 1253.3 353.7L1221.3 335.3L1189.4 317C1184.8 314.4 1184.8 307.7 1189.4 305.1L1221.3 286.8L1253.3 268.4C1257.9 265.8 1263.6 269.1 1263.6 274.4V279.7C1263.6 306.5 1292.6 323.3 1315.8 309.9L1350.9 289.8L1385.5 270C1408.8 256.5 1408.8 222.9 1385.5 209.5Z" fill="black"/>
-    <path d="M338.3 290.7C331.9 318 308.4 336.8 278 336.8C242.9 336.8 216.5 310.4 216.5 275.5C216.5 240.6 243.3 214 278.5 214C308.2 214 331.4 232.8 338.1 259.6H307.2C301.7 248.9 290.8 241.8 278.7 241.8C261.1 241.8 246.4 256.3 246.4 275.5C246.4 294.7 261.1 309.2 278.7 309.2C291.3 309.2 302.2 302.1 307.4 290.7H338.3Z" fill="black"/>
-    <path d="M352.2 334.9V167.3H382.1V334.9H352.2Z" fill="black"/>
-    <path d="M427.1 283.4C429.2 301 441.8 310.9 457.7 310.9C470.3 310.9 480 305 485.7 295.5H516.3C509.2 320.7 486.6 336.8 456.9 336.8C421 336.8 395.9 311.6 395.9 275.8C395.9 240 422.5 214.1 457.6 214.1C493 214.1 518.6 240.2 518.6 275.8C518.6 278.4 518.4 281 517.9 283.4H427.1V283.4ZM486.7 264.8C484.6 249.4 472 240.1 457.3 240.1C442.6 240.1 430.5 249.1 427.4 264.8H486.7Z" fill="black"/>
-    <path d="M625.4 334.9V319.5C616.6 330.2 603.6 336.8 587.9 336.8C555.4 336.8 530.7 310.7 530.7 275.5C530.7 240.6 556.3 214 588.6 214C603.8 214 616.6 220.6 625.4 231.1V216.1H654.6V334.8H625.4V334.9ZM625.1 275.5C625.1 255.6 610.6 241.5 592.8 241.5C575 241.5 560.5 255.7 560.5 275.5C560.5 295 575 309.5 592.8 309.5C610.6 309.5 625.1 295.5 625.1 275.5Z" fill="black"/>
-    </svg>
-    """
-  )
-
-  static let afterpayPayNowBlackOnMint = SVG(
-    size: CGSize(width: 1200, height: 180),
-    minimumWidth: 256,
-    svgString: """
-    <svg width="1200" height="180" viewBox="0 0 1200 180" fill="none" xmlns="http://www.w3.org/2000/svg">
-    <g clip-path="url(#clip0)">
-    <path d="M1169.8 180H30.2C13.6 180 0 166.4 0 149.8V30.2C0 13.6 13.6 0 30.2 0H1169.8C1186.4 0 1200 13.6 1200 30.2V149.8C1200 166.4 1186.4 180 1169.8 180Z" fill="#B2FCE4"/>
-    <path d="M161.2 107V63.4001H180.8C190.3 63.4001 196.4 68.4001 196.4 77.0001C196.4 85.7001 189.6 91.0001 180.5 91.0001H169.7V107H161.2ZM180.3 84.1001C184.8 84.1001 187.8 81.0001 187.8 77.1001C187.8 73.4001 184.8 70.6001 180.4 70.6001H169.7V84.2001H180.3V84.1001Z" fill="black"/>
-    <path d="M223.6 107V102.8C221.2 105.7 217.7 107.5 213.5 107.5C204.7 107.5 198.1 100.5 198.1 91.0001C198.1 81.6001 205 74.4001 213.7 74.4001C217.8 74.4001 221.3 76.2001 223.6 79.0001V75.0001H231.5V107H223.6ZM223.6 91.0001C223.6 85.6001 219.7 81.8001 214.9 81.8001C210.1 81.8001 206.2 85.6001 206.2 91.0001C206.2 96.2001 210.1 100.2 214.9 100.2C219.7 100.2 223.6 96.3001 223.6 91.0001Z" fill="black"/>
-    <path d="M239.1 118.7V111.7H241.7C243.9 111.7 246 110.6 246.6 107.3C246.8 106.3 246.7 105.8 246.3 104.8L234.5 75H242.8L251.3 97.1L259.6 75H267.7L255.8 105C251.9 114.9 250.8 118.8 241.5 118.8H239.1V118.7Z" fill="black"/>
-    <path d="M301.9 107V87.1001C301.9 83.6001 299.5 81.7001 296.3 81.7001C293 81.7001 290.4 83.8001 290.4 87.1001V107.1H282.3V75.0001H290.2V79.0001C291.4 77.1001 294.3 74.4001 299.1 74.4001C305.9 74.4001 310 78.8001 310 85.6001V107H301.9Z" fill="black"/>
-    <path d="M331.7 107.5C322.2 107.5 314.9 100.4 314.9 91.0001C314.9 81.6001 322.1 74.4001 331.7 74.4001C341.2 74.4001 348.5 81.6001 348.5 91.0001C348.4 100.4 341.1 107.5 331.7 107.5ZM340.3 91.0001C340.3 85.8001 336.4 81.9001 331.7 81.9001C327 81.9001 323 85.8001 323 91.0001C323 96.2001 327 100.1 331.7 100.1C336.4 100.1 340.3 96.1001 340.3 91.0001Z" fill="black"/>
-    <path d="M380.1 107L373.5 87.2H373.3L366.5 107H359.1L349.6 75H358L364 95.7H364.1L370.9 75H376.6L383.1 95.9H383.2L389.4 75H397.6L387.6 107H380.1Z" fill="black"/>
-    <path d="M440.6 107L434 87.2H433.8L427 107H419.6L410.1 75H418.5L424.5 95.7H424.6L431.4 75H437.1L443.6 95.9H443.7L449.9 75H458.1L448.1 107H440.6Z" fill="black"/>
-    <path d="M461.5 66.4001C461.5 64.0001 463.6 61.8 466.1 61.8C468.6 61.8 470.7 64.0001 470.7 66.4001C470.7 68.9001 468.6 71 466.1 71C463.6 71 461.5 68.9001 461.5 66.4001ZM462 107V75H470.1V107H462Z" fill="black"/>
-    <path d="M487.4 107C480.9 107 478.4 104.2 478.4 97.7001V81.9001H474.2V74.9001H478.4V67.1001H486.5V75.0001H491.6V82.0001H486.5V96.2001C486.5 98.9001 487.3 99.9001 490 99.9001H491.6V107H487.4Z" fill="black"/>
-    <path d="M515.7 107V87.1001C515.7 83.6001 513.3 81.7001 510.1 81.7001C506.8 81.7001 504.2 83.8001 504.2 87.1001V107.1H496.1V63.4001H504.2V78.6001C505.5 76.7001 508.3 74.4001 512.8 74.4001C519.6 74.4001 523.7 78.8001 523.7 85.6001V107H515.7Z" fill="black"/>
-    <path d="M1035.2 61L1020.3 52.5L1005.2 43.9C995.2 38.2 982.8 45.4 982.8 56.9V58.8C982.8 59.9 983.4 60.8 984.3 61.4L991.3 65.4C993.2 66.5 995.7 65.1 995.7 62.9V58.3C995.7 56 998.2 54.6 1000.1 55.7L1013.9 63.6L1027.6 71.5C1029.6 72.6 1029.6 75.5 1027.6 76.6L1013.9 84.5L1000.1 92.4C998.1 93.5 995.7 92.1 995.7 89.8V87.5C995.7 76 983.2 68.8 973.3 74.5L958.2 83.1L943.3 91.6C933.3 97.4 933.3 111.8 943.3 117.6L958.2 126.1L973.3 134.7C983.3 140.4 995.7 133.2 995.7 121.7V119.8C995.7 118.7 995.1 117.8 994.2 117.2L987.2 113.2C985.3 112.1 982.8 113.5 982.8 115.7V120.3C982.8 122.6 980.3 124 978.4 122.9L964.6 115L950.9 107.1C948.9 106 948.9 103.1 950.9 102L964.6 94.1L978.4 86.2C980.4 85.1 982.8 86.5 982.8 88.8V91.1C982.8 102.6 995.3 109.8 1005.2 104.1L1020.3 95.5L1035.2 87C1045.2 81.2 1045.2 66.8 1035.2 61Z" fill="black"/>
-    <path d="M937.6 63.8L902.7 135.8H888.2L901.3 108.9L880.8 63.8H895.6L908.8 94L923.2 63.8H937.6Z" fill="black"/>
-    <path d="M589.3 89.3999C589.3 80.7999 583.1 74.7999 575.4 74.7999C567.7 74.7999 561.5 80.8999 561.5 89.3999C561.5 97.7999 567.7 104 575.4 104C583.1 104 589.3 97.9999 589.3 89.3999ZM589.4 114.9V108.3C585.6 112.9 580 115.8 573.3 115.8C559.3 115.8 548.7 104.6 548.7 89.3999C548.7 74.3999 559.7 62.8999 573.6 62.8999C580.1 62.8999 585.7 65.7999 589.4 70.2999V63.8999H602V115L589.4 114.9Z" fill="black"/>
-    <path d="M663 103.6C658.6 103.6 657.4 102 657.4 97.7V75H665.5V63.8H657.4V51.4H644.5V63.9H627.9V58.8C627.9 54.5 629.5 52.9 634 52.9H636.8V43H630.6C620 43 615 46.5 615 57.1V63.9H607.8V75H615V115H627.9V75H644.5V100C644.5 110.4 648.5 114.9 658.9 114.9H665.5V103.6H663Z" fill="black"/>
-    <path d="M709.2 84.8C708.3 78.2 702.9 74.2 696.5 74.2C690.2 74.2 685 78.1 683.6 84.8H709.2ZM683.6 92.8C684.5 100.4 689.9 104.7 696.8 104.7C702.2 104.7 706.4 102.1 708.9 98.1H722.1C719 108.9 709.3 115.9 696.6 115.9C681.2 115.9 670.3 105.1 670.3 89.6C670.3 74.2 681.7 63 696.9 63C712.1 63 723.2 74.2 723.2 89.6C723.2 90.7 723.1 91.8 722.9 92.9H683.6V92.8Z" fill="black"/>
-    <path d="M804.9 89.4001C804.9 81.1001 798.7 74.8001 791 74.8001C783.3 74.8001 777.1 80.9001 777.1 89.4001C777.1 97.8 783.3 104 791 104C798.6 104 804.9 97.7001 804.9 89.4001ZM764.4 135.8V63.8H777V70.4001C780.8 65.7001 786.4 62.8 793.1 62.8C806.9 62.8 817.7 74.1 817.7 89.2C817.7 104.2 806.7 115.7 792.8 115.7C786.4 115.7 781 113.1 777.4 108.9V135.8H764.4Z" fill="black"/>
-    <path d="M863 89.3999C863 80.7999 856.8 74.7999 849.1 74.7999C841.4 74.7999 835.2 80.8999 835.2 89.3999C835.2 97.7999 841.4 104 849.1 104C856.8 104 863 97.9999 863 89.3999ZM863.2 114.9V108.3C859.4 112.9 853.8 115.8 847.1 115.8C833.1 115.8 822.5 104.6 822.5 89.3999C822.5 74.3999 833.5 62.8999 847.4 62.8999C853.9 62.8999 859.5 65.7999 863.2 70.2999V63.8999H875.8V115L863.2 114.9Z" fill="black"/>
-    <path d="M741.8 68.7999C741.8 68.7999 745 62.8999 752.8 62.8999C756.2 62.8999 758.3 64.0999 758.3 64.0999V77.0999C758.3 77.0999 753.6 74.1999 749.2 74.7999C744.9 75.3999 742.1 79.3999 742.1 84.6999V114.9H729.1V63.7999H741.7L741.8 68.7999Z" fill="black"/>
-    </g>
-    <defs>
-    <clipPath id="clip0">
-    <rect width="1200" height="180" fill="white"/>
-    </clipPath>
-    </defs>
-    </svg>
-    """
-  )
-
-  static let clearpayPayNowBlackOnMint = SVG(
-    size: CGSize(width: 1200, height: 180),
-    minimumWidth: 256,
-    svgString: """
-    <svg width="1200" height="180" viewBox="0 0 1200 180" fill="none" xmlns="http://www.w3.org/2000/svg">
-    <g clip-path="url(#clip0)">
-    <path d="M1169.8 180H30.2C13.6 180 0 166.4 0 149.8V30.2C0 13.6 13.6 0 30.2 0H1169.8C1186.4 0 1200 13.6 1200 30.2V149.8C1200 166.4 1186.4 180 1169.8 180Z" fill="#B2FCE4"/>
-    <path d="M944.4 64.4036L909.6 136.104H895.2L908.2 109.304L887.8 64.4036H902.6L915.7 94.5036L930 64.4036H944.4Z" fill="black"/>
-    <path d="M812.1 89.9037C812.1 81.6037 805.9 75.3037 798.2 75.3037C790.6 75.3037 784.3 81.4037 784.3 89.9037C784.3 98.3037 790.5 104.504 798.2 104.504C805.9 104.504 812.1 98.1037 812.1 89.9037ZM771.7 136.204V64.4037H784.2V71.0037C788 66.3037 793.6 63.5037 800.3 63.5037C814.1 63.5037 824.9 74.8037 824.9 89.8037C824.9 104.804 813.9 116.204 800 116.204C793.6 116.204 788.3 113.704 784.6 109.404V136.204H771.7Z" fill="black"/>
-    <path d="M870.1 89.9037C870.1 81.3037 863.9 75.3037 856.2 75.3037C848.6 75.3037 842.3 81.4037 842.3 89.9037C842.3 98.3037 848.5 104.504 856.2 104.504C863.9 104.504 870.1 98.4037 870.1 89.9037ZM870.2 115.404V108.804C866.4 113.404 860.8 116.204 854.1 116.204C840.1 116.204 829.5 105.004 829.5 89.9037C829.5 74.9037 840.5 63.5037 854.4 63.5037C860.9 63.5037 866.4 66.4037 870.2 70.8037V64.4037H882.7V115.304L870.2 115.404Z" fill="black"/>
-    <path d="M749.1 69.4037C749.1 69.4037 752.3 63.5037 760.1 63.5037C763.4 63.5037 765.6 64.7037 765.6 64.7037V77.7037C765.6 77.7037 760.9 74.8037 756.6 75.4037C752.3 76.0037 749.5 80.0037 749.5 85.3037V115.504H736.6V64.4037H749.1V69.4037Z" fill="black"/>
-    <path d="M1041.7 61.5037L1026.9 53.0037L1011.9 44.4037C1001.9 38.7037 989.5 45.9037 989.5 57.4037V59.3037C989.5 60.4037 990.1 61.3037 991 61.9037L998 65.9037C999.9 67.0037 1002.3 65.6037 1002.3 63.4037V58.8037C1002.3 56.5037 1004.8 55.1037 1006.7 56.2037L1020.4 64.1037L1034.1 72.0037C1036.1 73.1037 1036.1 76.0037 1034.1 77.1037L1020.4 85.0037L1006.7 92.9037C1004.7 94.0037 1002.3 92.6037 1002.3 90.3037V88.0037C1002.3 76.5037 989.9 69.3037 979.9 75.0037L964.9 83.6037L950 92.1037C940 97.8037 940 112.304 950 118.004L964.8 126.504L979.8 135.104C989.8 140.804 1002.2 133.604 1002.2 122.104V120.204C1002.2 119.104 1001.6 118.204 1000.7 117.604L993.7 113.604C991.8 112.504 989.4 113.904 989.4 116.104V120.704C989.4 123.004 986.9 124.404 985 123.304L971.3 115.404L957.6 107.504C955.6 106.404 955.6 103.504 957.6 102.404L971.3 94.5037L985 86.6037C987 85.5037 989.4 86.9037 989.4 89.2037V91.5037C989.4 103.004 1001.8 110.204 1011.8 104.504L1026.8 95.9037L1041.6 87.4037C1051.7 81.7037 1051.7 67.3037 1041.7 61.5037V61.5037Z" fill="black"/>
-    <path d="M592.6 96.4037C589.9 108.104 579.8 116.204 566.7 116.204C551.6 116.204 540.3 104.904 540.3 89.9037C540.3 74.9037 551.8 63.5037 566.9 63.5037C579.6 63.5037 589.6 71.5037 592.5 83.0037H579.3C577 78.4037 572.3 75.4037 567.1 75.4037C559.6 75.4037 553.3 81.6037 553.3 89.9037C553.3 98.1037 559.6 104.404 567.1 104.404C572.5 104.404 577.2 101.304 579.4 96.5037L592.6 96.4037Z" fill="black"/>
-    <path d="M598.6 115.304V43.5037H611.4V115.304H598.6Z" fill="black"/>
-    <path d="M630.7 93.2036C631.6 100.704 637 105.004 643.8 105.004C649.2 105.004 653.4 102.504 655.8 98.4036H669C665.9 109.204 656.3 116.104 643.5 116.104C628.1 116.104 617.3 105.304 617.3 89.9036C617.3 74.5036 628.7 63.4036 643.8 63.4036C659 63.5036 670 74.7036 670 90.0036C670 91.1036 669.9 92.2036 669.7 93.3036H630.7V93.2036ZM656.2 85.3036C655.3 78.7036 649.9 74.7036 643.6 74.7036C637.3 74.7036 632.1 78.6036 630.8 85.3036H656.2Z" fill="black"/>
-    <path d="M715.7 115.304V108.704C711.9 113.304 706.3 116.104 699.6 116.104C685.7 116.104 675.1 104.904 675.1 89.8036C675.1 74.8036 686.1 63.4036 699.9 63.4036C706.4 63.4036 711.9 66.3036 715.7 70.7036V64.3036H728.2V115.204L715.7 115.304ZM715.6 89.9036C715.6 81.3036 709.4 75.3036 701.8 75.3036C694.2 75.3036 688 81.4036 688 89.9036C688 98.2036 694.2 104.504 701.8 104.504C709.4 104.504 715.6 98.4036 715.6 89.9036Z" fill="black"/>
-    <path d="M150 111.204V67.6038H169.6C179.1 67.6038 185.2 72.6038 185.2 81.2038C185.2 89.9038 178.4 95.2038 169.3 95.2038H158.5V111.204H150ZM169.1 88.3038C173.6 88.3038 176.6 85.2038 176.6 81.3038C176.6 77.6038 173.6 74.8038 169.2 74.8038H158.5V88.4038H169.1V88.3038Z" fill="black"/>
-    <path d="M212.4 111.204V107.004C210 109.904 206.5 111.704 202.3 111.704C193.5 111.704 186.9 104.704 186.9 95.2038C186.9 85.8038 193.8 78.6038 202.5 78.6038C206.6 78.6038 210.1 80.4038 212.4 83.2038V79.2038H220.3V111.204H212.4ZM212.4 95.2038C212.4 89.8038 208.5 86.0038 203.7 86.0038C198.9 86.0038 195 89.8038 195 95.2038C195 100.404 198.9 104.404 203.7 104.404C208.5 104.404 212.4 100.504 212.4 95.2038Z" fill="black"/>
-    <path d="M227.9 122.904V115.904H230.5C232.7 115.904 234.8 114.804 235.4 111.504C235.6 110.504 235.5 110.004 235.1 109.004L223.3 79.2036H231.6L240.1 101.304L248.4 79.2036H256.5L244.6 109.204C240.7 119.104 239.6 123.004 230.3 123.004H227.9V122.904Z" fill="black"/>
-    <path d="M290.7 111.204V91.3038C290.7 87.8038 288.3 85.9038 285.1 85.9038C281.8 85.9038 279.2 88.0038 279.2 91.3038V111.304H271.1V79.2038H279V83.2038C280.2 81.3038 283.1 78.6038 287.9 78.6038C294.7 78.6038 298.8 83.0038 298.8 89.8038V111.204H290.7Z" fill="black"/>
-    <path d="M320.5 111.704C311 111.704 303.7 104.604 303.7 95.2038C303.7 85.8038 310.9 78.6038 320.5 78.6038C330 78.6038 337.3 85.8038 337.3 95.2038C337.2 104.604 329.9 111.704 320.5 111.704ZM329.1 95.2038C329.1 90.0038 325.2 86.1038 320.5 86.1038C315.8 86.1038 311.8 90.0038 311.8 95.2038C311.8 100.404 315.8 104.304 320.5 104.304C325.2 104.304 329.1 100.304 329.1 95.2038Z" fill="black"/>
-    <path d="M368.9 111.204L362.3 91.4036H362.1L355.3 111.204H347.9L338.4 79.2036H346.8L352.8 99.9036H352.9L359.7 79.2036H365.4L371.9 100.104H372L378.2 79.2036H386.4L376.4 111.204H368.9Z" fill="black"/>
-    <path d="M429.4 111.204L422.8 91.4036H422.6L415.8 111.204H408.4L398.9 79.2036H407.3L413.3 99.9036H413.4L420.2 79.2036H425.9L432.4 100.104H432.5L438.7 79.2036H446.9L436.9 111.204H429.4Z" fill="black"/>
-    <path d="M450.3 70.6037C450.3 68.2037 452.4 66.0037 454.9 66.0037C457.4 66.0037 459.5 68.2037 459.5 70.6037C459.5 73.1037 457.4 75.2037 454.9 75.2037C452.4 75.2037 450.3 73.1037 450.3 70.6037ZM450.8 111.204V79.2037H458.9V111.204H450.8Z" fill="black"/>
-    <path d="M476.2 111.204C469.7 111.204 467.2 108.404 467.2 101.904V86.1037H463V79.1037H467.2V71.3037H475.3V79.2037H480.4V86.2037H475.3V100.404C475.3 103.104 476.1 104.104 478.8 104.104H480.4V111.204H476.2Z" fill="black"/>
-    <path d="M504.5 111.204V91.3038C504.5 87.8038 502.1 85.9038 498.9 85.9038C495.6 85.9038 493 88.0038 493 91.3038V111.304H484.9V67.6038H493V82.8038C494.3 80.9038 497.1 78.6038 501.6 78.6038C508.4 78.6038 512.5 83.0038 512.5 89.8038V111.204H504.5Z" fill="black"/>
-    </g>
-    <defs>
-    <clipPath id="clip0">
-    <rect width="1200" height="180" fill="white"/>
-    </clipPath>
-    </defs>
-    </svg>
-    """
-  )
-
-  static let afterpayPayNowMintOnBlack = SVG(
-    size: CGSize(width: 1200, height: 180),
-    minimumWidth: 256,
-    svgString: """
-    <svg width="1200" height="180" viewBox="0 0 1200 180" fill="none" xmlns="http://www.w3.org/2000/svg">
-    <g clip-path="url(#clip0)">
-    <path d="M1169.8 180H30.2C13.6 180 0 166.4 0 149.8V30.2C0 13.6 13.6 0 30.2 0H1169.8C1186.4 0 1200 13.6 1200 30.2V149.8C1200 166.4 1186.4 180 1169.8 180Z" fill="black"/>
-    <path d="M161.2 107V63.4001H180.8C190.3 63.4001 196.4 68.4001 196.4 77.0001C196.4 85.7001 189.6 91.0001 180.5 91.0001H169.7V107H161.2ZM180.3 84.1001C184.8 84.1001 187.8 81.0001 187.8 77.1001C187.8 73.4001 184.8 70.6001 180.4 70.6001H169.7V84.2001H180.3V84.1001Z" fill="#B2FCE4"/>
-    <path d="M223.6 107V102.8C221.2 105.7 217.7 107.5 213.5 107.5C204.7 107.5 198.1 100.5 198.1 91.0001C198.1 81.6001 205 74.4001 213.7 74.4001C217.8 74.4001 221.3 76.2001 223.6 79.0001V75.0001H231.5V107H223.6ZM223.6 91.0001C223.6 85.6001 219.7 81.8001 214.9 81.8001C210.1 81.8001 206.2 85.6001 206.2 91.0001C206.2 96.2001 210.1 100.2 214.9 100.2C219.7 100.2 223.6 96.3001 223.6 91.0001Z" fill="#B2FCE4"/>
-    <path d="M239.1 118.7V111.7H241.7C243.9 111.7 246 110.6 246.6 107.3C246.8 106.3 246.7 105.8 246.3 104.8L234.5 75H242.8L251.3 97.1L259.6 75H267.7L255.8 105C251.9 114.9 250.8 118.8 241.5 118.8H239.1V118.7Z" fill="#B2FCE4"/>
-    <path d="M301.9 107V87.1001C301.9 83.6001 299.5 81.7001 296.3 81.7001C293 81.7001 290.4 83.8001 290.4 87.1001V107.1H282.3V75.0001H290.2V79.0001C291.4 77.1001 294.3 74.4001 299.1 74.4001C305.9 74.4001 310 78.8001 310 85.6001V107H301.9Z" fill="#B2FCE4"/>
-    <path d="M331.7 107.5C322.2 107.5 314.9 100.4 314.9 91.0001C314.9 81.6001 322.1 74.4001 331.7 74.4001C341.2 74.4001 348.5 81.6001 348.5 91.0001C348.4 100.4 341.1 107.5 331.7 107.5ZM340.3 91.0001C340.3 85.8001 336.4 81.9001 331.7 81.9001C327 81.9001 323 85.8001 323 91.0001C323 96.2001 327 100.1 331.7 100.1C336.4 100.1 340.3 96.1001 340.3 91.0001Z" fill="#B2FCE4"/>
-    <path d="M380.1 107L373.5 87.2H373.3L366.5 107H359.1L349.6 75H358L364 95.7H364.1L370.9 75H376.6L383.1 95.9H383.2L389.4 75H397.6L387.6 107H380.1Z" fill="#B2FCE4"/>
-    <path d="M440.6 107L434 87.2H433.8L427 107H419.6L410.1 75H418.5L424.5 95.7H424.6L431.4 75H437.1L443.6 95.9H443.7L449.9 75H458.1L448.1 107H440.6Z" fill="#B2FCE4"/>
-    <path d="M461.5 66.4001C461.5 64.0001 463.6 61.8 466.1 61.8C468.6 61.8 470.7 64.0001 470.7 66.4001C470.7 68.9001 468.6 71 466.1 71C463.6 71 461.5 68.9001 461.5 66.4001ZM462 107V75H470.1V107H462Z" fill="#B2FCE4"/>
-    <path d="M487.4 107C480.9 107 478.4 104.2 478.4 97.7001V81.9001H474.2V74.9001H478.4V67.1001H486.5V75.0001H491.6V82.0001H486.5V96.2001C486.5 98.9001 487.3 99.9001 490 99.9001H491.6V107H487.4Z" fill="#B2FCE4"/>
-    <path d="M515.7 107V87.1001C515.7 83.6001 513.3 81.7001 510.1 81.7001C506.8 81.7001 504.2 83.8001 504.2 87.1001V107.1H496.1V63.4001H504.2V78.6001C505.5 76.7001 508.3 74.4001 512.8 74.4001C519.6 74.4001 523.7 78.8001 523.7 85.6001V107H515.7Z" fill="#B2FCE4"/>
-    <path d="M1035.2 61L1020.3 52.5L1005.2 43.9C995.2 38.2 982.8 45.4 982.8 56.9V58.8C982.8 59.9 983.4 60.8 984.3 61.4L991.3 65.4C993.2 66.5 995.7 65.1 995.7 62.9V58.3C995.7 56 998.2 54.6 1000.1 55.7L1013.9 63.6L1027.6 71.5C1029.6 72.6 1029.6 75.5 1027.6 76.6L1013.9 84.5L1000.1 92.4C998.1 93.5 995.7 92.1 995.7 89.8V87.5C995.7 76 983.2 68.8 973.3 74.5L958.2 83.1L943.3 91.6C933.3 97.4 933.3 111.8 943.3 117.6L958.2 126.1L973.3 134.7C983.3 140.4 995.7 133.2 995.7 121.7V119.8C995.7 118.7 995.1 117.8 994.2 117.2L987.2 113.2C985.3 112.1 982.8 113.5 982.8 115.7V120.3C982.8 122.6 980.3 124 978.4 122.9L964.6 115L950.9 107.1C948.9 106 948.9 103.1 950.9 102L964.6 94.1L978.4 86.2C980.4 85.1 982.8 86.5 982.8 88.8V91.1C982.8 102.6 995.3 109.8 1005.2 104.1L1020.3 95.5L1035.2 87C1045.2 81.2 1045.2 66.8 1035.2 61Z" fill="#B2FCE4"/>
-    <path d="M937.6 63.8L902.7 135.8H888.2L901.3 108.9L880.8 63.8H895.6L908.8 94L923.2 63.8H937.6Z" fill="#B2FCE4"/>
-    <path d="M589.3 89.3999C589.3 80.7999 583.1 74.7999 575.4 74.7999C567.7 74.7999 561.5 80.8999 561.5 89.3999C561.5 97.7999 567.7 104 575.4 104C583.1 104 589.3 97.9999 589.3 89.3999ZM589.4 114.9V108.3C585.6 112.9 580 115.8 573.3 115.8C559.3 115.8 548.7 104.6 548.7 89.3999C548.7 74.3999 559.7 62.8999 573.6 62.8999C580.1 62.8999 585.7 65.7999 589.4 70.2999V63.8999H602V115L589.4 114.9Z" fill="#B2FCE4"/>
-    <path d="M663 103.6C658.6 103.6 657.4 102 657.4 97.7V75H665.5V63.8H657.4V51.4H644.5V63.9H627.9V58.8C627.9 54.5 629.5 52.9 634 52.9H636.8V43H630.6C620 43 615 46.5 615 57.1V63.9H607.8V75H615V115H627.9V75H644.5V100C644.5 110.4 648.5 114.9 658.9 114.9H665.5V103.6H663Z" fill="#B2FCE4"/>
-    <path d="M709.2 84.8C708.3 78.2 702.9 74.2 696.5 74.2C690.2 74.2 685 78.1 683.6 84.8H709.2ZM683.6 92.8C684.5 100.4 689.9 104.7 696.8 104.7C702.2 104.7 706.4 102.1 708.9 98.1H722.1C719 108.9 709.3 115.9 696.6 115.9C681.2 115.9 670.3 105.1 670.3 89.6C670.3 74.2 681.7 63 696.9 63C712.1 63 723.2 74.2 723.2 89.6C723.2 90.7 723.1 91.8 722.9 92.9H683.6V92.8Z" fill="#B2FCE4"/>
-    <path d="M804.9 89.4001C804.9 81.1001 798.7 74.8001 791 74.8001C783.3 74.8001 777.1 80.9001 777.1 89.4001C777.1 97.8 783.3 104 791 104C798.6 104 804.9 97.7001 804.9 89.4001ZM764.4 135.8V63.8H777V70.4001C780.8 65.7001 786.4 62.8 793.1 62.8C806.9 62.8 817.7 74.1 817.7 89.2C817.7 104.2 806.7 115.7 792.8 115.7C786.4 115.7 781 113.1 777.4 108.9V135.8H764.4Z" fill="#B2FCE4"/>
-    <path d="M863 89.3999C863 80.7999 856.8 74.7999 849.1 74.7999C841.4 74.7999 835.2 80.8999 835.2 89.3999C835.2 97.7999 841.4 104 849.1 104C856.8 104 863 97.9999 863 89.3999ZM863.2 114.9V108.3C859.4 112.9 853.8 115.8 847.1 115.8C833.1 115.8 822.5 104.6 822.5 89.3999C822.5 74.3999 833.5 62.8999 847.4 62.8999C853.9 62.8999 859.5 65.7999 863.2 70.2999V63.8999H875.8V115L863.2 114.9Z" fill="#B2FCE4"/>
-    <path d="M741.8 68.7999C741.8 68.7999 745 62.8999 752.8 62.8999C756.2 62.8999 758.3 64.0999 758.3 64.0999V77.0999C758.3 77.0999 753.6 74.1999 749.2 74.7999C744.9 75.3999 742.1 79.3999 742.1 84.6999V114.9H729.1V63.7999H741.7L741.8 68.7999Z" fill="#B2FCE4"/>
-    </g>
-    <defs>
-    <clipPath id="clip0">
-    <rect width="1200" height="180" fill="white"/>
-    </clipPath>
-    </defs>
-    </svg>
-    """
-  )
-
-  static let clearpayPayNowMintOnBlack = SVG(
-    size: CGSize(width: 1200, height: 180),
-    minimumWidth: 256,
-    svgString: """
-    <svg width="1200" height="180" viewBox="0 0 1200 180" fill="none" xmlns="http://www.w3.org/2000/svg">
-    <g clip-path="url(#clip0)">
-    <path d="M1169.8 180H30.2C13.6 180 0 166.4 0 149.8V30.2C0 13.6 13.6 0 30.2 0H1169.8C1186.4 0 1200 13.6 1200 30.2V149.8C1200 166.4 1186.4 180 1169.8 180Z" fill="black"/>
-    <path d="M944.4 64.4036L909.6 136.104H895.2L908.2 109.304L887.8 64.4036H902.6L915.7 94.5036L930 64.4036H944.4Z" fill="#B2FCE4"/>
-    <path d="M812.1 89.9037C812.1 81.6037 805.9 75.3037 798.2 75.3037C790.6 75.3037 784.3 81.4037 784.3 89.9037C784.3 98.3037 790.5 104.504 798.2 104.504C805.9 104.504 812.1 98.1037 812.1 89.9037ZM771.7 136.204V64.4037H784.2V71.0037C788 66.3037 793.6 63.5037 800.3 63.5037C814.1 63.5037 824.9 74.8037 824.9 89.8037C824.9 104.804 813.9 116.204 800 116.204C793.6 116.204 788.3 113.704 784.6 109.404V136.204H771.7Z" fill="#B2FCE4"/>
-    <path d="M870.1 89.9037C870.1 81.3037 863.9 75.3037 856.2 75.3037C848.6 75.3037 842.3 81.4037 842.3 89.9037C842.3 98.3037 848.5 104.504 856.2 104.504C863.9 104.504 870.1 98.4037 870.1 89.9037ZM870.2 115.404V108.804C866.4 113.404 860.8 116.204 854.1 116.204C840.1 116.204 829.5 105.004 829.5 89.9037C829.5 74.9037 840.5 63.5037 854.4 63.5037C860.9 63.5037 866.4 66.4037 870.2 70.8037V64.4037H882.7V115.304L870.2 115.404Z" fill="#B2FCE4"/>
-    <path d="M749.1 69.4037C749.1 69.4037 752.3 63.5037 760.1 63.5037C763.4 63.5037 765.6 64.7037 765.6 64.7037V77.7037C765.6 77.7037 760.9 74.8037 756.6 75.4037C752.3 76.0037 749.5 80.0037 749.5 85.3037V115.504H736.6V64.4037H749.1V69.4037Z" fill="#B2FCE4"/>
-    <path d="M1041.7 61.5037L1026.9 53.0037L1011.9 44.4037C1001.9 38.7037 989.5 45.9037 989.5 57.4037V59.3037C989.5 60.4037 990.1 61.3037 991 61.9037L998 65.9037C999.9 67.0037 1002.3 65.6037 1002.3 63.4037V58.8037C1002.3 56.5037 1004.8 55.1037 1006.7 56.2037L1020.4 64.1037L1034.1 72.0037C1036.1 73.1037 1036.1 76.0037 1034.1 77.1037L1020.4 85.0037L1006.7 92.9037C1004.7 94.0037 1002.3 92.6037 1002.3 90.3037V88.0037C1002.3 76.5037 989.9 69.3037 979.9 75.0037L964.9 83.6037L950 92.1037C940 97.8037 940 112.304 950 118.004L964.8 126.504L979.8 135.104C989.8 140.804 1002.2 133.604 1002.2 122.104V120.204C1002.2 119.104 1001.6 118.204 1000.7 117.604L993.7 113.604C991.8 112.504 989.4 113.904 989.4 116.104V120.704C989.4 123.004 986.9 124.404 985 123.304L971.3 115.404L957.6 107.504C955.6 106.404 955.6 103.504 957.6 102.404L971.3 94.5037L985 86.6037C987 85.5037 989.4 86.9037 989.4 89.2037V91.5037C989.4 103.004 1001.8 110.204 1011.8 104.504L1026.8 95.9037L1041.6 87.4037C1051.7 81.7037 1051.7 67.3037 1041.7 61.5037V61.5037Z" fill="#B2FCE4"/>
-    <path d="M592.6 96.4037C589.9 108.104 579.8 116.204 566.7 116.204C551.6 116.204 540.3 104.904 540.3 89.9037C540.3 74.9037 551.8 63.5037 566.9 63.5037C579.6 63.5037 589.6 71.5037 592.5 83.0037H579.3C577 78.4037 572.3 75.4037 567.1 75.4037C559.6 75.4037 553.3 81.6037 553.3 89.9037C553.3 98.1037 559.6 104.404 567.1 104.404C572.5 104.404 577.2 101.304 579.4 96.5037L592.6 96.4037Z" fill="#B2FCE4"/>
-    <path d="M598.6 115.304V43.5037H611.4V115.304H598.6Z" fill="#B2FCE4"/>
-    <path d="M630.7 93.2036C631.6 100.704 637 105.004 643.8 105.004C649.2 105.004 653.4 102.504 655.8 98.4036H669C665.9 109.204 656.3 116.104 643.5 116.104C628.1 116.104 617.3 105.304 617.3 89.9036C617.3 74.5036 628.7 63.4036 643.8 63.4036C659 63.5036 670 74.7036 670 90.0036C670 91.1036 669.9 92.2036 669.7 93.3036H630.7V93.2036ZM656.2 85.3036C655.3 78.7036 649.9 74.7036 643.6 74.7036C637.3 74.7036 632.1 78.6036 630.8 85.3036H656.2Z" fill="#B2FCE4"/>
-    <path d="M715.7 115.304V108.704C711.9 113.304 706.3 116.104 699.6 116.104C685.7 116.104 675.1 104.904 675.1 89.8036C675.1 74.8036 686.1 63.4036 699.9 63.4036C706.4 63.4036 711.9 66.3036 715.7 70.7036V64.3036H728.2V115.204L715.7 115.304ZM715.6 89.9036C715.6 81.3036 709.4 75.3036 701.8 75.3036C694.2 75.3036 688 81.4036 688 89.9036C688 98.2036 694.2 104.504 701.8 104.504C709.4 104.504 715.6 98.4036 715.6 89.9036Z" fill="#B2FCE4"/>
-    <path d="M150 111.204V67.6038H169.6C179.1 67.6038 185.2 72.6038 185.2 81.2038C185.2 89.9038 178.4 95.2038 169.3 95.2038H158.5V111.204H150ZM169.1 88.3038C173.6 88.3038 176.6 85.2038 176.6 81.3038C176.6 77.6038 173.6 74.8038 169.2 74.8038H158.5V88.4038H169.1V88.3038Z" fill="#B2FCE4"/>
-    <path d="M212.4 111.204V107.004C210 109.904 206.5 111.704 202.3 111.704C193.5 111.704 186.9 104.704 186.9 95.2038C186.9 85.8038 193.8 78.6038 202.5 78.6038C206.6 78.6038 210.1 80.4038 212.4 83.2038V79.2038H220.3V111.204H212.4ZM212.4 95.2038C212.4 89.8038 208.5 86.0038 203.7 86.0038C198.9 86.0038 195 89.8038 195 95.2038C195 100.404 198.9 104.404 203.7 104.404C208.5 104.404 212.4 100.504 212.4 95.2038Z" fill="#B2FCE4"/>
-    <path d="M227.9 122.904V115.904H230.5C232.7 115.904 234.8 114.804 235.4 111.504C235.6 110.504 235.5 110.004 235.1 109.004L223.3 79.2036H231.6L240.1 101.304L248.4 79.2036H256.5L244.6 109.204C240.7 119.104 239.6 123.004 230.3 123.004H227.9V122.904Z" fill="#B2FCE4"/>
-    <path d="M290.7 111.204V91.3038C290.7 87.8038 288.3 85.9038 285.1 85.9038C281.8 85.9038 279.2 88.0038 279.2 91.3038V111.304H271.1V79.2038H279V83.2038C280.2 81.3038 283.1 78.6038 287.9 78.6038C294.7 78.6038 298.8 83.0038 298.8 89.8038V111.204H290.7Z" fill="#B2FCE4"/>
-    <path d="M320.5 111.704C311 111.704 303.7 104.604 303.7 95.2038C303.7 85.8038 310.9 78.6038 320.5 78.6038C330 78.6038 337.3 85.8038 337.3 95.2038C337.2 104.604 329.9 111.704 320.5 111.704ZM329.1 95.2038C329.1 90.0038 325.2 86.1038 320.5 86.1038C315.8 86.1038 311.8 90.0038 311.8 95.2038C311.8 100.404 315.8 104.304 320.5 104.304C325.2 104.304 329.1 100.304 329.1 95.2038Z" fill="#B2FCE4"/>
-    <path d="M368.9 111.204L362.3 91.4036H362.1L355.3 111.204H347.9L338.4 79.2036H346.8L352.8 99.9036H352.9L359.7 79.2036H365.4L371.9 100.104H372L378.2 79.2036H386.4L376.4 111.204H368.9Z" fill="#B2FCE4"/>
-    <path d="M429.4 111.204L422.8 91.4036H422.6L415.8 111.204H408.4L398.9 79.2036H407.3L413.3 99.9036H413.4L420.2 79.2036H425.9L432.4 100.104H432.5L438.7 79.2036H446.9L436.9 111.204H429.4Z" fill="#B2FCE4"/>
-    <path d="M450.3 70.6037C450.3 68.2037 452.4 66.0037 454.9 66.0037C457.4 66.0037 459.5 68.2037 459.5 70.6037C459.5 73.1037 457.4 75.2037 454.9 75.2037C452.4 75.2037 450.3 73.1037 450.3 70.6037ZM450.8 111.204V79.2037H458.9V111.204H450.8Z" fill="#B2FCE4"/>
-    <path d="M476.2 111.204C469.7 111.204 467.2 108.404 467.2 101.904V86.1037H463V79.1037H467.2V71.3037H475.3V79.2037H480.4V86.2037H475.3V100.404C475.3 103.104 476.1 104.104 478.8 104.104H480.4V111.204H476.2Z" fill="#B2FCE4"/>
-    <path d="M504.5 111.204V91.3038C504.5 87.8038 502.1 85.9038 498.9 85.9038C495.6 85.9038 493 88.0038 493 91.3038V111.304H484.9V67.6038H493V82.8038C494.3 80.9038 497.1 78.6038 501.6 78.6038C508.4 78.6038 512.5 83.0038 512.5 89.8038V111.204H504.5Z" fill="#B2FCE4"/>
-    </g>
-    <defs>
-    <clipPath id="clip0">
-    <rect width="1200" height="180" fill="white"/>
-    </clipPath>
-    </defs>
-    </svg>
-    """
-  )
-
-  static let afterpayPayNowWhiteOnBlack = SVG(
-    size: CGSize(width: 1200, height: 180),
-    minimumWidth: 256,
-    svgString: """
-    <svg width="1200" height="180" viewBox="0 0 1200 180" fill="none" xmlns="http://www.w3.org/2000/svg">
-    <g clip-path="url(#clip0)">
-    <path d="M1169.8 180H30.2C13.6 180 0 166.4 0 149.8V30.2C0 13.6 13.6 0 30.2 0H1169.8C1186.4 0 1200 13.6 1200 30.2V149.8C1200 166.4 1186.4 180 1169.8 180Z" fill="black"/>
-    <path d="M161.2 107V63.4001H180.8C190.3 63.4001 196.4 68.4001 196.4 77.0001C196.4 85.7001 189.6 91.0001 180.5 91.0001H169.7V107H161.2ZM180.3 84.1001C184.8 84.1001 187.8 81.0001 187.8 77.1001C187.8 73.4001 184.8 70.6001 180.4 70.6001H169.7V84.2001H180.3V84.1001Z" fill="white"/>
-    <path d="M223.6 107V102.8C221.2 105.7 217.7 107.5 213.5 107.5C204.7 107.5 198.1 100.5 198.1 91.0001C198.1 81.6001 205 74.4001 213.7 74.4001C217.8 74.4001 221.3 76.2001 223.6 79.0001V75.0001H231.5V107H223.6ZM223.6 91.0001C223.6 85.6001 219.7 81.8001 214.9 81.8001C210.1 81.8001 206.2 85.6001 206.2 91.0001C206.2 96.2001 210.1 100.2 214.9 100.2C219.7 100.2 223.6 96.3001 223.6 91.0001Z" fill="white"/>
-    <path d="M239.1 118.7V111.7H241.7C243.9 111.7 246 110.6 246.6 107.3C246.8 106.3 246.7 105.8 246.3 104.8L234.5 75H242.8L251.3 97.1L259.6 75H267.7L255.8 105C251.9 114.9 250.8 118.8 241.5 118.8H239.1V118.7Z" fill="white"/>
-    <path d="M301.9 107V87.1001C301.9 83.6001 299.5 81.7001 296.3 81.7001C293 81.7001 290.4 83.8001 290.4 87.1001V107.1H282.3V75.0001H290.2V79.0001C291.4 77.1001 294.3 74.4001 299.1 74.4001C305.9 74.4001 310 78.8001 310 85.6001V107H301.9Z" fill="white"/>
-    <path d="M331.7 107.5C322.2 107.5 314.9 100.4 314.9 91.0001C314.9 81.6001 322.1 74.4001 331.7 74.4001C341.2 74.4001 348.5 81.6001 348.5 91.0001C348.4 100.4 341.1 107.5 331.7 107.5ZM340.3 91.0001C340.3 85.8001 336.4 81.9001 331.7 81.9001C327 81.9001 323 85.8001 323 91.0001C323 96.2001 327 100.1 331.7 100.1C336.4 100.1 340.3 96.1001 340.3 91.0001Z" fill="white"/>
-    <path d="M380.1 107L373.5 87.2H373.3L366.5 107H359.1L349.6 75H358L364 95.7H364.1L370.9 75H376.6L383.1 95.9H383.2L389.4 75H397.6L387.6 107H380.1Z" fill="white"/>
-    <path d="M440.6 107L434 87.2H433.8L427 107H419.6L410.1 75H418.5L424.5 95.7H424.6L431.4 75H437.1L443.6 95.9H443.7L449.9 75H458.1L448.1 107H440.6Z" fill="white"/>
-    <path d="M461.5 66.4001C461.5 64.0001 463.6 61.8 466.1 61.8C468.6 61.8 470.7 64.0001 470.7 66.4001C470.7 68.9001 468.6 71 466.1 71C463.6 71 461.5 68.9001 461.5 66.4001ZM462 107V75H470.1V107H462Z" fill="white"/>
-    <path d="M487.4 107C480.9 107 478.4 104.2 478.4 97.7001V81.9001H474.2V74.9001H478.4V67.1001H486.5V75.0001H491.6V82.0001H486.5V96.2001C486.5 98.9001 487.3 99.9001 490 99.9001H491.6V107H487.4Z" fill="white"/>
-    <path d="M515.7 107V87.1001C515.7 83.6001 513.3 81.7001 510.1 81.7001C506.8 81.7001 504.2 83.8001 504.2 87.1001V107.1H496.1V63.4001H504.2V78.6001C505.5 76.7001 508.3 74.4001 512.8 74.4001C519.6 74.4001 523.7 78.8001 523.7 85.6001V107H515.7Z" fill="white"/>
-    <path d="M1035.2 61L1020.3 52.5L1005.2 43.9C995.2 38.2 982.8 45.4 982.8 56.9V58.8C982.8 59.9 983.4 60.8 984.3 61.4L991.3 65.4C993.2 66.5 995.7 65.1 995.7 62.9V58.3C995.7 56 998.2 54.6 1000.1 55.7L1013.9 63.6L1027.6 71.5C1029.6 72.6 1029.6 75.5 1027.6 76.6L1013.9 84.5L1000.1 92.4C998.1 93.5 995.7 92.1 995.7 89.8V87.5C995.7 76 983.2 68.8 973.3 74.5L958.2 83.1L943.3 91.6C933.3 97.4 933.3 111.8 943.3 117.6L958.2 126.1L973.3 134.7C983.3 140.4 995.7 133.2 995.7 121.7V119.8C995.7 118.7 995.1 117.8 994.2 117.2L987.2 113.2C985.3 112.1 982.8 113.5 982.8 115.7V120.3C982.8 122.6 980.3 124 978.4 122.9L964.6 115L950.9 107.1C948.9 106 948.9 103.1 950.9 102L964.6 94.1L978.4 86.2C980.4 85.1 982.8 86.5 982.8 88.8V91.1C982.8 102.6 995.3 109.8 1005.2 104.1L1020.3 95.5L1035.2 87C1045.2 81.2 1045.2 66.8 1035.2 61Z" fill="white"/>
-    <path d="M937.6 63.8L902.7 135.8H888.2L901.3 108.9L880.8 63.8H895.6L908.8 94L923.2 63.8H937.6Z" fill="white"/>
-    <path d="M589.3 89.3999C589.3 80.7999 583.1 74.7999 575.4 74.7999C567.7 74.7999 561.5 80.8999 561.5 89.3999C561.5 97.7999 567.7 104 575.4 104C583.1 104 589.3 97.9999 589.3 89.3999ZM589.4 114.9V108.3C585.6 112.9 580 115.8 573.3 115.8C559.3 115.8 548.7 104.6 548.7 89.3999C548.7 74.3999 559.7 62.8999 573.6 62.8999C580.1 62.8999 585.7 65.7999 589.4 70.2999V63.8999H602V115L589.4 114.9Z" fill="white"/>
-    <path d="M663 103.6C658.6 103.6 657.4 102 657.4 97.7V75H665.5V63.8H657.4V51.4H644.5V63.9H627.9V58.8C627.9 54.5 629.5 52.9 634 52.9H636.8V43H630.6C620 43 615 46.5 615 57.1V63.9H607.8V75H615V115H627.9V75H644.5V100C644.5 110.4 648.5 114.9 658.9 114.9H665.5V103.6H663Z" fill="white"/>
-    <path d="M709.2 84.8C708.3 78.2 702.9 74.2 696.5 74.2C690.2 74.2 685 78.1 683.6 84.8H709.2ZM683.6 92.8C684.5 100.4 689.9 104.7 696.8 104.7C702.2 104.7 706.4 102.1 708.9 98.1H722.1C719 108.9 709.3 115.9 696.6 115.9C681.2 115.9 670.3 105.1 670.3 89.6C670.3 74.2 681.7 63 696.9 63C712.1 63 723.2 74.2 723.2 89.6C723.2 90.7 723.1 91.8 722.9 92.9H683.6V92.8Z" fill="white"/>
-    <path d="M804.9 89.4001C804.9 81.1001 798.7 74.8001 791 74.8001C783.3 74.8001 777.1 80.9001 777.1 89.4001C777.1 97.8 783.3 104 791 104C798.6 104 804.9 97.7001 804.9 89.4001ZM764.4 135.8V63.8H777V70.4001C780.8 65.7001 786.4 62.8 793.1 62.8C806.9 62.8 817.7 74.1 817.7 89.2C817.7 104.2 806.7 115.7 792.8 115.7C786.4 115.7 781 113.1 777.4 108.9V135.8H764.4Z" fill="white"/>
-    <path d="M863 89.3999C863 80.7999 856.8 74.7999 849.1 74.7999C841.4 74.7999 835.2 80.8999 835.2 89.3999C835.2 97.7999 841.4 104 849.1 104C856.8 104 863 97.9999 863 89.3999ZM863.2 114.9V108.3C859.4 112.9 853.8 115.8 847.1 115.8C833.1 115.8 822.5 104.6 822.5 89.3999C822.5 74.3999 833.5 62.8999 847.4 62.8999C853.9 62.8999 859.5 65.7999 863.2 70.2999V63.8999H875.8V115L863.2 114.9Z" fill="white"/>
-    <path d="M741.8 68.7999C741.8 68.7999 745 62.8999 752.8 62.8999C756.2 62.8999 758.3 64.0999 758.3 64.0999V77.0999C758.3 77.0999 753.6 74.1999 749.2 74.7999C744.9 75.3999 742.1 79.3999 742.1 84.6999V114.9H729.1V63.7999H741.7L741.8 68.7999Z" fill="white"/>
-    </g>
-    <defs>
-    <clipPath id="clip0">
-    <rect width="1200" height="180" fill="white"/>
-    </clipPath>
-    </defs>
-    </svg>
-    """
-  )
-
-  static let clearpayPayNowWhiteOnBlack = SVG(
-    size: CGSize(width: 1200, height: 180),
-    minimumWidth: 256,
-    svgString: """
-    <svg width="1200" height="180" viewBox="0 0 1200 180" fill="none" xmlns="http://www.w3.org/2000/svg">
-    <g clip-path="url(#clip0)">
-    <path d="M1169.8 180H30.2C13.6 180 0 166.4 0 149.8V30.2C0 13.6 13.6 0 30.2 0H1169.8C1186.4 0 1200 13.6 1200 30.2V149.8C1200 166.4 1186.4 180 1169.8 180Z" fill="black"/>
-    <path d="M944.4 64.4036L909.6 136.104H895.2L908.2 109.304L887.8 64.4036H902.6L915.7 94.5036L930 64.4036H944.4Z" fill="white"/>
-    <path d="M812.1 89.9037C812.1 81.6037 805.9 75.3037 798.2 75.3037C790.6 75.3037 784.3 81.4037 784.3 89.9037C784.3 98.3037 790.5 104.504 798.2 104.504C805.9 104.504 812.1 98.1037 812.1 89.9037ZM771.7 136.204V64.4037H784.2V71.0037C788 66.3037 793.6 63.5037 800.3 63.5037C814.1 63.5037 824.9 74.8037 824.9 89.8037C824.9 104.804 813.9 116.204 800 116.204C793.6 116.204 788.3 113.704 784.6 109.404V136.204H771.7Z" fill="white"/>
-    <path d="M870.1 89.9037C870.1 81.3037 863.9 75.3037 856.2 75.3037C848.6 75.3037 842.3 81.4037 842.3 89.9037C842.3 98.3037 848.5 104.504 856.2 104.504C863.9 104.504 870.1 98.4037 870.1 89.9037ZM870.2 115.404V108.804C866.4 113.404 860.8 116.204 854.1 116.204C840.1 116.204 829.5 105.004 829.5 89.9037C829.5 74.9037 840.5 63.5037 854.4 63.5037C860.9 63.5037 866.4 66.4037 870.2 70.8037V64.4037H882.7V115.304L870.2 115.404Z" fill="white"/>
-    <path d="M749.1 69.4037C749.1 69.4037 752.3 63.5037 760.1 63.5037C763.4 63.5037 765.6 64.7037 765.6 64.7037V77.7037C765.6 77.7037 760.9 74.8037 756.6 75.4037C752.3 76.0037 749.5 80.0037 749.5 85.3037V115.504H736.6V64.4037H749.1V69.4037Z" fill="white"/>
-    <path d="M1041.7 61.5037L1026.9 53.0037L1011.9 44.4037C1001.9 38.7037 989.5 45.9037 989.5 57.4037V59.3037C989.5 60.4037 990.1 61.3037 991 61.9037L998 65.9037C999.9 67.0037 1002.3 65.6037 1002.3 63.4037V58.8037C1002.3 56.5037 1004.8 55.1037 1006.7 56.2037L1020.4 64.1037L1034.1 72.0037C1036.1 73.1037 1036.1 76.0037 1034.1 77.1037L1020.4 85.0037L1006.7 92.9037C1004.7 94.0037 1002.3 92.6037 1002.3 90.3037V88.0037C1002.3 76.5037 989.9 69.3037 979.9 75.0037L964.9 83.6037L950 92.1037C940 97.8037 940 112.304 950 118.004L964.8 126.504L979.8 135.104C989.8 140.804 1002.2 133.604 1002.2 122.104V120.204C1002.2 119.104 1001.6 118.204 1000.7 117.604L993.7 113.604C991.8 112.504 989.4 113.904 989.4 116.104V120.704C989.4 123.004 986.9 124.404 985 123.304L971.3 115.404L957.6 107.504C955.6 106.404 955.6 103.504 957.6 102.404L971.3 94.5037L985 86.6037C987 85.5037 989.4 86.9037 989.4 89.2037V91.5037C989.4 103.004 1001.8 110.204 1011.8 104.504L1026.8 95.9037L1041.6 87.4037C1051.7 81.7037 1051.7 67.3037 1041.7 61.5037V61.5037Z" fill="white"/>
-    <path d="M592.6 96.4037C589.9 108.104 579.8 116.204 566.7 116.204C551.6 116.204 540.3 104.904 540.3 89.9037C540.3 74.9037 551.8 63.5037 566.9 63.5037C579.6 63.5037 589.6 71.5037 592.5 83.0037H579.3C577 78.4037 572.3 75.4037 567.1 75.4037C559.6 75.4037 553.3 81.6037 553.3 89.9037C553.3 98.1037 559.6 104.404 567.1 104.404C572.5 104.404 577.2 101.304 579.4 96.5037L592.6 96.4037Z" fill="white"/>
-    <path d="M598.6 115.304V43.5037H611.4V115.304H598.6Z" fill="white"/>
-    <path d="M630.7 93.2036C631.6 100.704 637 105.004 643.8 105.004C649.2 105.004 653.4 102.504 655.8 98.4036H669C665.9 109.204 656.3 116.104 643.5 116.104C628.1 116.104 617.3 105.304 617.3 89.9036C617.3 74.5036 628.7 63.4036 643.8 63.4036C659 63.5036 670 74.7036 670 90.0036C670 91.1036 669.9 92.2036 669.7 93.3036H630.7V93.2036ZM656.2 85.3036C655.3 78.7036 649.9 74.7036 643.6 74.7036C637.3 74.7036 632.1 78.6036 630.8 85.3036H656.2Z" fill="white"/>
-    <path d="M715.7 115.304V108.704C711.9 113.304 706.3 116.104 699.6 116.104C685.7 116.104 675.1 104.904 675.1 89.8036C675.1 74.8036 686.1 63.4036 699.9 63.4036C706.4 63.4036 711.9 66.3036 715.7 70.7036V64.3036H728.2V115.204L715.7 115.304ZM715.6 89.9036C715.6 81.3036 709.4 75.3036 701.8 75.3036C694.2 75.3036 688 81.4036 688 89.9036C688 98.2036 694.2 104.504 701.8 104.504C709.4 104.504 715.6 98.4036 715.6 89.9036Z" fill="white"/>
-    <path d="M150 111.204V67.6038H169.6C179.1 67.6038 185.2 72.6038 185.2 81.2038C185.2 89.9038 178.4 95.2038 169.3 95.2038H158.5V111.204H150ZM169.1 88.3038C173.6 88.3038 176.6 85.2038 176.6 81.3038C176.6 77.6038 173.6 74.8038 169.2 74.8038H158.5V88.4038H169.1V88.3038Z" fill="white"/>
-    <path d="M212.4 111.204V107.004C210 109.904 206.5 111.704 202.3 111.704C193.5 111.704 186.9 104.704 186.9 95.2038C186.9 85.8038 193.8 78.6038 202.5 78.6038C206.6 78.6038 210.1 80.4038 212.4 83.2038V79.2038H220.3V111.204H212.4ZM212.4 95.2038C212.4 89.8038 208.5 86.0038 203.7 86.0038C198.9 86.0038 195 89.8038 195 95.2038C195 100.404 198.9 104.404 203.7 104.404C208.5 104.404 212.4 100.504 212.4 95.2038Z" fill="white"/>
-    <path d="M227.9 122.904V115.904H230.5C232.7 115.904 234.8 114.804 235.4 111.504C235.6 110.504 235.5 110.004 235.1 109.004L223.3 79.2036H231.6L240.1 101.304L248.4 79.2036H256.5L244.6 109.204C240.7 119.104 239.6 123.004 230.3 123.004H227.9V122.904Z" fill="white"/>
-    <path d="M290.7 111.204V91.3038C290.7 87.8038 288.3 85.9038 285.1 85.9038C281.8 85.9038 279.2 88.0038 279.2 91.3038V111.304H271.1V79.2038H279V83.2038C280.2 81.3038 283.1 78.6038 287.9 78.6038C294.7 78.6038 298.8 83.0038 298.8 89.8038V111.204H290.7Z" fill="white"/>
-    <path d="M320.5 111.704C311 111.704 303.7 104.604 303.7 95.2038C303.7 85.8038 310.9 78.6038 320.5 78.6038C330 78.6038 337.3 85.8038 337.3 95.2038C337.2 104.604 329.9 111.704 320.5 111.704ZM329.1 95.2038C329.1 90.0038 325.2 86.1038 320.5 86.1038C315.8 86.1038 311.8 90.0038 311.8 95.2038C311.8 100.404 315.8 104.304 320.5 104.304C325.2 104.304 329.1 100.304 329.1 95.2038Z" fill="white"/>
-    <path d="M368.9 111.204L362.3 91.4036H362.1L355.3 111.204H347.9L338.4 79.2036H346.8L352.8 99.9036H352.9L359.7 79.2036H365.4L371.9 100.104H372L378.2 79.2036H386.4L376.4 111.204H368.9Z" fill="white"/>
-    <path d="M429.4 111.204L422.8 91.4036H422.6L415.8 111.204H408.4L398.9 79.2036H407.3L413.3 99.9036H413.4L420.2 79.2036H425.9L432.4 100.104H432.5L438.7 79.2036H446.9L436.9 111.204H429.4Z" fill="white"/>
-    <path d="M450.3 70.6037C450.3 68.2037 452.4 66.0037 454.9 66.0037C457.4 66.0037 459.5 68.2037 459.5 70.6037C459.5 73.1037 457.4 75.2037 454.9 75.2037C452.4 75.2037 450.3 73.1037 450.3 70.6037ZM450.8 111.204V79.2037H458.9V111.204H450.8Z" fill="white"/>
-    <path d="M476.2 111.204C469.7 111.204 467.2 108.404 467.2 101.904V86.1037H463V79.1037H467.2V71.3037H475.3V79.2037H480.4V86.2037H475.3V100.404C475.3 103.104 476.1 104.104 478.8 104.104H480.4V111.204H476.2Z" fill="white"/>
-    <path d="M504.5 111.204V91.3038C504.5 87.8038 502.1 85.9038 498.9 85.9038C495.6 85.9038 493 88.0038 493 91.3038V111.304H484.9V67.6038H493V82.8038C494.3 80.9038 497.1 78.6038 501.6 78.6038C508.4 78.6038 512.5 83.0038 512.5 89.8038V111.204H504.5Z" fill="white"/>
-    </g>
-    <defs>
-    <clipPath id="clip0">
-    <rect width="1200" height="180" fill="white"/>
-    </clipPath>
-    </defs>
-    </svg>
-    """
-  )
-
-  static let afterpayPayNowBlackOnWhite = SVG(
-    size: CGSize(width: 1200, height: 180),
-    minimumWidth: 256,
-    svgString: """
-    <svg width="1200" height="180" viewBox="0 0 1200 180" fill="none" xmlns="http://www.w3.org/2000/svg">
-    <g clip-path="url(#clip0)">
-    <path d="M1169.8 180H30.2C13.6 180 0 166.4 0 149.8V30.2C0 13.6 13.6 0 30.2 0H1169.8C1186.4 0 1200 13.6 1200 30.2V149.8C1200 166.4 1186.4 180 1169.8 180Z" fill="white"/>
-    <path d="M161.2 107V63.4001H180.8C190.3 63.4001 196.4 68.4001 196.4 77.0001C196.4 85.7001 189.6 91.0001 180.5 91.0001H169.7V107H161.2ZM180.3 84.1001C184.8 84.1001 187.8 81.0001 187.8 77.1001C187.8 73.4001 184.8 70.6001 180.4 70.6001H169.7V84.2001H180.3V84.1001Z" fill="black"/>
-    <path d="M223.6 107V102.8C221.2 105.7 217.7 107.5 213.5 107.5C204.7 107.5 198.1 100.5 198.1 91.0001C198.1 81.6001 205 74.4001 213.7 74.4001C217.8 74.4001 221.3 76.2001 223.6 79.0001V75.0001H231.5V107H223.6ZM223.6 91.0001C223.6 85.6001 219.7 81.8001 214.9 81.8001C210.1 81.8001 206.2 85.6001 206.2 91.0001C206.2 96.2001 210.1 100.2 214.9 100.2C219.7 100.2 223.6 96.3001 223.6 91.0001Z" fill="black"/>
-    <path d="M239.1 118.7V111.7H241.7C243.9 111.7 246 110.6 246.6 107.3C246.8 106.3 246.7 105.8 246.3 104.8L234.5 75H242.8L251.3 97.1L259.6 75H267.7L255.8 105C251.9 114.9 250.8 118.8 241.5 118.8H239.1V118.7Z" fill="black"/>
-    <path d="M301.9 107V87.1001C301.9 83.6001 299.5 81.7001 296.3 81.7001C293 81.7001 290.4 83.8001 290.4 87.1001V107.1H282.3V75.0001H290.2V79.0001C291.4 77.1001 294.3 74.4001 299.1 74.4001C305.9 74.4001 310 78.8001 310 85.6001V107H301.9Z" fill="black"/>
-    <path d="M331.7 107.5C322.2 107.5 314.9 100.4 314.9 91.0001C314.9 81.6001 322.1 74.4001 331.7 74.4001C341.2 74.4001 348.5 81.6001 348.5 91.0001C348.4 100.4 341.1 107.5 331.7 107.5ZM340.3 91.0001C340.3 85.8001 336.4 81.9001 331.7 81.9001C327 81.9001 323 85.8001 323 91.0001C323 96.2001 327 100.1 331.7 100.1C336.4 100.1 340.3 96.1001 340.3 91.0001Z" fill="black"/>
-    <path d="M380.1 107L373.5 87.2H373.3L366.5 107H359.1L349.6 75H358L364 95.7H364.1L370.9 75H376.6L383.1 95.9H383.2L389.4 75H397.6L387.6 107H380.1Z" fill="black"/>
-    <path d="M440.6 107L434 87.2H433.8L427 107H419.6L410.1 75H418.5L424.5 95.7H424.6L431.4 75H437.1L443.6 95.9H443.7L449.9 75H458.1L448.1 107H440.6Z" fill="black"/>
-    <path d="M461.5 66.4001C461.5 64.0001 463.6 61.8 466.1 61.8C468.6 61.8 470.7 64.0001 470.7 66.4001C470.7 68.9001 468.6 71 466.1 71C463.6 71 461.5 68.9001 461.5 66.4001ZM462 107V75H470.1V107H462Z" fill="black"/>
-    <path d="M487.4 107C480.9 107 478.4 104.2 478.4 97.7001V81.9001H474.2V74.9001H478.4V67.1001H486.5V75.0001H491.6V82.0001H486.5V96.2001C486.5 98.9001 487.3 99.9001 490 99.9001H491.6V107H487.4Z" fill="black"/>
-    <path d="M515.7 107V87.1001C515.7 83.6001 513.3 81.7001 510.1 81.7001C506.8 81.7001 504.2 83.8001 504.2 87.1001V107.1H496.1V63.4001H504.2V78.6001C505.5 76.7001 508.3 74.4001 512.8 74.4001C519.6 74.4001 523.7 78.8001 523.7 85.6001V107H515.7Z" fill="black"/>
-    <path d="M1035.2 61L1020.3 52.5L1005.2 43.9C995.2 38.2 982.8 45.4 982.8 56.9V58.8C982.8 59.9 983.4 60.8 984.3 61.4L991.3 65.4C993.2 66.5 995.7 65.1 995.7 62.9V58.3C995.7 56 998.2 54.6 1000.1 55.7L1013.9 63.6L1027.6 71.5C1029.6 72.6 1029.6 75.5 1027.6 76.6L1013.9 84.5L1000.1 92.4C998.1 93.5 995.7 92.1 995.7 89.8V87.5C995.7 76 983.2 68.8 973.3 74.5L958.2 83.1L943.3 91.6C933.3 97.4 933.3 111.8 943.3 117.6L958.2 126.1L973.3 134.7C983.3 140.4 995.7 133.2 995.7 121.7V119.8C995.7 118.7 995.1 117.8 994.2 117.2L987.2 113.2C985.3 112.1 982.8 113.5 982.8 115.7V120.3C982.8 122.6 980.3 124 978.4 122.9L964.6 115L950.9 107.1C948.9 106 948.9 103.1 950.9 102L964.6 94.1L978.4 86.2C980.4 85.1 982.8 86.5 982.8 88.8V91.1C982.8 102.6 995.3 109.8 1005.2 104.1L1020.3 95.5L1035.2 87C1045.2 81.2 1045.2 66.8 1035.2 61Z" fill="black"/>
-    <path d="M937.6 63.8L902.7 135.8H888.2L901.3 108.9L880.8 63.8H895.6L908.8 94L923.2 63.8H937.6Z" fill="black"/>
-    <path d="M589.3 89.3999C589.3 80.7999 583.1 74.7999 575.4 74.7999C567.7 74.7999 561.5 80.8999 561.5 89.3999C561.5 97.7999 567.7 104 575.4 104C583.1 104 589.3 97.9999 589.3 89.3999ZM589.4 114.9V108.3C585.6 112.9 580 115.8 573.3 115.8C559.3 115.8 548.7 104.6 548.7 89.3999C548.7 74.3999 559.7 62.8999 573.6 62.8999C580.1 62.8999 585.7 65.7999 589.4 70.2999V63.8999H602V115L589.4 114.9Z" fill="black"/>
-    <path d="M663 103.6C658.6 103.6 657.4 102 657.4 97.7V75H665.5V63.8H657.4V51.4H644.5V63.9H627.9V58.8C627.9 54.5 629.5 52.9 634 52.9H636.8V43H630.6C620 43 615 46.5 615 57.1V63.9H607.8V75H615V115H627.9V75H644.5V100C644.5 110.4 648.5 114.9 658.9 114.9H665.5V103.6H663Z" fill="black"/>
-    <path d="M709.2 84.8C708.3 78.2 702.9 74.2 696.5 74.2C690.2 74.2 685 78.1 683.6 84.8H709.2ZM683.6 92.8C684.5 100.4 689.9 104.7 696.8 104.7C702.2 104.7 706.4 102.1 708.9 98.1H722.1C719 108.9 709.3 115.9 696.6 115.9C681.2 115.9 670.3 105.1 670.3 89.6C670.3 74.2 681.7 63 696.9 63C712.1 63 723.2 74.2 723.2 89.6C723.2 90.7 723.1 91.8 722.9 92.9H683.6V92.8Z" fill="black"/>
-    <path d="M804.9 89.4001C804.9 81.1001 798.7 74.8001 791 74.8001C783.3 74.8001 777.1 80.9001 777.1 89.4001C777.1 97.8 783.3 104 791 104C798.6 104 804.9 97.7001 804.9 89.4001ZM764.4 135.8V63.8H777V70.4001C780.8 65.7001 786.4 62.8 793.1 62.8C806.9 62.8 817.7 74.1 817.7 89.2C817.7 104.2 806.7 115.7 792.8 115.7C786.4 115.7 781 113.1 777.4 108.9V135.8H764.4Z" fill="black"/>
-    <path d="M863 89.3999C863 80.7999 856.8 74.7999 849.1 74.7999C841.4 74.7999 835.2 80.8999 835.2 89.3999C835.2 97.7999 841.4 104 849.1 104C856.8 104 863 97.9999 863 89.3999ZM863.2 114.9V108.3C859.4 112.9 853.8 115.8 847.1 115.8C833.1 115.8 822.5 104.6 822.5 89.3999C822.5 74.3999 833.5 62.8999 847.4 62.8999C853.9 62.8999 859.5 65.7999 863.2 70.2999V63.8999H875.8V115L863.2 114.9Z" fill="black"/>
-    <path d="M741.8 68.7999C741.8 68.7999 745 62.8999 752.8 62.8999C756.2 62.8999 758.3 64.0999 758.3 64.0999V77.0999C758.3 77.0999 753.6 74.1999 749.2 74.7999C744.9 75.3999 742.1 79.3999 742.1 84.6999V114.9H729.1V63.7999H741.7L741.8 68.7999Z" fill="black"/>
-    </g>
-    <defs>
-    <clipPath id="clip0">
-    <rect width="1200" height="180" fill="white"/>
-    </clipPath>
-    </defs>
-    </svg>
-    """
-  )
-
-  static let clearpayPayNowBlackOnWhite = SVG(
-    size: CGSize(width: 1200, height: 180),
-    minimumWidth: 256,
-    svgString: """
-    <svg width="1200" height="180" viewBox="0 0 1200 180" fill="none" xmlns="http://www.w3.org/2000/svg">
-    <g clip-path="url(#clip0)">
-    <path d="M1169.8 180H30.2C13.6 180 0 166.4 0 149.8V30.2C0 13.6 13.6 0 30.2 0H1169.8C1186.4 0 1200 13.6 1200 30.2V149.8C1200 166.4 1186.4 180 1169.8 180Z" fill="white"/>
-    <path d="M944.4 64.4036L909.6 136.104H895.2L908.2 109.304L887.8 64.4036H902.6L915.7 94.5036L930 64.4036H944.4Z" fill="black"/>
-    <path d="M812.1 89.9037C812.1 81.6037 805.9 75.3037 798.2 75.3037C790.6 75.3037 784.3 81.4037 784.3 89.9037C784.3 98.3037 790.5 104.504 798.2 104.504C805.9 104.504 812.1 98.1037 812.1 89.9037ZM771.7 136.204V64.4037H784.2V71.0037C788 66.3037 793.6 63.5037 800.3 63.5037C814.1 63.5037 824.9 74.8037 824.9 89.8037C824.9 104.804 813.9 116.204 800 116.204C793.6 116.204 788.3 113.704 784.6 109.404V136.204H771.7Z" fill="black"/>
-    <path d="M870.1 89.9037C870.1 81.3037 863.9 75.3037 856.2 75.3037C848.6 75.3037 842.3 81.4037 842.3 89.9037C842.3 98.3037 848.5 104.504 856.2 104.504C863.9 104.504 870.1 98.4037 870.1 89.9037ZM870.2 115.404V108.804C866.4 113.404 860.8 116.204 854.1 116.204C840.1 116.204 829.5 105.004 829.5 89.9037C829.5 74.9037 840.5 63.5037 854.4 63.5037C860.9 63.5037 866.4 66.4037 870.2 70.8037V64.4037H882.7V115.304L870.2 115.404Z" fill="black"/>
-    <path d="M749.1 69.4037C749.1 69.4037 752.3 63.5037 760.1 63.5037C763.4 63.5037 765.6 64.7037 765.6 64.7037V77.7037C765.6 77.7037 760.9 74.8037 756.6 75.4037C752.3 76.0037 749.5 80.0037 749.5 85.3037V115.504H736.6V64.4037H749.1V69.4037Z" fill="black"/>
-    <path d="M1041.7 61.5037L1026.9 53.0037L1011.9 44.4037C1001.9 38.7037 989.5 45.9037 989.5 57.4037V59.3037C989.5 60.4037 990.1 61.3037 991 61.9037L998 65.9037C999.9 67.0037 1002.3 65.6037 1002.3 63.4037V58.8037C1002.3 56.5037 1004.8 55.1037 1006.7 56.2037L1020.4 64.1037L1034.1 72.0037C1036.1 73.1037 1036.1 76.0037 1034.1 77.1037L1020.4 85.0037L1006.7 92.9037C1004.7 94.0037 1002.3 92.6037 1002.3 90.3037V88.0037C1002.3 76.5037 989.9 69.3037 979.9 75.0037L964.9 83.6037L950 92.1037C940 97.8037 940 112.304 950 118.004L964.8 126.504L979.8 135.104C989.8 140.804 1002.2 133.604 1002.2 122.104V120.204C1002.2 119.104 1001.6 118.204 1000.7 117.604L993.7 113.604C991.8 112.504 989.4 113.904 989.4 116.104V120.704C989.4 123.004 986.9 124.404 985 123.304L971.3 115.404L957.6 107.504C955.6 106.404 955.6 103.504 957.6 102.404L971.3 94.5037L985 86.6037C987 85.5037 989.4 86.9037 989.4 89.2037V91.5037C989.4 103.004 1001.8 110.204 1011.8 104.504L1026.8 95.9037L1041.6 87.4037C1051.7 81.7037 1051.7 67.3037 1041.7 61.5037V61.5037Z" fill="black"/>
-    <path d="M592.6 96.4037C589.9 108.104 579.8 116.204 566.7 116.204C551.6 116.204 540.3 104.904 540.3 89.9037C540.3 74.9037 551.8 63.5037 566.9 63.5037C579.6 63.5037 589.6 71.5037 592.5 83.0037H579.3C577 78.4037 572.3 75.4037 567.1 75.4037C559.6 75.4037 553.3 81.6037 553.3 89.9037C553.3 98.1037 559.6 104.404 567.1 104.404C572.5 104.404 577.2 101.304 579.4 96.5037L592.6 96.4037Z" fill="black"/>
-    <path d="M598.6 115.304V43.5037H611.4V115.304H598.6Z" fill="black"/>
-    <path d="M630.7 93.2036C631.6 100.704 637 105.004 643.8 105.004C649.2 105.004 653.4 102.504 655.8 98.4036H669C665.9 109.204 656.3 116.104 643.5 116.104C628.1 116.104 617.3 105.304 617.3 89.9036C617.3 74.5036 628.7 63.4036 643.8 63.4036C659 63.5036 670 74.7036 670 90.0036C670 91.1036 669.9 92.2036 669.7 93.3036H630.7V93.2036ZM656.2 85.3036C655.3 78.7036 649.9 74.7036 643.6 74.7036C637.3 74.7036 632.1 78.6036 630.8 85.3036H656.2Z" fill="black"/>
-    <path d="M715.7 115.304V108.704C711.9 113.304 706.3 116.104 699.6 116.104C685.7 116.104 675.1 104.904 675.1 89.8036C675.1 74.8036 686.1 63.4036 699.9 63.4036C706.4 63.4036 711.9 66.3036 715.7 70.7036V64.3036H728.2V115.204L715.7 115.304ZM715.6 89.9036C715.6 81.3036 709.4 75.3036 701.8 75.3036C694.2 75.3036 688 81.4036 688 89.9036C688 98.2036 694.2 104.504 701.8 104.504C709.4 104.504 715.6 98.4036 715.6 89.9036Z" fill="black"/>
-    <path d="M150 111.204V67.6038H169.6C179.1 67.6038 185.2 72.6038 185.2 81.2038C185.2 89.9038 178.4 95.2038 169.3 95.2038H158.5V111.204H150ZM169.1 88.3038C173.6 88.3038 176.6 85.2038 176.6 81.3038C176.6 77.6038 173.6 74.8038 169.2 74.8038H158.5V88.4038H169.1V88.3038Z" fill="black"/>
-    <path d="M212.4 111.204V107.004C210 109.904 206.5 111.704 202.3 111.704C193.5 111.704 186.9 104.704 186.9 95.2038C186.9 85.8038 193.8 78.6038 202.5 78.6038C206.6 78.6038 210.1 80.4038 212.4 83.2038V79.2038H220.3V111.204H212.4ZM212.4 95.2038C212.4 89.8038 208.5 86.0038 203.7 86.0038C198.9 86.0038 195 89.8038 195 95.2038C195 100.404 198.9 104.404 203.7 104.404C208.5 104.404 212.4 100.504 212.4 95.2038Z" fill="black"/>
-    <path d="M227.9 122.904V115.904H230.5C232.7 115.904 234.8 114.804 235.4 111.504C235.6 110.504 235.5 110.004 235.1 109.004L223.3 79.2036H231.6L240.1 101.304L248.4 79.2036H256.5L244.6 109.204C240.7 119.104 239.6 123.004 230.3 123.004H227.9V122.904Z" fill="black"/>
-    <path d="M290.7 111.204V91.3038C290.7 87.8038 288.3 85.9038 285.1 85.9038C281.8 85.9038 279.2 88.0038 279.2 91.3038V111.304H271.1V79.2038H279V83.2038C280.2 81.3038 283.1 78.6038 287.9 78.6038C294.7 78.6038 298.8 83.0038 298.8 89.8038V111.204H290.7Z" fill="black"/>
-    <path d="M320.5 111.704C311 111.704 303.7 104.604 303.7 95.2038C303.7 85.8038 310.9 78.6038 320.5 78.6038C330 78.6038 337.3 85.8038 337.3 95.2038C337.2 104.604 329.9 111.704 320.5 111.704ZM329.1 95.2038C329.1 90.0038 325.2 86.1038 320.5 86.1038C315.8 86.1038 311.8 90.0038 311.8 95.2038C311.8 100.404 315.8 104.304 320.5 104.304C325.2 104.304 329.1 100.304 329.1 95.2038Z" fill="black"/>
-    <path d="M368.9 111.204L362.3 91.4036H362.1L355.3 111.204H347.9L338.4 79.2036H346.8L352.8 99.9036H352.9L359.7 79.2036H365.4L371.9 100.104H372L378.2 79.2036H386.4L376.4 111.204H368.9Z" fill="black"/>
-    <path d="M429.4 111.204L422.8 91.4036H422.6L415.8 111.204H408.4L398.9 79.2036H407.3L413.3 99.9036H413.4L420.2 79.2036H425.9L432.4 100.104H432.5L438.7 79.2036H446.9L436.9 111.204H429.4Z" fill="black"/>
-    <path d="M450.3 70.6037C450.3 68.2037 452.4 66.0037 454.9 66.0037C457.4 66.0037 459.5 68.2037 459.5 70.6037C459.5 73.1037 457.4 75.2037 454.9 75.2037C452.4 75.2037 450.3 73.1037 450.3 70.6037ZM450.8 111.204V79.2037H458.9V111.204H450.8Z" fill="black"/>
-    <path d="M476.2 111.204C469.7 111.204 467.2 108.404 467.2 101.904V86.1037H463V79.1037H467.2V71.3037H475.3V79.2037H480.4V86.2037H475.3V100.404C475.3 103.104 476.1 104.104 478.8 104.104H480.4V111.204H476.2Z" fill="black"/>
-    <path d="M504.5 111.204V91.3038C504.5 87.8038 502.1 85.9038 498.9 85.9038C495.6 85.9038 493 88.0038 493 91.3038V111.304H484.9V67.6038H493V82.8038C494.3 80.9038 497.1 78.6038 501.6 78.6038C508.4 78.6038 512.5 83.0038 512.5 89.8038V111.204H504.5Z" fill="black"/>
-    </g>
-    <defs>
-    <clipPath id="clip0">
-    <rect width="1200" height="180" fill="white"/>
-    </clipPath>
-    </defs>
-    </svg>
-    """
-  )
-
-  static let afterpayBuyNowWhiteOnBlack = SVG(
-    size: CGSize(width: 1200, height: 180),
-    minimumWidth: 256,
-    svgString: """
-  <svg width="1200" height="180" viewBox="0 0 1200 180" fill="none" xmlns="http://www.w3.org/2000/svg">
-  <rect width="1200" height="180" rx="30" fill="black"/>
-  <path d="M578.657 91.6002C578.657 83.0412 572.421 77.0257 564.761 77.0257C557.1 77.0257 550.864 83.1426 550.864 91.6002C550.864 99.9563 557.1 106.175 564.761 106.175C572.429 106.167 578.657 100.159 578.657 91.6002ZM578.767 117.074V110.45C574.988 115.038 569.362 117.886 562.624 117.886C548.626 117.886 538 106.674 538 91.6002C538 76.62 549.033 65.2132 562.937 65.2132C569.479 65.2132 574.995 68.0688 578.775 72.5472V66.126H591.341V117.074H578.767Z" fill="white"/>
-  <path d="M652.382 105.761C647.985 105.761 646.764 104.13 646.764 99.8549V77.2363H654.839V66.126H646.764V53.6971H633.884V66.126H617.296V63.0285C617.296 58.7451 618.931 57.1145 623.431 57.1145H626.255V47.2291H620.058C609.432 47.2291 604.424 50.6933 604.424 61.2886V66.1182H597.272V77.2285H604.424V117.067H617.304V77.2285H633.892V102.196C633.892 112.588 637.875 117.074 648.298 117.074H654.941V105.761H652.382Z" fill="white"/>
-  <path d="M698.572 87.0125C697.656 80.3884 692.234 76.4171 685.903 76.4171C679.565 76.4171 674.354 80.287 673.032 87.0125H698.572ZM672.922 94.9629C673.845 102.5 679.26 106.783 686.107 106.783C691.522 106.783 695.716 104.232 698.165 100.159H711.349C708.282 110.957 698.572 117.886 685.802 117.886C670.371 117.886 659.542 107.088 659.542 91.7016C659.542 76.3157 670.989 65.2053 686.115 65.2053C701.342 65.2053 712.375 76.4171 712.375 91.7016C712.375 92.8251 712.273 93.9408 712.069 94.9629H672.922Z" fill="white"/>
-  <path d="M794.261 91.6002C794.261 83.3455 788.024 77.0257 780.364 77.0257C772.703 77.0257 766.467 83.1426 766.467 91.6002C766.467 99.9563 772.703 106.175 780.364 106.175C788.024 106.167 794.261 99.8549 794.261 91.6002ZM753.791 66.126H766.358V72.7501C770.137 68.061 775.755 65.2132 782.5 65.2132C796.295 65.2132 807.125 76.5263 807.125 91.4987C807.125 106.479 796.092 117.894 782.195 117.894C775.755 117.894 770.442 115.342 766.764 111.067V138.039H753.791V66.126Z" fill="white"/>
-  <path d="M852.422 91.6002C852.422 83.0412 846.193 77.0257 838.525 77.0257C830.865 77.0257 824.628 83.1426 824.628 91.6002C824.628 99.9563 830.865 106.175 838.525 106.175C846.193 106.167 852.422 100.159 852.422 91.6002ZM852.531 117.074V110.45C848.752 115.038 843.126 117.886 836.389 117.886C822.391 117.886 811.765 106.674 811.765 91.6002C811.765 76.62 822.797 65.2132 836.694 65.2132C843.236 65.2132 848.752 68.0688 852.531 72.5472V66.126H865.098V117.074H852.531Z" fill="white"/>
-  <path d="M731.138 71.1194C731.138 71.1194 734.339 65.2132 742.171 65.2132C745.52 65.2132 747.688 66.3601 747.688 66.3601V79.3585C747.688 79.3585 742.962 76.4483 738.619 77.0335C734.276 77.6186 731.53 81.5978 731.545 86.9189V117.082H718.572V66.1338H731.138V71.1194Z" fill="white"/>
-  <path d="M923.792 66.126L892.062 137.875H878.682L891.139 110.138L870.075 66.126H885.239L897.563 95.5325L910.208 66.126H923.792Z" fill="white"/>
-  <path d="M1025.08 64.9713L993.62 46.8702C984.386 41.5569 972.845 48.1966 972.845 58.831V60.6879C972.845 62.3654 973.745 63.918 975.2 64.7528L981.139 68.1702C982.884 69.1767 985.059 67.9205 985.059 65.9154V61.2965C985.059 58.987 987.563 57.5436 989.566 58.6983L1016.81 74.3807C1018.82 75.5355 1018.82 78.4223 1016.81 79.5692L989.566 95.2516C987.563 96.4063 985.059 94.9629 985.059 92.6535V90.1958C985.059 79.5614 973.518 72.9139 964.277 78.235L932.821 96.3361C923.588 101.649 923.588 114.944 932.821 120.258L964.277 138.359C973.51 143.672 985.059 137.032 985.059 126.398V124.541C985.059 122.864 984.159 121.319 982.704 120.476L976.765 117.051C975.02 116.044 972.845 117.301 972.845 119.306V123.925C972.845 126.234 970.341 127.678 968.338 126.523L941.092 110.84C939.089 109.686 939.089 106.799 941.092 105.644L968.338 89.9617C970.341 88.807 972.845 90.2504 972.845 92.5598V95.0175C972.845 105.652 984.386 112.299 993.62 106.978L1025.08 88.8772C1034.31 83.5795 1034.31 70.2846 1025.08 64.9713Z" fill="white"/>
-  <path d="M190.42 109.478C198.673 109.478 204.406 103.934 204.406 97.004C204.406 92.594 201.571 89.318 198.547 87.554C201.319 85.79 203.902 82.64 203.902 78.419C203.902 71.426 197.98 66.638 190.483 66.638H169V109.478H190.42ZM189.475 73.505C193.192 73.505 195.649 75.584 195.649 78.86C195.649 81.884 193.318 84.467 189.79 84.467H177.253V73.505H189.475ZM189.79 90.956C193.696 90.956 196.027 93.791 196.027 96.626C196.027 99.965 193.192 102.548 189.664 102.548H177.253V90.956H189.79Z" fill="white"/>
-  <path d="M236.063 109.478V77.978H228.125V97.949C228.125 101.099 225.542 103.115 222.266 103.115C219.116 103.115 216.848 101.288 216.848 97.823V77.978H208.847V99.083C208.847 105.824 212.816 109.982 219.62 109.982C224.345 109.982 227.243 107.462 228.314 105.572V109.478H236.063Z" fill="white"/>
-  <path d="M245.648 121.07C254.783 121.07 255.854 117.29 259.697 107.525L271.415 77.978H263.477L255.35 99.776L246.971 77.978H238.781L250.499 107.336C250.877 108.344 250.94 108.848 250.751 109.793C250.121 113.069 248.042 114.14 245.9 114.14H243.317V121.07H245.648Z" fill="white"/>
-  <path d="M312.391 109.478V88.436C312.391 81.695 308.422 77.411 301.681 77.411C296.956 77.411 294.058 80.057 292.924 81.947V77.978H285.175V109.478H293.176V89.822C293.176 86.609 295.696 84.53 298.972 84.53C302.122 84.53 304.453 86.42 304.453 89.885V109.478H312.391Z" fill="white"/>
-  <path d="M349.949 93.728C349.949 84.467 342.767 77.411 333.443 77.411C324.056 77.411 316.937 84.467 316.937 93.728C316.937 102.989 324.056 109.982 333.443 109.982C342.767 109.982 349.949 102.989 349.949 93.728ZM333.443 102.674C328.781 102.674 324.875 98.831 324.875 93.728C324.875 88.625 328.781 84.782 333.443 84.782C338.105 84.782 341.948 88.625 341.948 93.728C341.948 98.831 338.105 102.674 333.443 102.674Z" fill="white"/>
-  <path d="M388.163 109.478L397.991 77.978H389.927L383.816 98.579H383.753L377.39 77.978H371.783L365.105 98.39H364.979L359.12 77.978H350.867L360.191 109.478H367.436L374.177 89.948H374.366L380.855 109.478H388.163Z" fill="white"/>
-  <path d="M447.087 109.478L456.915 77.978H448.851L442.74 98.579H442.677L436.314 77.978H430.707L424.029 98.39H423.903L418.044 77.978H409.791L419.115 109.478H426.36L433.101 89.948H433.29L439.779 109.478H447.087Z" fill="white"/>
-  <path d="M464.515 74.072C466.972 74.072 469.051 71.993 469.051 69.536C469.051 67.142 466.972 65 464.515 65C462.058 65 459.979 67.142 459.979 69.536C459.979 71.993 462.058 74.072 464.515 74.072ZM468.421 109.478V77.978H460.483V109.478H468.421Z" fill="white"/>
-  <path d="M489.299 109.478V102.485H487.724C485.015 102.485 484.259 101.477 484.259 98.831V84.845H489.236V77.978H484.259V70.292H476.321V77.978H472.226V84.845H476.321V100.28C476.321 106.706 478.778 109.478 485.204 109.478H489.299Z" fill="white"/>
-  <path d="M520.658 109.478V88.436C520.658 81.695 516.689 77.411 509.948 77.411C505.475 77.411 502.703 79.679 501.443 81.569V66.638H493.442V109.478H501.443V89.822C501.443 86.609 503.963 84.53 507.239 84.53C510.389 84.53 512.72 86.42 512.72 89.885V109.478H520.658Z" fill="white"/>
-  </svg>
-  """
-  )
-
-  static let afterpayBuyNowBlackOnMint = SVG(
-  size: CGSize(width: 1200, height: 180),
-  minimumWidth: 256,
-  svgString: """
-  <svg width="1200" height="180" viewBox="0 0 1200 180" fill="none" xmlns="http://www.w3.org/2000/svg">
-  <rect width="1200" height="180" rx="30" fill="#B2FCE4"/>
-  <path d="M578.657 91.6002C578.657 83.0412 572.421 77.0257 564.761 77.0257C557.1 77.0257 550.864 83.1426 550.864 91.6002C550.864 99.9563 557.1 106.175 564.761 106.175C572.429 106.167 578.657 100.159 578.657 91.6002ZM578.767 117.074V110.45C574.988 115.038 569.362 117.886 562.624 117.886C548.626 117.886 538 106.674 538 91.6002C538 76.62 549.033 65.2132 562.937 65.2132C569.479 65.2132 574.995 68.0688 578.775 72.5472V66.126H591.341V117.074H578.767Z" fill="black"/>
-  <path d="M652.382 105.761C647.985 105.761 646.764 104.13 646.764 99.8549V77.2363H654.839V66.126H646.764V53.6971H633.884V66.126H617.296V63.0285C617.296 58.7451 618.931 57.1145 623.431 57.1145H626.255V47.2291H620.058C609.432 47.2291 604.424 50.6933 604.424 61.2886V66.1182H597.272V77.2285H604.424V117.067H617.304V77.2285H633.892V102.196C633.892 112.588 637.875 117.074 648.298 117.074H654.941V105.761H652.382Z" fill="black"/>
-  <path d="M698.572 87.0125C697.656 80.3884 692.234 76.4171 685.903 76.4171C679.565 76.4171 674.354 80.287 673.032 87.0125H698.572ZM672.922 94.9629C673.845 102.5 679.26 106.783 686.107 106.783C691.522 106.783 695.716 104.232 698.165 100.159H711.349C708.282 110.957 698.572 117.886 685.802 117.886C670.371 117.886 659.542 107.088 659.542 91.7016C659.542 76.3157 670.989 65.2053 686.115 65.2053C701.342 65.2053 712.375 76.4171 712.375 91.7016C712.375 92.8251 712.273 93.9408 712.069 94.9629H672.922Z" fill="black"/>
-  <path d="M794.261 91.6002C794.261 83.3455 788.024 77.0257 780.364 77.0257C772.703 77.0257 766.467 83.1426 766.467 91.6002C766.467 99.9563 772.703 106.175 780.364 106.175C788.024 106.167 794.261 99.8549 794.261 91.6002ZM753.791 66.126H766.358V72.7501C770.137 68.061 775.755 65.2132 782.5 65.2132C796.295 65.2132 807.125 76.5263 807.125 91.4987C807.125 106.479 796.092 117.894 782.195 117.894C775.755 117.894 770.442 115.342 766.764 111.067V138.039H753.791V66.126Z" fill="black"/>
-  <path d="M852.422 91.6002C852.422 83.0412 846.193 77.0257 838.525 77.0257C830.865 77.0257 824.628 83.1426 824.628 91.6002C824.628 99.9563 830.865 106.175 838.525 106.175C846.193 106.167 852.422 100.159 852.422 91.6002ZM852.531 117.074V110.45C848.752 115.038 843.126 117.886 836.389 117.886C822.391 117.886 811.765 106.674 811.765 91.6002C811.765 76.62 822.797 65.2132 836.694 65.2132C843.236 65.2132 848.752 68.0688 852.531 72.5472V66.126H865.098V117.074H852.531Z" fill="black"/>
-  <path d="M731.138 71.1194C731.138 71.1194 734.339 65.2132 742.171 65.2132C745.52 65.2132 747.688 66.3601 747.688 66.3601V79.3585C747.688 79.3585 742.962 76.4483 738.619 77.0335C734.276 77.6186 731.53 81.5978 731.545 86.9189V117.082H718.572V66.1338H731.138V71.1194Z" fill="black"/>
-  <path d="M923.792 66.126L892.062 137.875H878.682L891.139 110.138L870.075 66.126H885.239L897.563 95.5325L910.208 66.126H923.792Z" fill="black"/>
-  <path d="M1025.08 64.9713L993.62 46.8702C984.386 41.5569 972.845 48.1966 972.845 58.831V60.6879C972.845 62.3654 973.745 63.918 975.2 64.7528L981.139 68.1702C982.884 69.1767 985.059 67.9205 985.059 65.9154V61.2965C985.059 58.987 987.563 57.5436 989.566 58.6983L1016.81 74.3807C1018.82 75.5355 1018.82 78.4223 1016.81 79.5692L989.566 95.2516C987.563 96.4063 985.059 94.9629 985.059 92.6535V90.1958C985.059 79.5614 973.518 72.9139 964.277 78.235L932.821 96.3361C923.588 101.649 923.588 114.944 932.821 120.258L964.277 138.359C973.51 143.672 985.059 137.032 985.059 126.398V124.541C985.059 122.864 984.159 121.319 982.704 120.476L976.765 117.051C975.02 116.044 972.845 117.301 972.845 119.306V123.925C972.845 126.234 970.341 127.678 968.338 126.523L941.092 110.84C939.089 109.686 939.089 106.799 941.092 105.644L968.338 89.9617C970.341 88.807 972.845 90.2504 972.845 92.5598V95.0175C972.845 105.652 984.386 112.299 993.62 106.978L1025.08 88.8772C1034.31 83.5795 1034.31 70.2846 1025.08 64.9713Z" fill="black"/>
-  <path d="M190.42 109.478C198.673 109.478 204.406 103.934 204.406 97.004C204.406 92.594 201.571 89.318 198.547 87.554C201.319 85.79 203.902 82.64 203.902 78.419C203.902 71.426 197.98 66.638 190.483 66.638H169V109.478H190.42ZM189.475 73.505C193.192 73.505 195.649 75.584 195.649 78.86C195.649 81.884 193.318 84.467 189.79 84.467H177.253V73.505H189.475ZM189.79 90.956C193.696 90.956 196.027 93.791 196.027 96.626C196.027 99.965 193.192 102.548 189.664 102.548H177.253V90.956H189.79Z" fill="black"/>
-  <path d="M236.063 109.478V77.978H228.125V97.949C228.125 101.099 225.542 103.115 222.266 103.115C219.116 103.115 216.848 101.288 216.848 97.823V77.978H208.847V99.083C208.847 105.824 212.816 109.982 219.62 109.982C224.345 109.982 227.243 107.462 228.314 105.572V109.478H236.063Z" fill="black"/>
-  <path d="M245.648 121.07C254.783 121.07 255.854 117.29 259.697 107.525L271.415 77.978H263.477L255.35 99.776L246.971 77.978H238.781L250.499 107.336C250.877 108.344 250.94 108.848 250.751 109.793C250.121 113.069 248.042 114.14 245.9 114.14H243.317V121.07H245.648Z" fill="black"/>
-  <path d="M312.391 109.478V88.436C312.391 81.695 308.422 77.411 301.681 77.411C296.956 77.411 294.058 80.057 292.924 81.947V77.978H285.175V109.478H293.176V89.822C293.176 86.609 295.696 84.53 298.972 84.53C302.122 84.53 304.453 86.42 304.453 89.885V109.478H312.391Z" fill="black"/>
-  <path d="M349.949 93.728C349.949 84.467 342.767 77.411 333.443 77.411C324.056 77.411 316.937 84.467 316.937 93.728C316.937 102.989 324.056 109.982 333.443 109.982C342.767 109.982 349.949 102.989 349.949 93.728ZM333.443 102.674C328.781 102.674 324.875 98.831 324.875 93.728C324.875 88.625 328.781 84.782 333.443 84.782C338.105 84.782 341.948 88.625 341.948 93.728C341.948 98.831 338.105 102.674 333.443 102.674Z" fill="black"/>
-  <path d="M388.163 109.478L397.991 77.978H389.927L383.816 98.579H383.753L377.39 77.978H371.783L365.105 98.39H364.979L359.12 77.978H350.867L360.191 109.478H367.436L374.177 89.948H374.366L380.855 109.478H388.163Z" fill="black"/>
-  <path d="M447.087 109.478L456.915 77.978H448.851L442.74 98.579H442.677L436.314 77.978H430.707L424.029 98.39H423.903L418.044 77.978H409.791L419.115 109.478H426.36L433.101 89.948H433.29L439.779 109.478H447.087Z" fill="black"/>
-  <path d="M464.515 74.072C466.972 74.072 469.051 71.993 469.051 69.536C469.051 67.142 466.972 65 464.515 65C462.058 65 459.979 67.142 459.979 69.536C459.979 71.993 462.058 74.072 464.515 74.072ZM468.421 109.478V77.978H460.483V109.478H468.421Z" fill="black"/>
-  <path d="M489.299 109.478V102.485H487.724C485.015 102.485 484.259 101.477 484.259 98.831V84.845H489.236V77.978H484.259V70.292H476.321V77.978H472.226V84.845H476.321V100.28C476.321 106.706 478.778 109.478 485.204 109.478H489.299Z" fill="black"/>
-  <path d="M520.658 109.478V88.436C520.658 81.695 516.689 77.411 509.948 77.411C505.475 77.411 502.703 79.679 501.443 81.569V66.638H493.442V109.478H501.443V89.822C501.443 86.609 503.963 84.53 507.239 84.53C510.389 84.53 512.72 86.42 512.72 89.885V109.478H520.658Z" fill="black"/>
-  </svg>
-  """
-  )
-
-  static let afterpayCheckoutWhiteOnBlack = SVG(
-  size: CGSize(width: 1200, height: 180),
-  minimumWidth: 256,
-  svgString: """
-  <svg width="1200" height="180" viewBox="0 0 1200 180" fill="none" xmlns="http://www.w3.org/2000/svg">
-  <rect width="1200" height="180" rx="30" fill="black"/>
-  <path d="M179.351 95.455C176.894 99.424 172.673 101.944 167.696 101.944C160.01 101.944 153.899 95.833 153.899 87.58C153.899 79.39 160.01 73.216 167.696 73.216C172.61 73.216 176.768 75.673 179.162 79.579H188.612C185.399 71.2 177.272 65.656 167.759 65.656C155.348 65.656 145.268 75.043 145.268 87.706C145.268 100.243 155.033 109.504 167.633 109.504C177.461 109.504 185.651 103.897 188.738 95.455H179.351Z" fill="white"/>
-  <path d="M219.81 109V87.958C219.81 81.217 215.841 76.933 209.1 76.933C204.627 76.933 201.855 79.201 200.595 81.091V66.16H192.594V109H200.595V89.344C200.595 86.131 203.115 84.052 206.391 84.052C209.541 84.052 211.872 85.942 211.872 89.407V109H219.81Z" fill="white"/>
-  <path d="M256.738 95.329C256.864 94.699 256.927 94.006 256.927 93.313C256.927 83.863 250.123 76.933 240.736 76.933C231.412 76.933 224.356 83.8 224.356 93.313C224.356 102.826 231.034 109.504 240.547 109.504C248.422 109.504 254.407 105.22 256.297 98.542H248.17C246.658 101.062 244.075 102.637 240.736 102.637C236.515 102.637 233.176 99.991 232.609 95.329H256.738ZM232.672 90.415C233.491 86.257 236.704 83.863 240.61 83.863C244.516 83.863 247.855 86.32 248.422 90.415H232.672Z" fill="white"/>
-  <path d="M283.915 97.282C282.529 100.306 279.631 102.196 276.292 102.196C271.63 102.196 267.724 98.353 267.724 93.25C267.724 88.147 271.63 84.304 276.292 84.304C279.505 84.304 282.403 86.194 283.852 89.029H292.042C290.278 81.91 284.104 76.933 276.229 76.933C266.905 76.933 259.786 83.989 259.786 93.25C259.786 102.511 266.779 109.504 276.103 109.504C284.167 109.504 290.404 104.527 292.105 97.282H283.915Z" fill="white"/>
-  <path d="M326.166 109L313.188 89.47L325.158 77.5H316.149L304.242 89.785H303.99V66.16H296.178V109H303.99V98.164L307.77 94.636L316.842 109H326.166Z" fill="white"/>
-  <path d="M357.937 93.25C357.937 83.989 350.755 76.933 341.431 76.933C332.044 76.933 324.925 83.989 324.925 93.25C324.925 102.511 332.044 109.504 341.431 109.504C350.755 109.504 357.937 102.511 357.937 93.25ZM341.431 102.196C336.769 102.196 332.863 98.353 332.863 93.25C332.863 88.147 336.769 84.304 341.431 84.304C346.093 84.304 349.936 88.147 349.936 93.25C349.936 98.353 346.093 102.196 341.431 102.196Z" fill="white"/>
-  <path d="M389.64 109V77.5H381.702V97.471C381.702 100.621 379.119 102.637 375.843 102.637C372.693 102.637 370.425 100.81 370.425 97.345V77.5H362.424V98.605C362.424 105.346 366.393 109.504 373.197 109.504C377.922 109.504 380.82 106.984 381.891 105.094V109H389.64Z" fill="white"/>
-  <path d="M410.376 109V102.007H408.801C406.092 102.007 405.336 100.999 405.336 98.353V84.367H410.313V77.5H405.336V69.814H397.398V77.5H393.303V84.367H397.398V99.802C397.398 106.228 399.855 109 406.281 109H410.376Z" fill="white"/>
-  <path d="M460.174 109L470.003 77.5H461.939L455.827 98.101H455.764L449.401 77.5H443.794L437.116 97.912H436.99L431.132 77.5H422.878L432.202 109H439.447L446.189 89.47H446.378L452.866 109H460.174Z" fill="white"/>
-  <path d="M477.602 73.594C480.059 73.594 482.138 71.515 482.138 69.058C482.138 66.664 480.059 64.522 477.602 64.522C475.145 64.522 473.066 66.664 473.066 69.058C473.066 71.515 475.145 73.594 477.602 73.594ZM481.508 109V77.5H473.57V109H481.508Z" fill="white"/>
-  <path d="M502.386 109V102.007H500.811C498.102 102.007 497.346 100.999 497.346 98.353V84.367H502.323V77.5H497.346V69.814H489.408V77.5H485.313V84.367H489.408V99.802C489.408 106.228 491.865 109 498.291 109H502.386Z" fill="white"/>
-  <path d="M533.745 109V87.958C533.745 81.217 529.776 76.933 523.035 76.933C518.562 76.933 515.79 79.201 514.53 81.091V66.16H506.529V109H514.53V89.344C514.53 86.131 517.05 84.052 520.326 84.052C523.476 84.052 525.807 85.942 525.807 89.407V109H533.745Z" fill="white"/>
-  <path d="M591.657 91.6002C591.657 83.0412 585.421 77.0257 577.761 77.0257C570.1 77.0257 563.864 83.1426 563.864 91.6002C563.864 99.9563 570.1 106.175 577.761 106.175C585.429 106.167 591.657 100.159 591.657 91.6002ZM591.767 117.074V110.45C587.988 115.038 582.362 117.886 575.624 117.886C561.626 117.886 551 106.674 551 91.6002C551 76.62 562.033 65.2132 575.937 65.2132C582.479 65.2132 587.995 68.0688 591.775 72.5472V66.126H604.341V117.074H591.767Z" fill="white"/>
-  <path d="M665.382 105.761C660.985 105.761 659.764 104.13 659.764 99.8549V77.2363H667.839V66.126H659.764V53.6971H646.884V66.126H630.296V63.0285C630.296 58.7451 631.931 57.1145 636.431 57.1145H639.255V47.2291H633.058C622.432 47.2291 617.424 50.6933 617.424 61.2886V66.1182H610.272V77.2285H617.424V117.067H630.304V77.2285H646.892V102.196C646.892 112.588 650.875 117.074 661.298 117.074H667.941V105.761H665.382Z" fill="white"/>
-  <path d="M711.572 87.0125C710.656 80.3884 705.234 76.4171 698.903 76.4171C692.565 76.4171 687.354 80.287 686.032 87.0125H711.572ZM685.922 94.9629C686.845 102.5 692.26 106.783 699.107 106.783C704.522 106.783 708.716 104.232 711.165 100.159H724.349C721.282 110.957 711.572 117.886 698.802 117.886C683.371 117.886 672.542 107.088 672.542 91.7016C672.542 76.3157 683.989 65.2053 699.115 65.2053C714.342 65.2053 725.375 76.4171 725.375 91.7016C725.375 92.8251 725.273 93.9408 725.069 94.9629H685.922Z" fill="white"/>
-  <path d="M807.261 91.6002C807.261 83.3455 801.024 77.0257 793.364 77.0257C785.703 77.0257 779.467 83.1426 779.467 91.6002C779.467 99.9563 785.703 106.175 793.364 106.175C801.024 106.167 807.261 99.8549 807.261 91.6002ZM766.791 66.126H779.358V72.7501C783.137 68.061 788.755 65.2132 795.5 65.2132C809.295 65.2132 820.125 76.5263 820.125 91.4987C820.125 106.479 809.092 117.894 795.195 117.894C788.755 117.894 783.442 115.342 779.764 111.067V138.039H766.791V66.126Z" fill="white"/>
-  <path d="M865.422 91.6002C865.422 83.0412 859.193 77.0257 851.525 77.0257C843.865 77.0257 837.628 83.1426 837.628 91.6002C837.628 99.9563 843.865 106.175 851.525 106.175C859.193 106.167 865.422 100.159 865.422 91.6002ZM865.531 117.074V110.45C861.752 115.038 856.126 117.886 849.389 117.886C835.391 117.886 824.765 106.674 824.765 91.6002C824.765 76.62 835.797 65.2132 849.694 65.2132C856.236 65.2132 861.752 68.0688 865.531 72.5472V66.126H878.098V117.074H865.531Z" fill="white"/>
-  <path d="M744.138 71.1194C744.138 71.1194 747.339 65.2132 755.171 65.2132C758.52 65.2132 760.688 66.3601 760.688 66.3601V79.3585C760.688 79.3585 755.962 76.4483 751.619 77.0335C747.276 77.6186 744.53 81.5978 744.545 86.9189V117.082H731.572V66.1338H744.138V71.1194Z" fill="white"/>
-  <path d="M936.792 66.126L905.062 137.875H891.682L904.139 110.138L883.075 66.126H898.239L910.563 95.5325L923.208 66.126H936.792Z" fill="white"/>
-  <path d="M1038.08 64.9713L1006.62 46.8702C997.386 41.5569 985.845 48.1966 985.845 58.831V60.6879C985.845 62.3654 986.745 63.918 988.2 64.7528L994.139 68.1702C995.884 69.1767 998.059 67.9205 998.059 65.9154V61.2965C998.059 58.987 1000.56 57.5436 1002.57 58.6983L1029.81 74.3807C1031.82 75.5355 1031.82 78.4223 1029.81 79.5692L1002.57 95.2516C1000.56 96.4063 998.059 94.9629 998.059 92.6535V90.1958C998.059 79.5614 986.518 72.9139 977.277 78.235L945.821 96.3361C936.588 101.649 936.588 114.944 945.821 120.258L977.277 138.359C986.51 143.672 998.059 137.032 998.059 126.398V124.541C998.059 122.864 997.159 121.319 995.704 120.476L989.765 117.051C988.02 116.044 985.845 117.301 985.845 119.306V123.925C985.845 126.234 983.341 127.678 981.338 126.523L954.092 110.84C952.089 109.686 952.089 106.799 954.092 105.644L981.338 89.9617C983.341 88.807 985.845 90.2504 985.845 92.5598V95.0175C985.845 105.652 997.386 112.299 1006.62 106.978L1038.08 88.8772C1047.31 83.5795 1047.31 70.2846 1038.08 64.9713Z" fill="white"/>
-  </svg>
-  """
-  )
-
-  static let afterpayCheckoutBlackOnMint = SVG(
-  size: CGSize(width: 1200, height: 180),
-  minimumWidth: 256,
-  svgString: """
-  <svg width="1200" height="180" viewBox="0 0 1200 180" fill="none" xmlns="http://www.w3.org/2000/svg">
-  <rect width="1200" height="180" rx="30" fill="#B2FCE4"/>
-  <path d="M179.351 95.455C176.894 99.424 172.673 101.944 167.696 101.944C160.01 101.944 153.899 95.833 153.899 87.58C153.899 79.39 160.01 73.216 167.696 73.216C172.61 73.216 176.768 75.673 179.162 79.579H188.612C185.399 71.2 177.272 65.656 167.759 65.656C155.348 65.656 145.268 75.043 145.268 87.706C145.268 100.243 155.033 109.504 167.633 109.504C177.461 109.504 185.651 103.897 188.738 95.455H179.351Z" fill="black"/>
-  <path d="M219.81 109V87.958C219.81 81.217 215.841 76.933 209.1 76.933C204.627 76.933 201.855 79.201 200.595 81.091V66.16H192.594V109H200.595V89.344C200.595 86.131 203.115 84.052 206.391 84.052C209.541 84.052 211.872 85.942 211.872 89.407V109H219.81Z" fill="black"/>
-  <path d="M256.738 95.329C256.864 94.699 256.927 94.006 256.927 93.313C256.927 83.863 250.123 76.933 240.736 76.933C231.412 76.933 224.356 83.8 224.356 93.313C224.356 102.826 231.034 109.504 240.547 109.504C248.422 109.504 254.407 105.22 256.297 98.542H248.17C246.658 101.062 244.075 102.637 240.736 102.637C236.515 102.637 233.176 99.991 232.609 95.329H256.738ZM232.672 90.415C233.491 86.257 236.704 83.863 240.61 83.863C244.516 83.863 247.855 86.32 248.422 90.415H232.672Z" fill="black"/>
-  <path d="M283.915 97.282C282.529 100.306 279.631 102.196 276.292 102.196C271.63 102.196 267.724 98.353 267.724 93.25C267.724 88.147 271.63 84.304 276.292 84.304C279.505 84.304 282.403 86.194 283.852 89.029H292.042C290.278 81.91 284.104 76.933 276.229 76.933C266.905 76.933 259.786 83.989 259.786 93.25C259.786 102.511 266.779 109.504 276.103 109.504C284.167 109.504 290.404 104.527 292.105 97.282H283.915Z" fill="black"/>
-  <path d="M326.166 109L313.188 89.47L325.158 77.5H316.149L304.242 89.785H303.99V66.16H296.178V109H303.99V98.164L307.77 94.636L316.842 109H326.166Z" fill="black"/>
-  <path d="M357.937 93.25C357.937 83.989 350.755 76.933 341.431 76.933C332.044 76.933 324.925 83.989 324.925 93.25C324.925 102.511 332.044 109.504 341.431 109.504C350.755 109.504 357.937 102.511 357.937 93.25ZM341.431 102.196C336.769 102.196 332.863 98.353 332.863 93.25C332.863 88.147 336.769 84.304 341.431 84.304C346.093 84.304 349.936 88.147 349.936 93.25C349.936 98.353 346.093 102.196 341.431 102.196Z" fill="black"/>
-  <path d="M389.64 109V77.5H381.702V97.471C381.702 100.621 379.119 102.637 375.843 102.637C372.693 102.637 370.425 100.81 370.425 97.345V77.5H362.424V98.605C362.424 105.346 366.393 109.504 373.197 109.504C377.922 109.504 380.82 106.984 381.891 105.094V109H389.64Z" fill="black"/>
-  <path d="M410.376 109V102.007H408.801C406.092 102.007 405.336 100.999 405.336 98.353V84.367H410.313V77.5H405.336V69.814H397.398V77.5H393.303V84.367H397.398V99.802C397.398 106.228 399.855 109 406.281 109H410.376Z" fill="black"/>
-  <path d="M460.174 109L470.003 77.5H461.939L455.827 98.101H455.764L449.401 77.5H443.794L437.116 97.912H436.99L431.132 77.5H422.878L432.202 109H439.447L446.189 89.47H446.378L452.866 109H460.174Z" fill="black"/>
-  <path d="M477.602 73.594C480.059 73.594 482.138 71.515 482.138 69.058C482.138 66.664 480.059 64.522 477.602 64.522C475.145 64.522 473.066 66.664 473.066 69.058C473.066 71.515 475.145 73.594 477.602 73.594ZM481.508 109V77.5H473.57V109H481.508Z" fill="black"/>
-  <path d="M502.386 109V102.007H500.811C498.102 102.007 497.346 100.999 497.346 98.353V84.367H502.323V77.5H497.346V69.814H489.408V77.5H485.313V84.367H489.408V99.802C489.408 106.228 491.865 109 498.291 109H502.386Z" fill="black"/>
-  <path d="M533.745 109V87.958C533.745 81.217 529.776 76.933 523.035 76.933C518.562 76.933 515.79 79.201 514.53 81.091V66.16H506.529V109H514.53V89.344C514.53 86.131 517.05 84.052 520.326 84.052C523.476 84.052 525.807 85.942 525.807 89.407V109H533.745Z" fill="black"/>
-  <path d="M591.657 91.6002C591.657 83.0412 585.421 77.0257 577.761 77.0257C570.1 77.0257 563.864 83.1426 563.864 91.6002C563.864 99.9563 570.1 106.175 577.761 106.175C585.429 106.167 591.657 100.159 591.657 91.6002ZM591.767 117.074V110.45C587.988 115.038 582.362 117.886 575.624 117.886C561.626 117.886 551 106.674 551 91.6002C551 76.62 562.033 65.2132 575.937 65.2132C582.479 65.2132 587.995 68.0688 591.775 72.5472V66.126H604.341V117.074H591.767Z" fill="black"/>
-  <path d="M665.382 105.761C660.985 105.761 659.764 104.13 659.764 99.8549V77.2363H667.839V66.126H659.764V53.6971H646.884V66.126H630.296V63.0285C630.296 58.7451 631.931 57.1145 636.431 57.1145H639.255V47.2291H633.058C622.432 47.2291 617.424 50.6933 617.424 61.2886V66.1182H610.272V77.2285H617.424V117.067H630.304V77.2285H646.892V102.196C646.892 112.588 650.875 117.074 661.298 117.074H667.941V105.761H665.382Z" fill="black"/>
-  <path d="M711.572 87.0125C710.656 80.3884 705.234 76.4171 698.903 76.4171C692.565 76.4171 687.354 80.287 686.032 87.0125H711.572ZM685.922 94.9629C686.845 102.5 692.26 106.783 699.107 106.783C704.522 106.783 708.716 104.232 711.165 100.159H724.349C721.282 110.957 711.572 117.886 698.802 117.886C683.371 117.886 672.542 107.088 672.542 91.7016C672.542 76.3157 683.989 65.2053 699.115 65.2053C714.342 65.2053 725.375 76.4171 725.375 91.7016C725.375 92.8251 725.273 93.9408 725.069 94.9629H685.922Z" fill="black"/>
-  <path d="M807.261 91.6002C807.261 83.3455 801.024 77.0257 793.364 77.0257C785.703 77.0257 779.467 83.1426 779.467 91.6002C779.467 99.9563 785.703 106.175 793.364 106.175C801.024 106.167 807.261 99.8549 807.261 91.6002ZM766.791 66.126H779.358V72.7501C783.137 68.061 788.755 65.2132 795.5 65.2132C809.295 65.2132 820.125 76.5263 820.125 91.4987C820.125 106.479 809.092 117.894 795.195 117.894C788.755 117.894 783.442 115.342 779.764 111.067V138.039H766.791V66.126Z" fill="black"/>
-  <path d="M865.422 91.6002C865.422 83.0412 859.193 77.0257 851.525 77.0257C843.865 77.0257 837.628 83.1426 837.628 91.6002C837.628 99.9563 843.865 106.175 851.525 106.175C859.193 106.167 865.422 100.159 865.422 91.6002ZM865.531 117.074V110.45C861.752 115.038 856.126 117.886 849.389 117.886C835.391 117.886 824.765 106.674 824.765 91.6002C824.765 76.62 835.797 65.2132 849.694 65.2132C856.236 65.2132 861.752 68.0688 865.531 72.5472V66.126H878.098V117.074H865.531Z" fill="black"/>
-  <path d="M744.138 71.1194C744.138 71.1194 747.339 65.2132 755.171 65.2132C758.52 65.2132 760.688 66.3601 760.688 66.3601V79.3585C760.688 79.3585 755.962 76.4483 751.619 77.0335C747.276 77.6186 744.53 81.5978 744.545 86.9189V117.082H731.572V66.1338H744.138V71.1194Z" fill="black"/>
-  <path d="M936.792 66.126L905.062 137.875H891.682L904.139 110.138L883.075 66.126H898.239L910.563 95.5325L923.208 66.126H936.792Z" fill="black"/>
-  <path d="M1038.08 64.9713L1006.62 46.8702C997.386 41.5569 985.845 48.1966 985.845 58.831V60.6879C985.845 62.3654 986.745 63.918 988.2 64.7528L994.139 68.1702C995.884 69.1767 998.059 67.9205 998.059 65.9154V61.2965C998.059 58.987 1000.56 57.5436 1002.57 58.6983L1029.81 74.3807C1031.82 75.5355 1031.82 78.4223 1029.81 79.5692L1002.57 95.2516C1000.56 96.4063 998.059 94.9629 998.059 92.6535V90.1958C998.059 79.5614 986.518 72.9139 977.277 78.235L945.821 96.3361C936.588 101.649 936.588 114.944 945.821 120.258L977.277 138.359C986.51 143.672 998.059 137.032 998.059 126.398V124.541C998.059 122.864 997.159 121.319 995.704 120.476L989.765 117.051C988.02 116.044 985.845 117.301 985.845 119.306V123.925C985.845 126.234 983.341 127.678 981.338 126.523L954.092 110.84C952.089 109.686 952.089 106.799 954.092 105.644L981.338 89.9617C983.341 88.807 985.845 90.2504 985.845 92.5598V95.0175C985.845 105.652 997.386 112.299 1006.62 106.978L1038.08 88.8772C1047.31 83.5795 1047.31 70.2846 1038.08 64.9713Z" fill="black"/>
-  </svg>
-  """
-  )
-
-  static let afterpayPlaceOrderWhiteOnBlack = SVG(
-  size: CGSize(width: 1200, height: 180),
-  minimumWidth: 256,
-  svgString: """
-  <svg width="1200" height="180" viewBox="0 0 1200 180" fill="none" xmlns="http://www.w3.org/2000/svg">
-  <rect width="1200" height="180" rx="30" fill="black"/>
-  <path d="M623.657 91.6002C623.657 83.0412 617.421 77.0257 609.761 77.0257C602.1 77.0257 595.864 83.1426 595.864 91.6002C595.864 99.9563 602.1 106.175 609.761 106.175C617.429 106.167 623.657 100.159 623.657 91.6002ZM623.767 117.074V110.45C619.988 115.038 614.362 117.886 607.624 117.886C593.626 117.886 583 106.674 583 91.6002C583 76.62 594.033 65.2132 607.937 65.2132C614.479 65.2132 619.995 68.0688 623.775 72.5472V66.126H636.341V117.074H623.767Z" fill="white"/>
-  <path d="M697.382 105.761C692.985 105.761 691.764 104.13 691.764 99.8549V77.2363H699.839V66.126H691.764V53.6971H678.884V66.126H662.296V63.0285C662.296 58.7451 663.931 57.1145 668.431 57.1145H671.255V47.2291H665.058C654.432 47.2291 649.424 50.6933 649.424 61.2886V66.1182H642.272V77.2285H649.424V117.067H662.304V77.2285H678.892V102.196C678.892 112.588 682.875 117.074 693.298 117.074H699.941V105.761H697.382Z" fill="white"/>
-  <path d="M743.572 87.0125C742.656 80.3884 737.234 76.4171 730.903 76.4171C724.565 76.4171 719.354 80.287 718.032 87.0125H743.572ZM717.922 94.9629C718.845 102.5 724.26 106.783 731.107 106.783C736.522 106.783 740.716 104.232 743.165 100.159H756.349C753.282 110.957 743.572 117.886 730.802 117.886C715.371 117.886 704.542 107.088 704.542 91.7016C704.542 76.3157 715.989 65.2053 731.115 65.2053C746.342 65.2053 757.375 76.4171 757.375 91.7016C757.375 92.8251 757.273 93.9408 757.069 94.9629H717.922Z" fill="white"/>
-  <path d="M839.261 91.6002C839.261 83.3455 833.024 77.0257 825.364 77.0257C817.703 77.0257 811.467 83.1426 811.467 91.6002C811.467 99.9563 817.703 106.175 825.364 106.175C833.024 106.167 839.261 99.8549 839.261 91.6002ZM798.791 66.126H811.358V72.7501C815.137 68.061 820.755 65.2132 827.5 65.2132C841.295 65.2132 852.125 76.5263 852.125 91.4987C852.125 106.479 841.092 117.894 827.195 117.894C820.755 117.894 815.442 115.342 811.764 111.067V138.039H798.791V66.126Z" fill="white"/>
-  <path d="M897.422 91.6002C897.422 83.0412 891.193 77.0257 883.525 77.0257C875.865 77.0257 869.628 83.1426 869.628 91.6002C869.628 99.9563 875.865 106.175 883.525 106.175C891.193 106.167 897.422 100.159 897.422 91.6002ZM897.531 117.074V110.45C893.752 115.038 888.126 117.886 881.389 117.886C867.391 117.886 856.765 106.674 856.765 91.6002C856.765 76.62 867.797 65.2132 881.694 65.2132C888.236 65.2132 893.752 68.0688 897.531 72.5472V66.126H910.098V117.074H897.531Z" fill="white"/>
-  <path d="M776.138 71.1194C776.138 71.1194 779.339 65.2132 787.171 65.2132C790.52 65.2132 792.688 66.3601 792.688 66.3601V79.3585C792.688 79.3585 787.962 76.4483 783.619 77.0335C779.276 77.6186 776.53 81.5978 776.545 86.9189V117.082H763.572V66.1338H776.138V71.1194Z" fill="white"/>
-  <path d="M968.792 66.126L937.062 137.875H923.682L936.139 110.138L915.075 66.126H930.239L942.563 95.5325L955.208 66.126H968.792Z" fill="white"/>
-  <path d="M1070.08 64.9713L1038.62 46.8702C1029.39 41.5569 1017.84 48.1966 1017.84 58.831V60.6879C1017.84 62.3654 1018.74 63.918 1020.2 64.7528L1026.14 68.1702C1027.88 69.1767 1030.06 67.9205 1030.06 65.9154V61.2965C1030.06 58.987 1032.56 57.5436 1034.57 58.6983L1061.81 74.3807C1063.82 75.5355 1063.82 78.4223 1061.81 79.5692L1034.57 95.2516C1032.56 96.4063 1030.06 94.9629 1030.06 92.6535V90.1958C1030.06 79.5614 1018.52 72.9139 1009.28 78.235L977.821 96.3361C968.588 101.649 968.588 114.944 977.821 120.258L1009.28 138.359C1018.51 143.672 1030.06 137.032 1030.06 126.398V124.541C1030.06 122.864 1029.16 121.319 1027.7 120.476L1021.77 117.051C1020.02 116.044 1017.84 117.301 1017.84 119.306V123.925C1017.84 126.234 1015.34 127.678 1013.34 126.523L986.092 110.84C984.089 109.686 984.089 106.799 986.092 105.644L1013.34 89.9617C1015.34 88.807 1017.84 90.2504 1017.84 92.5598V95.0175C1017.84 105.652 1029.39 112.299 1038.62 106.978L1070.08 88.8772C1079.31 83.5795 1079.31 70.2846 1070.08 64.9713Z" fill="white"/>
-  <path d="M132.316 109.478V93.791H142.963C151.846 93.791 158.587 88.562 158.587 79.994C158.587 71.552 152.602 66.638 143.278 66.638H124V109.478H132.316ZM132.316 86.987V73.631H142.837C147.184 73.631 150.145 76.403 150.145 80.057C150.145 83.963 147.184 86.987 142.774 86.987H132.316Z" fill="white"/>
-  <path d="M170.68 109.478V66.638H162.742V109.478H170.68Z" fill="white"/>
-  <path d="M191.933 102.737C187.208 102.737 183.365 98.894 183.365 93.728C183.365 88.499 187.208 84.719 191.933 84.719C196.658 84.719 200.501 88.436 200.501 93.728C200.501 99.02 196.658 102.737 191.933 102.737ZM208.313 109.478V77.978H200.564V81.947C198.233 79.175 194.831 77.411 190.799 77.411C182.231 77.411 175.427 84.467 175.427 93.728C175.427 103.052 181.979 109.982 190.61 109.982C194.768 109.982 198.233 108.218 200.564 105.383V109.478H208.313Z" fill="white"/>
-  <path d="M236.894 97.76C235.508 100.784 232.61 102.674 229.271 102.674C224.609 102.674 220.703 98.831 220.703 93.728C220.703 88.625 224.609 84.782 229.271 84.782C232.484 84.782 235.382 86.672 236.831 89.507H245.021C243.257 82.388 237.083 77.411 229.208 77.411C219.884 77.411 212.765 84.467 212.765 93.728C212.765 102.989 219.758 109.982 229.082 109.982C237.146 109.982 243.383 105.005 245.084 97.76H236.894Z" fill="white"/>
-  <path d="M279.9 95.807C280.026 95.177 280.089 94.484 280.089 93.791C280.089 84.341 273.285 77.411 263.898 77.411C254.574 77.411 247.518 84.278 247.518 93.791C247.518 103.304 254.196 109.982 263.709 109.982C271.584 109.982 277.569 105.698 279.459 99.02H271.332C269.82 101.54 267.237 103.115 263.898 103.115C259.677 103.115 256.338 100.469 255.771 95.807H279.9ZM255.834 90.893C256.653 86.735 259.866 84.341 263.772 84.341C267.678 84.341 271.017 86.798 271.584 90.893H255.834Z" fill="white"/>
-  <path d="M326.965 93.728C326.965 84.467 319.783 77.411 310.459 77.411C301.072 77.411 293.953 84.467 293.953 93.728C293.953 102.989 301.072 109.982 310.459 109.982C319.783 109.982 326.965 102.989 326.965 93.728ZM310.459 102.674C305.797 102.674 301.891 98.831 301.891 93.728C301.891 88.625 305.797 84.782 310.459 84.782C315.121 84.782 318.964 88.625 318.964 93.728C318.964 98.831 315.121 102.674 310.459 102.674Z" fill="white"/>
-  <path d="M339.453 109.478V90.83C339.453 87.617 341.028 85.097 344.052 85.097C345.123 85.097 346.005 85.412 346.761 85.853L350.667 78.923C349.596 78.041 348.021 77.411 346.005 77.411C342.729 77.411 340.461 79.112 339.201 81.065V77.978H331.452V109.478H339.453Z" fill="white"/>
-  <path d="M366.554 102.737C361.829 102.737 357.986 98.831 357.986 93.728C357.986 88.625 361.829 84.719 366.554 84.719C371.279 84.719 375.122 88.499 375.122 93.728C375.122 98.894 371.279 102.737 366.554 102.737ZM382.934 109.478V66.638H374.933V81.695C372.602 79.049 369.389 77.411 365.483 77.411C356.915 77.411 350.048 84.467 350.048 93.728C350.048 102.989 356.789 109.982 365.294 109.982C369.452 109.982 372.854 108.155 375.185 105.32V109.478H382.934Z" fill="white"/>
-  <path d="M419.767 95.807C419.893 95.177 419.956 94.484 419.956 93.791C419.956 84.341 413.152 77.411 403.765 77.411C394.441 77.411 387.385 84.278 387.385 93.791C387.385 103.304 394.063 109.982 403.576 109.982C411.451 109.982 417.436 105.698 419.326 99.02H411.199C409.687 101.54 407.104 103.115 403.765 103.115C399.544 103.115 396.205 100.469 395.638 95.807H419.767ZM395.701 90.893C396.52 86.735 399.733 84.341 403.639 84.341C407.545 84.341 410.884 86.798 411.451 90.893H395.701Z" fill="white"/>
-  <path d="M432.454 109.478V90.83C432.454 87.617 434.029 85.097 437.053 85.097C438.124 85.097 439.006 85.412 439.762 85.853L443.668 78.923C442.597 78.041 441.022 77.411 439.006 77.411C435.73 77.411 433.462 79.112 432.202 81.065V77.978H424.453V109.478H432.454Z" fill="white"/>
-  <path d="M491.573 109.478L501.401 77.978H493.337L487.226 98.579H487.163L480.8 77.978H475.193L468.515 98.39H468.389L462.53 77.978H454.277L463.601 109.478H470.846L477.587 89.948H477.776L484.265 109.478H491.573Z" fill="white"/>
-  <path d="M509 74.072C511.457 74.072 513.536 71.993 513.536 69.536C513.536 67.142 511.457 65 509 65C506.543 65 504.464 67.142 504.464 69.536C504.464 71.993 506.543 74.072 509 74.072ZM512.906 109.478V77.978H504.968V109.478H512.906Z" fill="white"/>
-  <path d="M533.784 109.478V102.485H532.21C529.501 102.485 528.745 101.477 528.745 98.831V84.845H533.721V77.978H528.745V70.292H520.806V77.978H516.711V84.845H520.806V100.28C520.806 106.706 523.263 109.478 529.689 109.478H533.784Z" fill="white"/>
-  <path d="M565.143 109.478V88.436C565.143 81.695 561.174 77.411 554.433 77.411C549.96 77.411 547.188 79.679 545.928 81.569V66.638H537.927V109.478H545.928V89.822C545.928 86.609 548.448 84.53 551.724 84.53C554.874 84.53 557.205 86.42 557.205 89.885V109.478H565.143Z" fill="white"/>
-  </svg>
-  """
-  )
-
-  static let afterpayPlaceOrderBlackOnMint = SVG(
-  size: CGSize(width: 1200, height: 180),
-  minimumWidth: 256,
-  svgString: """
-  <svg width="1200" height="180" viewBox="0 0 1200 180" fill="none" xmlns="http://www.w3.org/2000/svg">
-  <rect width="1200" height="180" rx="30" fill="#B2FCE4"/>
-  <path d="M623.657 91.6002C623.657 83.0412 617.421 77.0257 609.761 77.0257C602.1 77.0257 595.864 83.1426 595.864 91.6002C595.864 99.9563 602.1 106.175 609.761 106.175C617.429 106.167 623.657 100.159 623.657 91.6002ZM623.767 117.074V110.45C619.988 115.038 614.362 117.886 607.624 117.886C593.626 117.886 583 106.674 583 91.6002C583 76.62 594.033 65.2132 607.937 65.2132C614.479 65.2132 619.995 68.0688 623.775 72.5472V66.126H636.341V117.074H623.767Z" fill="black"/>
-  <path d="M697.382 105.761C692.985 105.761 691.764 104.13 691.764 99.8549V77.2363H699.839V66.126H691.764V53.6971H678.884V66.126H662.296V63.0285C662.296 58.7451 663.931 57.1145 668.431 57.1145H671.255V47.2291H665.058C654.432 47.2291 649.424 50.6933 649.424 61.2886V66.1182H642.272V77.2285H649.424V117.067H662.304V77.2285H678.892V102.196C678.892 112.588 682.875 117.074 693.298 117.074H699.941V105.761H697.382Z" fill="black"/>
-  <path d="M743.572 87.0125C742.656 80.3884 737.234 76.4171 730.903 76.4171C724.565 76.4171 719.354 80.287 718.032 87.0125H743.572ZM717.922 94.9629C718.845 102.5 724.26 106.783 731.107 106.783C736.522 106.783 740.716 104.232 743.165 100.159H756.349C753.282 110.957 743.572 117.886 730.802 117.886C715.371 117.886 704.542 107.088 704.542 91.7016C704.542 76.3157 715.989 65.2053 731.115 65.2053C746.342 65.2053 757.375 76.4171 757.375 91.7016C757.375 92.8251 757.273 93.9408 757.069 94.9629H717.922Z" fill="black"/>
-  <path d="M839.261 91.6002C839.261 83.3455 833.024 77.0257 825.364 77.0257C817.703 77.0257 811.467 83.1426 811.467 91.6002C811.467 99.9563 817.703 106.175 825.364 106.175C833.024 106.167 839.261 99.8549 839.261 91.6002ZM798.791 66.126H811.358V72.7501C815.137 68.061 820.755 65.2132 827.5 65.2132C841.295 65.2132 852.125 76.5263 852.125 91.4987C852.125 106.479 841.092 117.894 827.195 117.894C820.755 117.894 815.442 115.342 811.764 111.067V138.039H798.791V66.126Z" fill="black"/>
-  <path d="M897.422 91.6002C897.422 83.0412 891.193 77.0257 883.525 77.0257C875.865 77.0257 869.628 83.1426 869.628 91.6002C869.628 99.9563 875.865 106.175 883.525 106.175C891.193 106.167 897.422 100.159 897.422 91.6002ZM897.531 117.074V110.45C893.752 115.038 888.126 117.886 881.389 117.886C867.391 117.886 856.765 106.674 856.765 91.6002C856.765 76.62 867.797 65.2132 881.694 65.2132C888.236 65.2132 893.752 68.0688 897.531 72.5472V66.126H910.098V117.074H897.531Z" fill="black"/>
-  <path d="M776.138 71.1194C776.138 71.1194 779.339 65.2132 787.171 65.2132C790.52 65.2132 792.688 66.3601 792.688 66.3601V79.3585C792.688 79.3585 787.962 76.4483 783.619 77.0335C779.276 77.6186 776.53 81.5978 776.545 86.9189V117.082H763.572V66.1338H776.138V71.1194Z" fill="black"/>
-  <path d="M968.792 66.126L937.062 137.875H923.682L936.139 110.138L915.075 66.126H930.239L942.563 95.5325L955.208 66.126H968.792Z" fill="black"/>
-  <path d="M1070.08 64.9713L1038.62 46.8702C1029.39 41.5569 1017.84 48.1966 1017.84 58.831V60.6879C1017.84 62.3654 1018.74 63.918 1020.2 64.7528L1026.14 68.1702C1027.88 69.1767 1030.06 67.9205 1030.06 65.9154V61.2965C1030.06 58.987 1032.56 57.5436 1034.57 58.6983L1061.81 74.3807C1063.82 75.5355 1063.82 78.4223 1061.81 79.5692L1034.57 95.2516C1032.56 96.4063 1030.06 94.9629 1030.06 92.6535V90.1958C1030.06 79.5614 1018.52 72.9139 1009.28 78.235L977.821 96.3361C968.588 101.649 968.588 114.944 977.821 120.258L1009.28 138.359C1018.51 143.672 1030.06 137.032 1030.06 126.398V124.541C1030.06 122.864 1029.16 121.319 1027.7 120.476L1021.77 117.051C1020.02 116.044 1017.84 117.301 1017.84 119.306V123.925C1017.84 126.234 1015.34 127.678 1013.34 126.523L986.092 110.84C984.089 109.686 984.089 106.799 986.092 105.644L1013.34 89.9617C1015.34 88.807 1017.84 90.2504 1017.84 92.5598V95.0175C1017.84 105.652 1029.39 112.299 1038.62 106.978L1070.08 88.8772C1079.31 83.5795 1079.31 70.2846 1070.08 64.9713Z" fill="black"/>
-  <path d="M132.316 109.478V93.791H142.963C151.846 93.791 158.587 88.562 158.587 79.994C158.587 71.552 152.602 66.638 143.278 66.638H124V109.478H132.316ZM132.316 86.987V73.631H142.837C147.184 73.631 150.145 76.403 150.145 80.057C150.145 83.963 147.184 86.987 142.774 86.987H132.316Z" fill="black"/>
-  <path d="M170.68 109.478V66.638H162.742V109.478H170.68Z" fill="black"/>
-  <path d="M191.933 102.737C187.208 102.737 183.365 98.894 183.365 93.728C183.365 88.499 187.208 84.719 191.933 84.719C196.658 84.719 200.501 88.436 200.501 93.728C200.501 99.02 196.658 102.737 191.933 102.737ZM208.313 109.478V77.978H200.564V81.947C198.233 79.175 194.831 77.411 190.799 77.411C182.231 77.411 175.427 84.467 175.427 93.728C175.427 103.052 181.979 109.982 190.61 109.982C194.768 109.982 198.233 108.218 200.564 105.383V109.478H208.313Z" fill="black"/>
-  <path d="M236.894 97.76C235.508 100.784 232.61 102.674 229.271 102.674C224.609 102.674 220.703 98.831 220.703 93.728C220.703 88.625 224.609 84.782 229.271 84.782C232.484 84.782 235.382 86.672 236.831 89.507H245.021C243.257 82.388 237.083 77.411 229.208 77.411C219.884 77.411 212.765 84.467 212.765 93.728C212.765 102.989 219.758 109.982 229.082 109.982C237.146 109.982 243.383 105.005 245.084 97.76H236.894Z" fill="black"/>
-  <path d="M279.9 95.807C280.026 95.177 280.089 94.484 280.089 93.791C280.089 84.341 273.285 77.411 263.898 77.411C254.574 77.411 247.518 84.278 247.518 93.791C247.518 103.304 254.196 109.982 263.709 109.982C271.584 109.982 277.569 105.698 279.459 99.02H271.332C269.82 101.54 267.237 103.115 263.898 103.115C259.677 103.115 256.338 100.469 255.771 95.807H279.9ZM255.834 90.893C256.653 86.735 259.866 84.341 263.772 84.341C267.678 84.341 271.017 86.798 271.584 90.893H255.834Z" fill="black"/>
-  <path d="M326.965 93.728C326.965 84.467 319.783 77.411 310.459 77.411C301.072 77.411 293.953 84.467 293.953 93.728C293.953 102.989 301.072 109.982 310.459 109.982C319.783 109.982 326.965 102.989 326.965 93.728ZM310.459 102.674C305.797 102.674 301.891 98.831 301.891 93.728C301.891 88.625 305.797 84.782 310.459 84.782C315.121 84.782 318.964 88.625 318.964 93.728C318.964 98.831 315.121 102.674 310.459 102.674Z" fill="black"/>
-  <path d="M339.453 109.478V90.83C339.453 87.617 341.028 85.097 344.052 85.097C345.123 85.097 346.005 85.412 346.761 85.853L350.667 78.923C349.596 78.041 348.021 77.411 346.005 77.411C342.729 77.411 340.461 79.112 339.201 81.065V77.978H331.452V109.478H339.453Z" fill="black"/>
-  <path d="M366.554 102.737C361.829 102.737 357.986 98.831 357.986 93.728C357.986 88.625 361.829 84.719 366.554 84.719C371.279 84.719 375.122 88.499 375.122 93.728C375.122 98.894 371.279 102.737 366.554 102.737ZM382.934 109.478V66.638H374.933V81.695C372.602 79.049 369.389 77.411 365.483 77.411C356.915 77.411 350.048 84.467 350.048 93.728C350.048 102.989 356.789 109.982 365.294 109.982C369.452 109.982 372.854 108.155 375.185 105.32V109.478H382.934Z" fill="black"/>
-  <path d="M419.767 95.807C419.893 95.177 419.956 94.484 419.956 93.791C419.956 84.341 413.152 77.411 403.765 77.411C394.441 77.411 387.385 84.278 387.385 93.791C387.385 103.304 394.063 109.982 403.576 109.982C411.451 109.982 417.436 105.698 419.326 99.02H411.199C409.687 101.54 407.104 103.115 403.765 103.115C399.544 103.115 396.205 100.469 395.638 95.807H419.767ZM395.701 90.893C396.52 86.735 399.733 84.341 403.639 84.341C407.545 84.341 410.884 86.798 411.451 90.893H395.701Z" fill="black"/>
-  <path d="M432.454 109.478V90.83C432.454 87.617 434.029 85.097 437.053 85.097C438.124 85.097 439.006 85.412 439.762 85.853L443.668 78.923C442.597 78.041 441.022 77.411 439.006 77.411C435.73 77.411 433.462 79.112 432.202 81.065V77.978H424.453V109.478H432.454Z" fill="black"/>
-  <path d="M491.573 109.478L501.401 77.978H493.337L487.226 98.579H487.163L480.8 77.978H475.193L468.515 98.39H468.389L462.53 77.978H454.277L463.601 109.478H470.846L477.587 89.948H477.776L484.265 109.478H491.573Z" fill="black"/>
-  <path d="M509 74.072C511.457 74.072 513.536 71.993 513.536 69.536C513.536 67.142 511.457 65 509 65C506.543 65 504.464 67.142 504.464 69.536C504.464 71.993 506.543 74.072 509 74.072ZM512.906 109.478V77.978H504.968V109.478H512.906Z" fill="black"/>
-  <path d="M533.784 109.478V102.485H532.21C529.501 102.485 528.745 101.477 528.745 98.831V84.845H533.721V77.978H528.745V70.292H520.806V77.978H516.711V84.845H520.806V100.28C520.806 106.706 523.263 109.478 529.689 109.478H533.784Z" fill="black"/>
-  <path d="M565.143 109.478V88.436C565.143 81.695 561.174 77.411 554.433 77.411C549.96 77.411 547.188 79.679 545.928 81.569V66.638H537.927V109.478H545.928V89.822C545.928 86.609 548.448 84.53 551.724 84.53C554.874 84.53 557.205 86.42 557.205 89.885V109.478H565.143Z" fill="black"/>
-  </svg>
-  """
-  )
-
-  static let clearpayBuyNowWhiteOnBlack = SVG(
-  size: CGSize(width: 1200, height: 180),
-  minimumWidth: 256,
-  svgString: """
-  <svg width="1200" height="180" viewBox="0 0 1200 180" fill="none" xmlns="http://www.w3.org/2000/svg">
-  <rect width="1200" height="180" rx="30" fill="black"/>
-  <path d="M930.292 67.1079L896.525 139.252H882.536L895.151 112.3L875.333 67.1079H889.697L902.438 97.3798L916.344 67.1079H930.292Z" fill="white"/>
-  <path d="M801.803 92.723C801.803 84.4435 795.766 78.0614 788.355 78.0614C780.944 78.0614 774.907 84.2279 774.907 92.723C774.907 101.132 780.944 107.385 788.355 107.385C795.766 107.385 801.803 101.003 801.803 92.723ZM762.624 139.252V67.1084H774.782V73.7492C778.446 69.0489 783.9 66.1597 790.395 66.1597C803.76 66.1597 814.252 77.544 814.252 92.5937C814.252 107.643 803.552 119.114 790.104 119.114C783.858 119.114 778.737 116.57 775.156 112.257V139.209H762.624V139.252Z" fill="white"/>
-  <path d="M858.135 92.7236C858.135 84.0991 852.098 78.062 844.687 78.062C837.275 78.062 831.238 84.2285 831.238 92.7236C831.238 101.132 837.275 107.385 844.687 107.385C852.098 107.385 858.135 101.305 858.135 92.7236ZM858.218 118.338V111.697C854.554 116.312 849.1 119.158 842.605 119.158C829.032 119.158 818.748 107.903 818.748 92.7236C818.748 77.6739 829.448 66.2034 842.896 66.2034C849.225 66.2034 854.554 69.0926 858.218 73.5773V67.1089H870.376V118.338H858.218Z" fill="white"/>
-  <path d="M740.682 72.1157C740.682 72.1157 743.763 66.1648 751.383 66.1648C754.63 66.1648 756.712 67.3291 756.712 67.3291V80.3952C756.712 80.3952 752.132 77.4629 747.927 78.0666C743.722 78.6703 741.057 82.6375 741.057 87.9847V118.343H728.483V67.1135H740.641V72.1157H740.682Z" fill="white"/>
-  <path d="M1024.72 64.2184L1010.32 55.6802L995.703 47.0126C986.043 41.2773 973.969 48.4787 973.969 60.0355V61.976C973.969 63.0541 974.51 64.0459 975.426 64.5634L982.213 68.5738C984.086 69.6949 986.418 68.2719 986.418 66.0295V61.4154C986.418 59.13 988.791 57.7069 990.706 58.8281L1004.03 66.7626L1017.31 74.654C1019.23 75.7752 1019.23 78.6644 1017.31 79.7856L1004.03 87.6769L990.706 95.6115C988.791 96.7326 986.418 95.3096 986.418 93.0241V90.7386C986.418 79.1819 974.344 71.9373 964.684 77.7157L950.07 86.3833L935.664 94.9215C925.963 100.7 925.963 115.232 935.664 121.011L950.07 129.549L964.684 138.216C974.344 143.952 986.418 136.75 986.418 125.193V123.253C986.418 122.175 985.877 121.183 984.961 120.666L978.174 116.655C976.3 115.534 973.969 116.957 973.969 119.199V123.813C973.969 126.099 971.596 127.522 969.68 126.401L956.357 118.466L943.075 110.575C941.16 109.454 941.16 106.565 943.075 105.443L956.357 97.552L969.68 89.6175C971.596 88.4963 973.969 89.9193 973.969 92.2048V94.4903C973.969 106.047 986.043 113.292 995.703 107.513L1010.32 98.8456L1024.72 90.3074C1034.42 84.4859 1034.42 69.9968 1024.72 64.2184Z" fill="white"/>
-  <path d="M588.712 99.236C586.048 111.008 576.263 119.115 563.606 119.115C548.992 119.115 538 107.731 538 92.6814C538 77.6317 549.158 66.1611 563.814 66.1611C576.18 66.1611 585.839 74.2681 588.629 85.8249H575.764C573.474 81.2108 568.935 78.1491 563.897 78.1491C556.57 78.1491 550.449 84.4019 550.449 92.6814C550.449 100.961 556.57 107.214 563.897 107.214C569.144 107.214 573.682 104.152 575.847 99.236H588.712Z" fill="white"/>
-  <path d="M594.498 118.295V46.0217H606.947V118.295H594.498Z" fill="white"/>
-  <path d="M625.687 96.0871C626.562 103.677 631.808 107.946 638.428 107.946C643.674 107.946 647.712 105.402 650.086 101.305H662.826C659.87 112.172 650.46 119.114 638.095 119.114C623.147 119.114 612.697 108.248 612.697 92.8098C612.697 77.372 623.772 66.2034 638.386 66.2034C653.125 66.2034 663.784 77.4583 663.784 92.8098C663.784 93.931 663.701 95.0522 663.492 96.0871H625.687V96.0871ZM650.502 88.0664C649.628 81.4255 644.382 77.4152 638.261 77.4152C632.141 77.4152 627.103 81.2962 625.812 88.0664H650.502Z" fill="white"/>
-  <path d="M708.249 118.296V111.655C704.586 116.269 699.173 119.115 692.636 119.115C679.104 119.115 668.82 107.86 668.82 92.6814C668.82 77.6317 679.479 66.1611 692.927 66.1611C699.256 66.1611 704.586 69.0072 708.249 73.5351V67.0667H720.407V118.253H708.249V118.296ZM708.125 92.6814C708.125 84.1 702.087 78.0198 694.676 78.0198C687.265 78.0198 681.228 84.1432 681.228 92.6814C681.228 101.09 687.265 107.343 694.676 107.343C702.087 107.343 708.125 101.306 708.125 92.6814Z" fill="white"/>
-  <path d="M190.42 109.478C198.673 109.478 204.406 103.934 204.406 97.004C204.406 92.594 201.571 89.318 198.547 87.554C201.319 85.79 203.902 82.64 203.902 78.419C203.902 71.426 197.98 66.638 190.483 66.638H169V109.478H190.42ZM189.475 73.505C193.192 73.505 195.649 75.584 195.649 78.86C195.649 81.884 193.318 84.467 189.79 84.467H177.253V73.505H189.475ZM189.79 90.956C193.696 90.956 196.027 93.791 196.027 96.626C196.027 99.965 193.192 102.548 189.664 102.548H177.253V90.956H189.79Z" fill="white"/>
-  <path d="M236.063 109.478V77.978H228.125V97.949C228.125 101.099 225.542 103.115 222.266 103.115C219.116 103.115 216.848 101.288 216.848 97.823V77.978H208.847V99.083C208.847 105.824 212.816 109.982 219.62 109.982C224.345 109.982 227.243 107.462 228.314 105.572V109.478H236.063Z" fill="white"/>
-  <path d="M245.648 121.07C254.783 121.07 255.854 117.29 259.697 107.525L271.415 77.978H263.477L255.35 99.776L246.971 77.978H238.781L250.499 107.336C250.877 108.344 250.94 108.848 250.751 109.793C250.121 113.069 248.042 114.14 245.9 114.14H243.317V121.07H245.648Z" fill="white"/>
-  <path d="M312.391 109.478V88.436C312.391 81.695 308.422 77.411 301.681 77.411C296.956 77.411 294.058 80.057 292.924 81.947V77.978H285.175V109.478H293.176V89.822C293.176 86.609 295.696 84.53 298.972 84.53C302.122 84.53 304.453 86.42 304.453 89.885V109.478H312.391Z" fill="white"/>
-  <path d="M349.949 93.728C349.949 84.467 342.767 77.411 333.443 77.411C324.056 77.411 316.937 84.467 316.937 93.728C316.937 102.989 324.056 109.982 333.443 109.982C342.767 109.982 349.949 102.989 349.949 93.728ZM333.443 102.674C328.781 102.674 324.875 98.831 324.875 93.728C324.875 88.625 328.781 84.782 333.443 84.782C338.105 84.782 341.948 88.625 341.948 93.728C341.948 98.831 338.105 102.674 333.443 102.674Z" fill="white"/>
-  <path d="M388.163 109.478L397.991 77.978H389.927L383.816 98.579H383.753L377.39 77.978H371.783L365.105 98.39H364.979L359.12 77.978H350.867L360.191 109.478H367.436L374.177 89.948H374.366L380.855 109.478H388.163Z" fill="white"/>
-  <path d="M447.087 109.478L456.915 77.978H448.851L442.74 98.579H442.677L436.314 77.978H430.707L424.029 98.39H423.903L418.044 77.978H409.791L419.115 109.478H426.36L433.101 89.948H433.29L439.779 109.478H447.087Z" fill="white"/>
-  <path d="M464.515 74.072C466.972 74.072 469.051 71.993 469.051 69.536C469.051 67.142 466.972 65 464.515 65C462.058 65 459.979 67.142 459.979 69.536C459.979 71.993 462.058 74.072 464.515 74.072ZM468.421 109.478V77.978H460.483V109.478H468.421Z" fill="white"/>
-  <path d="M489.299 109.478V102.485H487.724C485.015 102.485 484.259 101.477 484.259 98.831V84.845H489.236V77.978H484.259V70.292H476.321V77.978H472.226V84.845H476.321V100.28C476.321 106.706 478.778 109.478 485.204 109.478H489.299Z" fill="white"/>
-  <path d="M520.658 109.478V88.436C520.658 81.695 516.689 77.411 509.948 77.411C505.475 77.411 502.703 79.679 501.443 81.569V66.638H493.442V109.478H501.443V89.822C501.443 86.609 503.963 84.53 507.239 84.53C510.389 84.53 512.72 86.42 512.72 89.885V109.478H520.658Z" fill="white"/>
-  </svg>
-  """
-  )
-
-  static let clearpayBuyNowBlackOnMint = SVG(
-  size: CGSize(width: 1200, height: 180),
-  minimumWidth: 256,
-  svgString: """
-  <svg width="1200" height="180" viewBox="0 0 1200 180" fill="none" xmlns="http://www.w3.org/2000/svg">
-  <rect width="1200" height="180" rx="30" fill="#B2FCE4"/>
-  <path d="M930.292 67.1079L896.525 139.252H882.536L895.151 112.3L875.333 67.1079H889.697L902.438 97.3798L916.344 67.1079H930.292Z" fill="black"/>
-  <path d="M801.803 92.723C801.803 84.4435 795.766 78.0614 788.355 78.0614C780.944 78.0614 774.907 84.2279 774.907 92.723C774.907 101.132 780.944 107.385 788.355 107.385C795.766 107.385 801.803 101.003 801.803 92.723ZM762.624 139.252V67.1084H774.782V73.7492C778.446 69.0489 783.9 66.1597 790.395 66.1597C803.76 66.1597 814.252 77.544 814.252 92.5937C814.252 107.643 803.552 119.114 790.104 119.114C783.858 119.114 778.737 116.57 775.156 112.257V139.209H762.624V139.252Z" fill="black"/>
-  <path d="M858.135 92.7236C858.135 84.0991 852.098 78.062 844.687 78.062C837.275 78.062 831.238 84.2285 831.238 92.7236C831.238 101.132 837.275 107.385 844.687 107.385C852.098 107.385 858.135 101.305 858.135 92.7236ZM858.218 118.338V111.697C854.554 116.312 849.1 119.158 842.605 119.158C829.032 119.158 818.748 107.903 818.748 92.7236C818.748 77.6739 829.448 66.2034 842.896 66.2034C849.225 66.2034 854.554 69.0926 858.218 73.5773V67.1089H870.376V118.338H858.218Z" fill="black"/>
-  <path d="M740.682 72.1157C740.682 72.1157 743.763 66.1648 751.383 66.1648C754.63 66.1648 756.712 67.3291 756.712 67.3291V80.3952C756.712 80.3952 752.132 77.4629 747.927 78.0666C743.722 78.6703 741.057 82.6375 741.057 87.9847V118.343H728.483V67.1135H740.641V72.1157H740.682Z" fill="black"/>
-  <path d="M1024.72 64.2184L1010.32 55.6802L995.703 47.0126C986.043 41.2773 973.969 48.4787 973.969 60.0355V61.976C973.969 63.0541 974.51 64.0459 975.426 64.5634L982.213 68.5738C984.086 69.6949 986.418 68.2719 986.418 66.0295V61.4154C986.418 59.13 988.791 57.7069 990.706 58.8281L1004.03 66.7626L1017.31 74.654C1019.23 75.7752 1019.23 78.6644 1017.31 79.7856L1004.03 87.6769L990.706 95.6115C988.791 96.7326 986.418 95.3096 986.418 93.0241V90.7386C986.418 79.1819 974.344 71.9373 964.684 77.7157L950.07 86.3833L935.664 94.9215C925.963 100.7 925.963 115.232 935.664 121.011L950.07 129.549L964.684 138.216C974.344 143.952 986.418 136.75 986.418 125.193V123.253C986.418 122.175 985.877 121.183 984.961 120.666L978.174 116.655C976.3 115.534 973.969 116.957 973.969 119.199V123.813C973.969 126.099 971.596 127.522 969.68 126.401L956.357 118.466L943.075 110.575C941.16 109.454 941.16 106.565 943.075 105.443L956.357 97.552L969.68 89.6175C971.596 88.4963 973.969 89.9193 973.969 92.2048V94.4903C973.969 106.047 986.043 113.292 995.703 107.513L1010.32 98.8456L1024.72 90.3074C1034.42 84.4859 1034.42 69.9968 1024.72 64.2184Z" fill="black"/>
-  <path d="M588.712 99.236C586.048 111.008 576.263 119.115 563.606 119.115C548.992 119.115 538 107.731 538 92.6814C538 77.6317 549.158 66.1611 563.814 66.1611C576.18 66.1611 585.839 74.2681 588.629 85.8249H575.764C573.474 81.2108 568.935 78.1491 563.897 78.1491C556.57 78.1491 550.449 84.4019 550.449 92.6814C550.449 100.961 556.57 107.214 563.897 107.214C569.144 107.214 573.682 104.152 575.847 99.236H588.712Z" fill="black"/>
-  <path d="M594.498 118.295V46.0217H606.947V118.295H594.498Z" fill="black"/>
-  <path d="M625.687 96.0871C626.562 103.677 631.808 107.946 638.428 107.946C643.674 107.946 647.712 105.402 650.086 101.305H662.826C659.87 112.172 650.46 119.114 638.095 119.114C623.147 119.114 612.697 108.248 612.697 92.8098C612.697 77.372 623.772 66.2034 638.386 66.2034C653.125 66.2034 663.784 77.4583 663.784 92.8098C663.784 93.931 663.701 95.0522 663.492 96.0871H625.687V96.0871ZM650.502 88.0664C649.628 81.4255 644.382 77.4152 638.261 77.4152C632.141 77.4152 627.103 81.2962 625.812 88.0664H650.502Z" fill="black"/>
-  <path d="M708.249 118.296V111.655C704.586 116.269 699.173 119.115 692.636 119.115C679.104 119.115 668.82 107.86 668.82 92.6814C668.82 77.6317 679.479 66.1611 692.927 66.1611C699.256 66.1611 704.586 69.0072 708.249 73.5351V67.0667H720.407V118.253H708.249V118.296ZM708.125 92.6814C708.125 84.1 702.087 78.0198 694.676 78.0198C687.265 78.0198 681.228 84.1432 681.228 92.6814C681.228 101.09 687.265 107.343 694.676 107.343C702.087 107.343 708.125 101.306 708.125 92.6814Z" fill="black"/>
-  <path d="M190.42 109.478C198.673 109.478 204.406 103.934 204.406 97.004C204.406 92.594 201.571 89.318 198.547 87.554C201.319 85.79 203.902 82.64 203.902 78.419C203.902 71.426 197.98 66.638 190.483 66.638H169V109.478H190.42ZM189.475 73.505C193.192 73.505 195.649 75.584 195.649 78.86C195.649 81.884 193.318 84.467 189.79 84.467H177.253V73.505H189.475ZM189.79 90.956C193.696 90.956 196.027 93.791 196.027 96.626C196.027 99.965 193.192 102.548 189.664 102.548H177.253V90.956H189.79Z" fill="black"/>
-  <path d="M236.063 109.478V77.978H228.125V97.949C228.125 101.099 225.542 103.115 222.266 103.115C219.116 103.115 216.848 101.288 216.848 97.823V77.978H208.847V99.083C208.847 105.824 212.816 109.982 219.62 109.982C224.345 109.982 227.243 107.462 228.314 105.572V109.478H236.063Z" fill="black"/>
-  <path d="M245.648 121.07C254.783 121.07 255.854 117.29 259.697 107.525L271.415 77.978H263.477L255.35 99.776L246.971 77.978H238.781L250.499 107.336C250.877 108.344 250.94 108.848 250.751 109.793C250.121 113.069 248.042 114.14 245.9 114.14H243.317V121.07H245.648Z" fill="black"/>
-  <path d="M312.391 109.478V88.436C312.391 81.695 308.422 77.411 301.681 77.411C296.956 77.411 294.058 80.057 292.924 81.947V77.978H285.175V109.478H293.176V89.822C293.176 86.609 295.696 84.53 298.972 84.53C302.122 84.53 304.453 86.42 304.453 89.885V109.478H312.391Z" fill="black"/>
-  <path d="M349.949 93.728C349.949 84.467 342.767 77.411 333.443 77.411C324.056 77.411 316.937 84.467 316.937 93.728C316.937 102.989 324.056 109.982 333.443 109.982C342.767 109.982 349.949 102.989 349.949 93.728ZM333.443 102.674C328.781 102.674 324.875 98.831 324.875 93.728C324.875 88.625 328.781 84.782 333.443 84.782C338.105 84.782 341.948 88.625 341.948 93.728C341.948 98.831 338.105 102.674 333.443 102.674Z" fill="black"/>
-  <path d="M388.163 109.478L397.991 77.978H389.927L383.816 98.579H383.753L377.39 77.978H371.783L365.105 98.39H364.979L359.12 77.978H350.867L360.191 109.478H367.436L374.177 89.948H374.366L380.855 109.478H388.163Z" fill="black"/>
-  <path d="M447.087 109.478L456.915 77.978H448.851L442.74 98.579H442.677L436.314 77.978H430.707L424.029 98.39H423.903L418.044 77.978H409.791L419.115 109.478H426.36L433.101 89.948H433.29L439.779 109.478H447.087Z" fill="black"/>
-  <path d="M464.515 74.072C466.972 74.072 469.051 71.993 469.051 69.536C469.051 67.142 466.972 65 464.515 65C462.058 65 459.979 67.142 459.979 69.536C459.979 71.993 462.058 74.072 464.515 74.072ZM468.421 109.478V77.978H460.483V109.478H468.421Z" fill="black"/>
-  <path d="M489.299 109.478V102.485H487.724C485.015 102.485 484.259 101.477 484.259 98.831V84.845H489.236V77.978H484.259V70.292H476.321V77.978H472.226V84.845H476.321V100.28C476.321 106.706 478.778 109.478 485.204 109.478H489.299Z" fill="black"/>
-  <path d="M520.658 109.478V88.436C520.658 81.695 516.689 77.411 509.948 77.411C505.475 77.411 502.703 79.679 501.443 81.569V66.638H493.442V109.478H501.443V89.822C501.443 86.609 503.963 84.53 507.239 84.53C510.389 84.53 512.72 86.42 512.72 89.885V109.478H520.658Z" fill="black"/>
-  </svg>
-  """
-  )
-
-  static let clearpayCheckoutWhiteOnBlack = SVG(
-  size: CGSize(width: 1200, height: 180),
-  minimumWidth: 256,
-  svgString: """
-  <svg width="1200" height="180" viewBox="0 0 1200 180" fill="none" xmlns="http://www.w3.org/2000/svg">
-  <rect width="1200" height="180" rx="30" fill="black"/>
-  <path d="M179.351 95.455C176.894 99.424 172.673 101.944 167.696 101.944C160.01 101.944 153.899 95.833 153.899 87.58C153.899 79.39 160.01 73.216 167.696 73.216C172.61 73.216 176.768 75.673 179.162 79.579H188.612C185.399 71.2 177.272 65.656 167.759 65.656C155.348 65.656 145.268 75.043 145.268 87.706C145.268 100.243 155.033 109.504 167.633 109.504C177.461 109.504 185.651 103.897 188.738 95.455H179.351Z" fill="white"/>
-  <path d="M219.81 109V87.958C219.81 81.217 215.841 76.933 209.1 76.933C204.627 76.933 201.855 79.201 200.595 81.091V66.16H192.594V109H200.595V89.344C200.595 86.131 203.115 84.052 206.391 84.052C209.541 84.052 211.872 85.942 211.872 89.407V109H219.81Z" fill="white"/>
-  <path d="M256.738 95.329C256.864 94.699 256.927 94.006 256.927 93.313C256.927 83.863 250.123 76.933 240.736 76.933C231.412 76.933 224.356 83.8 224.356 93.313C224.356 102.826 231.034 109.504 240.547 109.504C248.422 109.504 254.407 105.22 256.297 98.542H248.17C246.658 101.062 244.075 102.637 240.736 102.637C236.515 102.637 233.176 99.991 232.609 95.329H256.738ZM232.672 90.415C233.491 86.257 236.704 83.863 240.61 83.863C244.516 83.863 247.855 86.32 248.422 90.415H232.672Z" fill="white"/>
-  <path d="M283.915 97.282C282.529 100.306 279.631 102.196 276.292 102.196C271.63 102.196 267.724 98.353 267.724 93.25C267.724 88.147 271.63 84.304 276.292 84.304C279.505 84.304 282.403 86.194 283.852 89.029H292.042C290.278 81.91 284.104 76.933 276.229 76.933C266.905 76.933 259.786 83.989 259.786 93.25C259.786 102.511 266.779 109.504 276.103 109.504C284.167 109.504 290.404 104.527 292.105 97.282H283.915Z" fill="white"/>
-  <path d="M326.166 109L313.188 89.47L325.158 77.5H316.149L304.242 89.785H303.99V66.16H296.178V109H303.99V98.164L307.77 94.636L316.842 109H326.166Z" fill="white"/>
-  <path d="M357.937 93.25C357.937 83.989 350.755 76.933 341.431 76.933C332.044 76.933 324.925 83.989 324.925 93.25C324.925 102.511 332.044 109.504 341.431 109.504C350.755 109.504 357.937 102.511 357.937 93.25ZM341.431 102.196C336.769 102.196 332.863 98.353 332.863 93.25C332.863 88.147 336.769 84.304 341.431 84.304C346.093 84.304 349.936 88.147 349.936 93.25C349.936 98.353 346.093 102.196 341.431 102.196Z" fill="white"/>
-  <path d="M389.64 109V77.5H381.702V97.471C381.702 100.621 379.119 102.637 375.843 102.637C372.693 102.637 370.425 100.81 370.425 97.345V77.5H362.424V98.605C362.424 105.346 366.393 109.504 373.197 109.504C377.922 109.504 380.82 106.984 381.891 105.094V109H389.64Z" fill="white"/>
-  <path d="M410.376 109V102.007H408.801C406.092 102.007 405.336 100.999 405.336 98.353V84.367H410.313V77.5H405.336V69.814H397.398V77.5H393.303V84.367H397.398V99.802C397.398 106.228 399.855 109 406.281 109H410.376Z" fill="white"/>
-  <path d="M460.174 109L470.003 77.5H461.939L455.827 98.101H455.764L449.401 77.5H443.794L437.116 97.912H436.99L431.132 77.5H422.878L432.202 109H439.447L446.189 89.47H446.378L452.866 109H460.174Z" fill="white"/>
-  <path d="M477.602 73.594C480.059 73.594 482.138 71.515 482.138 69.058C482.138 66.664 480.059 64.522 477.602 64.522C475.145 64.522 473.066 66.664 473.066 69.058C473.066 71.515 475.145 73.594 477.602 73.594ZM481.508 109V77.5H473.57V109H481.508Z" fill="white"/>
-  <path d="M502.386 109V102.007H500.811C498.102 102.007 497.346 100.999 497.346 98.353V84.367H502.323V77.5H497.346V69.814H489.408V77.5H485.313V84.367H489.408V99.802C489.408 106.228 491.865 109 498.291 109H502.386Z" fill="white"/>
-  <path d="M533.745 109V87.958C533.745 81.217 529.776 76.933 523.035 76.933C518.562 76.933 515.79 79.201 514.53 81.091V66.16H506.529V109H514.53V89.344C514.53 86.131 517.05 84.052 520.326 84.052C523.476 84.052 525.807 85.942 525.807 89.407V109H533.745Z" fill="white"/>
-  <path d="M943.292 67.1079L909.525 139.252H895.536L908.151 112.3L888.333 67.1079H902.697L915.438 97.3798L929.344 67.1079H943.292Z" fill="white"/>
-  <path d="M814.803 92.723C814.803 84.4435 808.766 78.0614 801.355 78.0614C793.944 78.0614 787.907 84.2279 787.907 92.723C787.907 101.132 793.944 107.385 801.355 107.385C808.766 107.385 814.803 101.003 814.803 92.723ZM775.624 139.252V67.1084H787.782V73.7492C791.446 69.0489 796.9 66.1597 803.395 66.1597C816.76 66.1597 827.252 77.544 827.252 92.5937C827.252 107.643 816.552 119.114 803.104 119.114C796.858 119.114 791.737 116.57 788.156 112.257V139.209H775.624V139.252Z" fill="white"/>
-  <path d="M871.135 92.7236C871.135 84.0991 865.098 78.062 857.687 78.062C850.275 78.062 844.238 84.2285 844.238 92.7236C844.238 101.132 850.275 107.385 857.687 107.385C865.098 107.385 871.135 101.305 871.135 92.7236ZM871.218 118.338V111.697C867.554 116.312 862.1 119.158 855.605 119.158C842.032 119.158 831.748 107.903 831.748 92.7236C831.748 77.6739 842.448 66.2034 855.896 66.2034C862.225 66.2034 867.554 69.0926 871.218 73.5773V67.1089H883.376V118.338H871.218Z" fill="white"/>
-  <path d="M753.682 72.1157C753.682 72.1157 756.763 66.1648 764.383 66.1648C767.63 66.1648 769.712 67.3291 769.712 67.3291V80.3952C769.712 80.3952 765.132 77.4629 760.927 78.0666C756.722 78.6703 754.057 82.6375 754.057 87.9847V118.343H741.483V67.1135H753.641V72.1157H753.682Z" fill="white"/>
-  <path d="M1037.72 64.2184L1023.32 55.6802L1008.7 47.0126C999.043 41.2773 986.969 48.4787 986.969 60.0355V61.976C986.969 63.0541 987.51 64.0459 988.426 64.5634L995.213 68.5738C997.086 69.6949 999.418 68.2719 999.418 66.0295V61.4154C999.418 59.13 1001.79 57.7069 1003.71 58.8281L1017.03 66.7626L1030.31 74.654C1032.23 75.7752 1032.23 78.6644 1030.31 79.7856L1017.03 87.6769L1003.71 95.6115C1001.79 96.7326 999.418 95.3096 999.418 93.0241V90.7386C999.418 79.1819 987.344 71.9373 977.684 77.7157L963.07 86.3833L948.664 94.9215C938.963 100.7 938.963 115.232 948.664 121.011L963.07 129.549L977.684 138.216C987.344 143.952 999.418 136.75 999.418 125.193V123.253C999.418 122.175 998.877 121.183 997.961 120.666L991.174 116.655C989.3 115.534 986.969 116.957 986.969 119.199V123.813C986.969 126.099 984.596 127.522 982.68 126.401L969.357 118.466L956.075 110.575C954.16 109.454 954.16 106.565 956.075 105.443L969.357 97.552L982.68 89.6175C984.596 88.4963 986.969 89.9193 986.969 92.2048V94.4903C986.969 106.047 999.043 113.292 1008.7 107.513L1023.32 98.8456L1037.72 90.3074C1047.42 84.4859 1047.42 69.9968 1037.72 64.2184Z" fill="white"/>
-  <path d="M601.712 99.236C599.048 111.008 589.263 119.115 576.606 119.115C561.992 119.115 551 107.731 551 92.6814C551 77.6317 562.158 66.1611 576.814 66.1611C589.18 66.1611 598.839 74.2681 601.629 85.8249H588.764C586.474 81.2108 581.935 78.1491 576.897 78.1491C569.57 78.1491 563.449 84.4019 563.449 92.6814C563.449 100.961 569.57 107.214 576.897 107.214C582.144 107.214 586.682 104.152 588.847 99.236H601.712Z" fill="white"/>
-  <path d="M607.498 118.295V46.0217H619.947V118.295H607.498Z" fill="white"/>
-  <path d="M638.687 96.0871C639.562 103.677 644.808 107.946 651.428 107.946C656.674 107.946 660.712 105.402 663.086 101.305H675.826C672.87 112.172 663.46 119.114 651.095 119.114C636.147 119.114 625.697 108.248 625.697 92.8098C625.697 77.372 636.772 66.2034 651.386 66.2034C666.125 66.2034 676.784 77.4583 676.784 92.8098C676.784 93.931 676.701 95.0522 676.492 96.0871H638.687V96.0871ZM663.502 88.0664C662.628 81.4255 657.382 77.4152 651.261 77.4152C645.141 77.4152 640.103 81.2962 638.812 88.0664H663.502Z" fill="white"/>
-  <path d="M721.249 118.296V111.655C717.586 116.269 712.173 119.115 705.636 119.115C692.104 119.115 681.82 107.86 681.82 92.6814C681.82 77.6317 692.479 66.1611 705.927 66.1611C712.256 66.1611 717.586 69.0072 721.249 73.5351V67.0667H733.407V118.253H721.249V118.296ZM721.125 92.6814C721.125 84.1 715.087 78.0198 707.676 78.0198C700.265 78.0198 694.228 84.1432 694.228 92.6814C694.228 101.09 700.265 107.343 707.676 107.343C715.087 107.343 721.125 101.306 721.125 92.6814Z" fill="white"/>
-  </svg>
-  """
-  )
-
-  static let clearpayCheckoutBlackOnMint = SVG(
-  size: CGSize(width: 1200, height: 180),
-  minimumWidth: 256,
-  svgString: """
-  <svg width="1200" height="180" viewBox="0 0 1200 180" fill="none" xmlns="http://www.w3.org/2000/svg">
-  <rect width="1200" height="180" rx="30" fill="#B2FCE4"/>
-  <path d="M179.351 95.455C176.894 99.424 172.673 101.944 167.696 101.944C160.01 101.944 153.899 95.833 153.899 87.58C153.899 79.39 160.01 73.216 167.696 73.216C172.61 73.216 176.768 75.673 179.162 79.579H188.612C185.399 71.2 177.272 65.656 167.759 65.656C155.348 65.656 145.268 75.043 145.268 87.706C145.268 100.243 155.033 109.504 167.633 109.504C177.461 109.504 185.651 103.897 188.738 95.455H179.351Z" fill="black"/>
-  <path d="M219.81 109V87.958C219.81 81.217 215.841 76.933 209.1 76.933C204.627 76.933 201.855 79.201 200.595 81.091V66.16H192.594V109H200.595V89.344C200.595 86.131 203.115 84.052 206.391 84.052C209.541 84.052 211.872 85.942 211.872 89.407V109H219.81Z" fill="black"/>
-  <path d="M256.738 95.329C256.864 94.699 256.927 94.006 256.927 93.313C256.927 83.863 250.123 76.933 240.736 76.933C231.412 76.933 224.356 83.8 224.356 93.313C224.356 102.826 231.034 109.504 240.547 109.504C248.422 109.504 254.407 105.22 256.297 98.542H248.17C246.658 101.062 244.075 102.637 240.736 102.637C236.515 102.637 233.176 99.991 232.609 95.329H256.738ZM232.672 90.415C233.491 86.257 236.704 83.863 240.61 83.863C244.516 83.863 247.855 86.32 248.422 90.415H232.672Z" fill="black"/>
-  <path d="M283.915 97.282C282.529 100.306 279.631 102.196 276.292 102.196C271.63 102.196 267.724 98.353 267.724 93.25C267.724 88.147 271.63 84.304 276.292 84.304C279.505 84.304 282.403 86.194 283.852 89.029H292.042C290.278 81.91 284.104 76.933 276.229 76.933C266.905 76.933 259.786 83.989 259.786 93.25C259.786 102.511 266.779 109.504 276.103 109.504C284.167 109.504 290.404 104.527 292.105 97.282H283.915Z" fill="black"/>
-  <path d="M326.166 109L313.188 89.47L325.158 77.5H316.149L304.242 89.785H303.99V66.16H296.178V109H303.99V98.164L307.77 94.636L316.842 109H326.166Z" fill="black"/>
-  <path d="M357.937 93.25C357.937 83.989 350.755 76.933 341.431 76.933C332.044 76.933 324.925 83.989 324.925 93.25C324.925 102.511 332.044 109.504 341.431 109.504C350.755 109.504 357.937 102.511 357.937 93.25ZM341.431 102.196C336.769 102.196 332.863 98.353 332.863 93.25C332.863 88.147 336.769 84.304 341.431 84.304C346.093 84.304 349.936 88.147 349.936 93.25C349.936 98.353 346.093 102.196 341.431 102.196Z" fill="black"/>
-  <path d="M389.64 109V77.5H381.702V97.471C381.702 100.621 379.119 102.637 375.843 102.637C372.693 102.637 370.425 100.81 370.425 97.345V77.5H362.424V98.605C362.424 105.346 366.393 109.504 373.197 109.504C377.922 109.504 380.82 106.984 381.891 105.094V109H389.64Z" fill="black"/>
-  <path d="M410.376 109V102.007H408.801C406.092 102.007 405.336 100.999 405.336 98.353V84.367H410.313V77.5H405.336V69.814H397.398V77.5H393.303V84.367H397.398V99.802C397.398 106.228 399.855 109 406.281 109H410.376Z" fill="black"/>
-  <path d="M460.174 109L470.003 77.5H461.939L455.827 98.101H455.764L449.401 77.5H443.794L437.116 97.912H436.99L431.132 77.5H422.878L432.202 109H439.447L446.189 89.47H446.378L452.866 109H460.174Z" fill="black"/>
-  <path d="M477.602 73.594C480.059 73.594 482.138 71.515 482.138 69.058C482.138 66.664 480.059 64.522 477.602 64.522C475.145 64.522 473.066 66.664 473.066 69.058C473.066 71.515 475.145 73.594 477.602 73.594ZM481.508 109V77.5H473.57V109H481.508Z" fill="black"/>
-  <path d="M502.386 109V102.007H500.811C498.102 102.007 497.346 100.999 497.346 98.353V84.367H502.323V77.5H497.346V69.814H489.408V77.5H485.313V84.367H489.408V99.802C489.408 106.228 491.865 109 498.291 109H502.386Z" fill="black"/>
-  <path d="M533.745 109V87.958C533.745 81.217 529.776 76.933 523.035 76.933C518.562 76.933 515.79 79.201 514.53 81.091V66.16H506.529V109H514.53V89.344C514.53 86.131 517.05 84.052 520.326 84.052C523.476 84.052 525.807 85.942 525.807 89.407V109H533.745Z" fill="black"/>
-  <path d="M943.292 67.1079L909.525 139.252H895.536L908.151 112.3L888.333 67.1079H902.697L915.438 97.3798L929.344 67.1079H943.292Z" fill="black"/>
-  <path d="M814.803 92.723C814.803 84.4435 808.766 78.0614 801.355 78.0614C793.944 78.0614 787.907 84.2279 787.907 92.723C787.907 101.132 793.944 107.385 801.355 107.385C808.766 107.385 814.803 101.003 814.803 92.723ZM775.624 139.252V67.1084H787.782V73.7492C791.446 69.0489 796.9 66.1597 803.395 66.1597C816.76 66.1597 827.252 77.544 827.252 92.5937C827.252 107.643 816.552 119.114 803.104 119.114C796.858 119.114 791.737 116.57 788.156 112.257V139.209H775.624V139.252Z" fill="black"/>
-  <path d="M871.135 92.7236C871.135 84.0991 865.098 78.062 857.687 78.062C850.275 78.062 844.238 84.2285 844.238 92.7236C844.238 101.132 850.275 107.385 857.687 107.385C865.098 107.385 871.135 101.305 871.135 92.7236ZM871.218 118.338V111.697C867.554 116.312 862.1 119.158 855.605 119.158C842.032 119.158 831.748 107.903 831.748 92.7236C831.748 77.6739 842.448 66.2034 855.896 66.2034C862.225 66.2034 867.554 69.0926 871.218 73.5773V67.1089H883.376V118.338H871.218Z" fill="black"/>
-  <path d="M753.682 72.1157C753.682 72.1157 756.763 66.1648 764.383 66.1648C767.63 66.1648 769.712 67.3291 769.712 67.3291V80.3952C769.712 80.3952 765.132 77.4629 760.927 78.0666C756.722 78.6703 754.057 82.6375 754.057 87.9847V118.343H741.483V67.1135H753.641V72.1157H753.682Z" fill="black"/>
-  <path d="M1037.72 64.2184L1023.32 55.6802L1008.7 47.0126C999.043 41.2773 986.969 48.4787 986.969 60.0355V61.976C986.969 63.0541 987.51 64.0459 988.426 64.5634L995.213 68.5738C997.086 69.6949 999.418 68.2719 999.418 66.0295V61.4154C999.418 59.13 1001.79 57.7069 1003.71 58.8281L1017.03 66.7626L1030.31 74.654C1032.23 75.7752 1032.23 78.6644 1030.31 79.7856L1017.03 87.6769L1003.71 95.6115C1001.79 96.7326 999.418 95.3096 999.418 93.0241V90.7386C999.418 79.1819 987.344 71.9373 977.684 77.7157L963.07 86.3833L948.664 94.9215C938.963 100.7 938.963 115.232 948.664 121.011L963.07 129.549L977.684 138.216C987.344 143.952 999.418 136.75 999.418 125.193V123.253C999.418 122.175 998.877 121.183 997.961 120.666L991.174 116.655C989.3 115.534 986.969 116.957 986.969 119.199V123.813C986.969 126.099 984.596 127.522 982.68 126.401L969.357 118.466L956.075 110.575C954.16 109.454 954.16 106.565 956.075 105.443L969.357 97.552L982.68 89.6175C984.596 88.4963 986.969 89.9193 986.969 92.2048V94.4903C986.969 106.047 999.043 113.292 1008.7 107.513L1023.32 98.8456L1037.72 90.3074C1047.42 84.4859 1047.42 69.9968 1037.72 64.2184Z" fill="black"/>
-  <path d="M601.712 99.236C599.048 111.008 589.263 119.115 576.606 119.115C561.992 119.115 551 107.731 551 92.6814C551 77.6317 562.158 66.1611 576.814 66.1611C589.18 66.1611 598.839 74.2681 601.629 85.8249H588.764C586.474 81.2108 581.935 78.1491 576.897 78.1491C569.57 78.1491 563.449 84.4019 563.449 92.6814C563.449 100.961 569.57 107.214 576.897 107.214C582.144 107.214 586.682 104.152 588.847 99.236H601.712Z" fill="black"/>
-  <path d="M607.498 118.295V46.0217H619.947V118.295H607.498Z" fill="black"/>
-  <path d="M638.687 96.0871C639.562 103.677 644.808 107.946 651.428 107.946C656.674 107.946 660.712 105.402 663.086 101.305H675.826C672.87 112.172 663.46 119.114 651.095 119.114C636.147 119.114 625.697 108.248 625.697 92.8098C625.697 77.372 636.772 66.2034 651.386 66.2034C666.125 66.2034 676.784 77.4583 676.784 92.8098C676.784 93.931 676.701 95.0522 676.492 96.0871H638.687V96.0871ZM663.502 88.0664C662.628 81.4255 657.382 77.4152 651.261 77.4152C645.141 77.4152 640.103 81.2962 638.812 88.0664H663.502Z" fill="black"/>
-  <path d="M721.249 118.296V111.655C717.586 116.269 712.173 119.115 705.636 119.115C692.104 119.115 681.82 107.86 681.82 92.6814C681.82 77.6317 692.479 66.1611 705.927 66.1611C712.256 66.1611 717.586 69.0072 721.249 73.5351V67.0667H733.407V118.253H721.249V118.296ZM721.125 92.6814C721.125 84.1 715.087 78.0198 707.676 78.0198C700.265 78.0198 694.228 84.1432 694.228 92.6814C694.228 101.09 700.265 107.343 707.676 107.343C715.087 107.343 721.125 101.306 721.125 92.6814Z" fill="black"/>
-  </svg>
-  """
-  )
-
-  static let clearpayPlaceOrderWhiteOnBlack = SVG(
-  size: CGSize(width: 1200, height: 180),
-  minimumWidth: 256,
-  svgString: """
-  <svg width="1200" height="180" viewBox="0 0 1200 180" fill="none" xmlns="http://www.w3.org/2000/svg">
-  <rect width="1200" height="180" rx="30" fill="black"/>
-  <path d="M975.292 67.1079L941.525 139.252H927.536L940.151 112.3L920.333 67.1079H934.697L947.438 97.3798L961.344 67.1079H975.292Z" fill="white"/>
-  <path d="M846.803 92.723C846.803 84.4435 840.766 78.0614 833.355 78.0614C825.944 78.0614 819.907 84.2279 819.907 92.723C819.907 101.132 825.944 107.385 833.355 107.385C840.766 107.385 846.803 101.003 846.803 92.723ZM807.624 139.252V67.1084H819.782V73.7492C823.446 69.0489 828.9 66.1597 835.395 66.1597C848.76 66.1597 859.252 77.544 859.252 92.5937C859.252 107.643 848.552 119.114 835.104 119.114C828.858 119.114 823.737 116.57 820.156 112.257V139.209H807.624V139.252Z" fill="white"/>
-  <path d="M903.135 92.7234C903.135 84.0989 897.098 78.0618 889.687 78.0618C882.275 78.0618 876.238 84.2283 876.238 92.7234C876.238 101.132 882.275 107.385 889.687 107.385C897.098 107.385 903.135 101.305 903.135 92.7234ZM903.218 118.338V111.697C899.554 116.311 894.1 119.157 887.605 119.157C874.032 119.157 863.748 107.902 863.748 92.7234C863.748 77.6737 874.448 66.2031 887.896 66.2031C894.225 66.2031 899.554 69.0923 903.218 73.577V67.1087H915.376V118.338H903.218Z" fill="white"/>
-  <path d="M785.682 72.1159C785.682 72.1159 788.763 66.165 796.383 66.165C799.63 66.165 801.712 67.3293 801.712 67.3293V80.3954C801.712 80.3954 797.132 77.4631 792.927 78.0668C788.722 78.6705 786.057 82.6378 786.057 87.9849V118.343H773.483V67.1137H785.641V72.1159H785.682Z" fill="white"/>
-  <path d="M1069.72 64.2184L1055.32 55.6802L1040.7 47.0126C1031.04 41.2773 1018.97 48.4787 1018.97 60.0355V61.976C1018.97 63.0541 1019.51 64.0459 1020.43 64.5634L1027.21 68.5738C1029.09 69.6949 1031.42 68.2719 1031.42 66.0295V61.4154C1031.42 59.13 1033.79 57.7069 1035.71 58.8281L1049.03 66.7626L1062.31 74.654C1064.23 75.7752 1064.23 78.6644 1062.31 79.7856L1049.03 87.6769L1035.71 95.6115C1033.79 96.7326 1031.42 95.3096 1031.42 93.0241V90.7386C1031.42 79.1819 1019.34 71.9373 1009.68 77.7157L995.07 86.3833L980.664 94.9215C970.963 100.7 970.963 115.232 980.664 121.011L995.07 129.549L1009.68 138.216C1019.34 143.952 1031.42 136.75 1031.42 125.193V123.253C1031.42 122.175 1030.88 121.183 1029.96 120.666L1023.17 116.655C1021.3 115.534 1018.97 116.957 1018.97 119.199V123.813C1018.97 126.099 1016.6 127.522 1014.68 126.401L1001.36 118.466L988.075 110.575C986.16 109.454 986.16 106.565 988.075 105.443L1001.36 97.552L1014.68 89.6175C1016.6 88.4963 1018.97 89.9193 1018.97 92.2048V94.4903C1018.97 106.047 1031.04 113.292 1040.7 107.513L1055.32 98.8456L1069.72 90.3074C1079.42 84.4859 1079.42 69.9968 1069.72 64.2184Z" fill="white"/>
-  <path d="M633.712 99.236C631.048 111.008 621.263 119.115 608.606 119.115C593.992 119.115 583 107.731 583 92.6814C583 77.6317 594.158 66.1611 608.814 66.1611C621.18 66.1611 630.839 74.2681 633.629 85.8249H620.764C618.474 81.2108 613.935 78.1491 608.897 78.1491C601.57 78.1491 595.449 84.4019 595.449 92.6814C595.449 100.961 601.57 107.214 608.897 107.214C614.144 107.214 618.682 104.152 620.847 99.236H633.712Z" fill="white"/>
-  <path d="M639.498 118.295V46.0215H651.947V118.295H639.498Z" fill="white"/>
-  <path d="M670.687 96.0869C671.562 103.676 676.808 107.946 683.428 107.946C688.674 107.946 692.712 105.401 695.086 101.305H707.826C704.87 112.172 695.46 119.114 683.095 119.114C668.147 119.114 657.697 108.247 657.697 92.8096C657.697 77.3718 668.772 66.2031 683.386 66.2031C698.125 66.2031 708.784 77.458 708.784 92.8096C708.784 93.9308 708.701 95.052 708.492 96.0869H670.687V96.0869ZM695.502 88.0661C694.628 81.4253 689.382 77.4149 683.261 77.4149C677.141 77.4149 672.103 81.2959 670.812 88.0661H695.502Z" fill="white"/>
-  <path d="M753.249 118.296V111.655C749.586 116.269 744.173 119.115 737.636 119.115C724.104 119.115 713.82 107.86 713.82 92.6814C713.82 77.6317 724.479 66.1611 737.927 66.1611C744.256 66.1611 749.586 69.0072 753.249 73.5351V67.0667H765.407V118.253H753.249V118.296ZM753.125 92.6814C753.125 84.1 747.087 78.0198 739.676 78.0198C732.265 78.0198 726.228 84.1432 726.228 92.6814C726.228 101.09 732.265 107.343 739.676 107.343C747.087 107.343 753.125 101.306 753.125 92.6814Z" fill="white"/>
-  <path d="M132.316 109.478V93.791H142.963C151.846 93.791 158.587 88.562 158.587 79.994C158.587 71.552 152.602 66.638 143.278 66.638H124V109.478H132.316ZM132.316 86.987V73.631H142.837C147.184 73.631 150.145 76.403 150.145 80.057C150.145 83.963 147.184 86.987 142.774 86.987H132.316Z" fill="white"/>
-  <path d="M170.68 109.478V66.638H162.742V109.478H170.68Z" fill="white"/>
-  <path d="M191.933 102.737C187.208 102.737 183.365 98.894 183.365 93.728C183.365 88.499 187.208 84.719 191.933 84.719C196.658 84.719 200.501 88.436 200.501 93.728C200.501 99.02 196.658 102.737 191.933 102.737ZM208.313 109.478V77.978H200.564V81.947C198.233 79.175 194.831 77.411 190.799 77.411C182.231 77.411 175.427 84.467 175.427 93.728C175.427 103.052 181.979 109.982 190.61 109.982C194.768 109.982 198.233 108.218 200.564 105.383V109.478H208.313Z" fill="white"/>
-  <path d="M236.894 97.76C235.508 100.784 232.61 102.674 229.271 102.674C224.609 102.674 220.703 98.831 220.703 93.728C220.703 88.625 224.609 84.782 229.271 84.782C232.484 84.782 235.382 86.672 236.831 89.507H245.021C243.257 82.388 237.083 77.411 229.208 77.411C219.884 77.411 212.765 84.467 212.765 93.728C212.765 102.989 219.758 109.982 229.082 109.982C237.146 109.982 243.383 105.005 245.084 97.76H236.894Z" fill="white"/>
-  <path d="M279.9 95.807C280.026 95.177 280.089 94.484 280.089 93.791C280.089 84.341 273.285 77.411 263.898 77.411C254.574 77.411 247.518 84.278 247.518 93.791C247.518 103.304 254.196 109.982 263.709 109.982C271.584 109.982 277.569 105.698 279.459 99.02H271.332C269.82 101.54 267.237 103.115 263.898 103.115C259.677 103.115 256.338 100.469 255.771 95.807H279.9ZM255.834 90.893C256.653 86.735 259.866 84.341 263.772 84.341C267.678 84.341 271.017 86.798 271.584 90.893H255.834Z" fill="white"/>
-  <path d="M326.965 93.728C326.965 84.467 319.783 77.411 310.459 77.411C301.072 77.411 293.953 84.467 293.953 93.728C293.953 102.989 301.072 109.982 310.459 109.982C319.783 109.982 326.965 102.989 326.965 93.728ZM310.459 102.674C305.797 102.674 301.891 98.831 301.891 93.728C301.891 88.625 305.797 84.782 310.459 84.782C315.121 84.782 318.964 88.625 318.964 93.728C318.964 98.831 315.121 102.674 310.459 102.674Z" fill="white"/>
-  <path d="M339.453 109.478V90.83C339.453 87.617 341.028 85.097 344.052 85.097C345.123 85.097 346.005 85.412 346.761 85.853L350.667 78.923C349.596 78.041 348.021 77.411 346.005 77.411C342.729 77.411 340.461 79.112 339.201 81.065V77.978H331.452V109.478H339.453Z" fill="white"/>
-  <path d="M366.554 102.737C361.829 102.737 357.986 98.831 357.986 93.728C357.986 88.625 361.829 84.719 366.554 84.719C371.279 84.719 375.122 88.499 375.122 93.728C375.122 98.894 371.279 102.737 366.554 102.737ZM382.934 109.478V66.638H374.933V81.695C372.602 79.049 369.389 77.411 365.483 77.411C356.915 77.411 350.048 84.467 350.048 93.728C350.048 102.989 356.789 109.982 365.294 109.982C369.452 109.982 372.854 108.155 375.185 105.32V109.478H382.934Z" fill="white"/>
-  <path d="M419.767 95.807C419.893 95.177 419.956 94.484 419.956 93.791C419.956 84.341 413.152 77.411 403.765 77.411C394.441 77.411 387.385 84.278 387.385 93.791C387.385 103.304 394.063 109.982 403.576 109.982C411.451 109.982 417.436 105.698 419.326 99.02H411.199C409.687 101.54 407.104 103.115 403.765 103.115C399.544 103.115 396.205 100.469 395.638 95.807H419.767ZM395.701 90.893C396.52 86.735 399.733 84.341 403.639 84.341C407.545 84.341 410.884 86.798 411.451 90.893H395.701Z" fill="white"/>
-  <path d="M432.454 109.478V90.83C432.454 87.617 434.029 85.097 437.053 85.097C438.124 85.097 439.006 85.412 439.762 85.853L443.668 78.923C442.597 78.041 441.022 77.411 439.006 77.411C435.73 77.411 433.462 79.112 432.202 81.065V77.978H424.453V109.478H432.454Z" fill="white"/>
-  <path d="M491.573 109.478L501.401 77.978H493.337L487.226 98.579H487.163L480.8 77.978H475.193L468.515 98.39H468.389L462.53 77.978H454.277L463.601 109.478H470.846L477.587 89.948H477.776L484.265 109.478H491.573Z" fill="white"/>
-  <path d="M509 74.072C511.457 74.072 513.536 71.993 513.536 69.536C513.536 67.142 511.457 65 509 65C506.543 65 504.464 67.142 504.464 69.536C504.464 71.993 506.543 74.072 509 74.072ZM512.906 109.478V77.978H504.968V109.478H512.906Z" fill="white"/>
-  <path d="M533.784 109.478V102.485H532.21C529.501 102.485 528.745 101.477 528.745 98.831V84.845H533.721V77.978H528.745V70.292H520.806V77.978H516.711V84.845H520.806V100.28C520.806 106.706 523.263 109.478 529.689 109.478H533.784Z" fill="white"/>
-  <path d="M565.143 109.478V88.436C565.143 81.695 561.174 77.411 554.433 77.411C549.96 77.411 547.188 79.679 545.928 81.569V66.638H537.927V109.478H545.928V89.822C545.928 86.609 548.448 84.53 551.724 84.53C554.874 84.53 557.205 86.42 557.205 89.885V109.478H565.143Z" fill="white"/>
-  </svg>
-  """
-  )
-
-  static let clearpayPlaceOrderBlackOnMint = SVG(
-  size: CGSize(width: 1200, height: 180),
-  minimumWidth: 256,
-  svgString: """
-  <svg width="1200" height="180" viewBox="0 0 1200 180" fill="none" xmlns="http://www.w3.org/2000/svg">
-  <rect width="1200" height="180" rx="30" fill="#B2FCE4"/>
-  <path d="M975.292 67.1079L941.525 139.252H927.536L940.151 112.3L920.333 67.1079H934.697L947.438 97.3798L961.344 67.1079H975.292Z" fill="black"/>
-  <path d="M846.803 92.723C846.803 84.4435 840.766 78.0614 833.355 78.0614C825.944 78.0614 819.907 84.2279 819.907 92.723C819.907 101.132 825.944 107.385 833.355 107.385C840.766 107.385 846.803 101.003 846.803 92.723ZM807.624 139.252V67.1084H819.782V73.7492C823.446 69.0489 828.9 66.1597 835.395 66.1597C848.76 66.1597 859.252 77.544 859.252 92.5937C859.252 107.643 848.552 119.114 835.104 119.114C828.858 119.114 823.737 116.57 820.156 112.257V139.209H807.624V139.252Z" fill="black"/>
-  <path d="M903.135 92.7236C903.135 84.0991 897.098 78.062 889.687 78.062C882.275 78.062 876.238 84.2285 876.238 92.7236C876.238 101.132 882.275 107.385 889.687 107.385C897.098 107.385 903.135 101.305 903.135 92.7236ZM903.218 118.338V111.697C899.554 116.312 894.1 119.158 887.605 119.158C874.032 119.158 863.748 107.903 863.748 92.7236C863.748 77.6739 874.448 66.2034 887.896 66.2034C894.225 66.2034 899.554 69.0926 903.218 73.5773V67.1089H915.376V118.338H903.218Z" fill="black"/>
-  <path d="M785.682 72.1157C785.682 72.1157 788.763 66.1648 796.383 66.1648C799.63 66.1648 801.712 67.3291 801.712 67.3291V80.3952C801.712 80.3952 797.132 77.4629 792.927 78.0666C788.722 78.6703 786.057 82.6375 786.057 87.9847V118.343H773.483V67.1135H785.641V72.1157H785.682Z" fill="black"/>
-  <path d="M1069.72 64.2184L1055.32 55.6802L1040.7 47.0126C1031.04 41.2773 1018.97 48.4787 1018.97 60.0355V61.976C1018.97 63.0541 1019.51 64.0459 1020.43 64.5634L1027.21 68.5738C1029.09 69.6949 1031.42 68.2719 1031.42 66.0295V61.4154C1031.42 59.13 1033.79 57.7069 1035.71 58.8281L1049.03 66.7626L1062.31 74.654C1064.23 75.7752 1064.23 78.6644 1062.31 79.7856L1049.03 87.6769L1035.71 95.6115C1033.79 96.7326 1031.42 95.3096 1031.42 93.0241V90.7386C1031.42 79.1819 1019.34 71.9373 1009.68 77.7157L995.07 86.3833L980.664 94.9215C970.963 100.7 970.963 115.232 980.664 121.011L995.07 129.549L1009.68 138.216C1019.34 143.952 1031.42 136.75 1031.42 125.193V123.253C1031.42 122.175 1030.88 121.183 1029.96 120.666L1023.17 116.655C1021.3 115.534 1018.97 116.957 1018.97 119.199V123.813C1018.97 126.099 1016.6 127.522 1014.68 126.401L1001.36 118.466L988.075 110.575C986.16 109.454 986.16 106.565 988.075 105.443L1001.36 97.552L1014.68 89.6175C1016.6 88.4963 1018.97 89.9193 1018.97 92.2048V94.4903C1018.97 106.047 1031.04 113.292 1040.7 107.513L1055.32 98.8456L1069.72 90.3074C1079.42 84.4859 1079.42 69.9968 1069.72 64.2184Z" fill="black"/>
-  <path d="M633.712 99.236C631.048 111.008 621.263 119.115 608.606 119.115C593.992 119.115 583 107.731 583 92.6814C583 77.6317 594.158 66.1611 608.814 66.1611C621.18 66.1611 630.839 74.2681 633.629 85.8249H620.764C618.474 81.2108 613.935 78.1491 608.897 78.1491C601.57 78.1491 595.449 84.4019 595.449 92.6814C595.449 100.961 601.57 107.214 608.897 107.214C614.144 107.214 618.682 104.152 620.847 99.236H633.712Z" fill="black"/>
-  <path d="M639.498 118.295V46.0217H651.947V118.295H639.498Z" fill="black"/>
-  <path d="M670.687 96.0871C671.562 103.677 676.808 107.946 683.428 107.946C688.674 107.946 692.712 105.402 695.086 101.305H707.826C704.87 112.172 695.46 119.114 683.095 119.114C668.147 119.114 657.697 108.248 657.697 92.8098C657.697 77.372 668.772 66.2034 683.386 66.2034C698.125 66.2034 708.784 77.4583 708.784 92.8098C708.784 93.931 708.701 95.0522 708.492 96.0871H670.687V96.0871ZM695.502 88.0664C694.628 81.4255 689.382 77.4152 683.261 77.4152C677.141 77.4152 672.103 81.2962 670.812 88.0664H695.502Z" fill="black"/>
-  <path d="M753.249 118.296V111.655C749.586 116.269 744.173 119.115 737.636 119.115C724.104 119.115 713.82 107.86 713.82 92.6814C713.82 77.6317 724.479 66.1611 737.927 66.1611C744.256 66.1611 749.586 69.0072 753.249 73.5351V67.0667H765.407V118.253H753.249V118.296ZM753.125 92.6814C753.125 84.1 747.087 78.0198 739.676 78.0198C732.265 78.0198 726.228 84.1432 726.228 92.6814C726.228 101.09 732.265 107.343 739.676 107.343C747.087 107.343 753.125 101.306 753.125 92.6814Z" fill="black"/>
-  <path d="M132.316 109.478V93.791H142.963C151.846 93.791 158.587 88.562 158.587 79.994C158.587 71.552 152.602 66.638 143.278 66.638H124V109.478H132.316ZM132.316 86.987V73.631H142.837C147.184 73.631 150.145 76.403 150.145 80.057C150.145 83.963 147.184 86.987 142.774 86.987H132.316Z" fill="black"/>
-  <path d="M170.68 109.478V66.638H162.742V109.478H170.68Z" fill="black"/>
-  <path d="M191.933 102.737C187.208 102.737 183.365 98.894 183.365 93.728C183.365 88.499 187.208 84.719 191.933 84.719C196.658 84.719 200.501 88.436 200.501 93.728C200.501 99.02 196.658 102.737 191.933 102.737ZM208.313 109.478V77.978H200.564V81.947C198.233 79.175 194.831 77.411 190.799 77.411C182.231 77.411 175.427 84.467 175.427 93.728C175.427 103.052 181.979 109.982 190.61 109.982C194.768 109.982 198.233 108.218 200.564 105.383V109.478H208.313Z" fill="black"/>
-  <path d="M236.894 97.76C235.508 100.784 232.61 102.674 229.271 102.674C224.609 102.674 220.703 98.831 220.703 93.728C220.703 88.625 224.609 84.782 229.271 84.782C232.484 84.782 235.382 86.672 236.831 89.507H245.021C243.257 82.388 237.083 77.411 229.208 77.411C219.884 77.411 212.765 84.467 212.765 93.728C212.765 102.989 219.758 109.982 229.082 109.982C237.146 109.982 243.383 105.005 245.084 97.76H236.894Z" fill="black"/>
-  <path d="M279.9 95.807C280.026 95.177 280.089 94.484 280.089 93.791C280.089 84.341 273.285 77.411 263.898 77.411C254.574 77.411 247.518 84.278 247.518 93.791C247.518 103.304 254.196 109.982 263.709 109.982C271.584 109.982 277.569 105.698 279.459 99.02H271.332C269.82 101.54 267.237 103.115 263.898 103.115C259.677 103.115 256.338 100.469 255.771 95.807H279.9ZM255.834 90.893C256.653 86.735 259.866 84.341 263.772 84.341C267.678 84.341 271.017 86.798 271.584 90.893H255.834Z" fill="black"/>
-  <path d="M326.965 93.728C326.965 84.467 319.783 77.411 310.459 77.411C301.072 77.411 293.953 84.467 293.953 93.728C293.953 102.989 301.072 109.982 310.459 109.982C319.783 109.982 326.965 102.989 326.965 93.728ZM310.459 102.674C305.797 102.674 301.891 98.831 301.891 93.728C301.891 88.625 305.797 84.782 310.459 84.782C315.121 84.782 318.964 88.625 318.964 93.728C318.964 98.831 315.121 102.674 310.459 102.674Z" fill="black"/>
-  <path d="M339.453 109.478V90.83C339.453 87.617 341.028 85.097 344.052 85.097C345.123 85.097 346.005 85.412 346.761 85.853L350.667 78.923C349.596 78.041 348.021 77.411 346.005 77.411C342.729 77.411 340.461 79.112 339.201 81.065V77.978H331.452V109.478H339.453Z" fill="black"/>
-  <path d="M366.554 102.737C361.829 102.737 357.986 98.831 357.986 93.728C357.986 88.625 361.829 84.719 366.554 84.719C371.279 84.719 375.122 88.499 375.122 93.728C375.122 98.894 371.279 102.737 366.554 102.737ZM382.934 109.478V66.638H374.933V81.695C372.602 79.049 369.389 77.411 365.483 77.411C356.915 77.411 350.048 84.467 350.048 93.728C350.048 102.989 356.789 109.982 365.294 109.982C369.452 109.982 372.854 108.155 375.185 105.32V109.478H382.934Z" fill="black"/>
-  <path d="M419.767 95.807C419.893 95.177 419.956 94.484 419.956 93.791C419.956 84.341 413.152 77.411 403.765 77.411C394.441 77.411 387.385 84.278 387.385 93.791C387.385 103.304 394.063 109.982 403.576 109.982C411.451 109.982 417.436 105.698 419.326 99.02H411.199C409.687 101.54 407.104 103.115 403.765 103.115C399.544 103.115 396.205 100.469 395.638 95.807H419.767ZM395.701 90.893C396.52 86.735 399.733 84.341 403.639 84.341C407.545 84.341 410.884 86.798 411.451 90.893H395.701Z" fill="black"/>
-  <path d="M432.454 109.478V90.83C432.454 87.617 434.029 85.097 437.053 85.097C438.124 85.097 439.006 85.412 439.762 85.853L443.668 78.923C442.597 78.041 441.022 77.411 439.006 77.411C435.73 77.411 433.462 79.112 432.202 81.065V77.978H424.453V109.478H432.454Z" fill="black"/>
-  <path d="M491.573 109.478L501.401 77.978H493.337L487.226 98.579H487.163L480.8 77.978H475.193L468.515 98.39H468.389L462.53 77.978H454.277L463.601 109.478H470.846L477.587 89.948H477.776L484.265 109.478H491.573Z" fill="black"/>
-  <path d="M509 74.072C511.457 74.072 513.536 71.993 513.536 69.536C513.536 67.142 511.457 65 509 65C506.543 65 504.464 67.142 504.464 69.536C504.464 71.993 506.543 74.072 509 74.072ZM512.906 109.478V77.978H504.968V109.478H512.906Z" fill="black"/>
-  <path d="M533.784 109.478V102.485H532.21C529.501 102.485 528.745 101.477 528.745 98.831V84.845H533.721V77.978H528.745V70.292H520.806V77.978H516.711V84.845H520.806V100.28C520.806 106.706 523.263 109.478 529.689 109.478H533.784Z" fill="black"/>
-  <path d="M565.143 109.478V88.436C565.143 81.695 561.174 77.411 554.433 77.411C549.96 77.411 547.188 79.679 545.928 81.569V66.638H537.927V109.478H545.928V89.822C545.928 86.609 548.448 84.53 551.724 84.53C554.874 84.53 557.205 86.42 557.205 89.885V109.478H565.143Z" fill="black"/>
-  </svg>
-  """
-  )
-
+  static func placeOrder(palette: ColorPalette, locale: Locale) -> SVG {
+    switch (palette, locale) {
+    case (.blackOnMint, Locales.greatBritain), (.blackOnWhite, Locales.greatBritain):
+      return .clearpayPlaceOrderBlackOnMint
+    case (.whiteOnBlack, Locales.greatBritain), (.mintOnBlack, Locales.greatBritain):
+      return .clearpayPlaceOrderWhiteOnBlack
+    case (.blackOnMint, _), (.blackOnWhite, _):
+      return .afterpayPlaceOrderBlackOnMint
+    case (.whiteOnBlack, _), (.mintOnBlack, _):
+      return .afterpayPlaceOrderWhiteOnBlack
+    }
+  }
 }
-// swiftlint:enable line_length file_length type_body_length
