+++ Sources/Afterpay/Resources/SVGConfiguration.swift
@@ -23,39 +23,14 @@ struct BadgeConfiguration: SVGConfiguration {
   var colorScheme: ColorScheme = .static(.blackOnMint)
 
   func svg(localizedFor locale: Locale, withTraits traitCollection: UITraitCollection) -> SVG {
-    let svgForPalette: (ColorPalette) -> SVG = { palette in
-      switch (palette, locale) {
-      case (.blackOnMint, Locales.greatBritain):
-        return .clearpayBadgeBlackOnMint
-      case (.mintOnBlack, Locales.greatBritain):
-        return .clearpayBadgeMintOnBlack
-      case (.whiteOnBlack, Locales.greatBritain):
-        return .clearpayBadgeWhiteOnBlack
-      case (.blackOnWhite, Locales.greatBritain):
-        return .clearpayBadgeBlackOnWhite
-      case (.blackOnMint, _):
-        return .afterpayBadgeBlackOnMint
-      case (.mintOnBlack, _):
-        return .afterpayBadgeMintOnBlack
-      case (.whiteOnBlack, _):
-        return .afterpayBadgeWhiteOnBlack
-      case (.blackOnWhite, _):
-        return .afterpayBadgeBlackOnWhite
-      }
+    switch traitCollection.userInterfaceStyle {
+    case .dark:
+      return SVG.badge(palette: colorScheme.darkPalette, locale: locale)
+    case .light, .unspecified:
+      fallthrough
+    @unknown default:
+      return SVG.badge(palette: colorScheme.lightPalette, locale: locale)
     }
-
-    let svg: SVG = {
-      switch traitCollection.userInterfaceStyle {
-      case .dark:
-        return svgForPalette(colorScheme.darkPalette)
-      case .light, .unspecified:
-        fallthrough
-      @unknown default:
-        return svgForPalette(colorScheme.lightPalette)
-      }
-    }()
-
-    return svg
   }
 
   func accessibilityLabel(localizedFor locale: Locale) -> String {
@@ -69,74 +44,21 @@ struct PaymentButtonConfiguration: SVGConfiguration {
   var colorScheme: ColorScheme
   var buttonKind: ButtonKind
 
-  // swiftlint:disable:next cyclomatic_complexity
   private func svg(for palette: ColorPalette, locale: Locale) -> SVG {
-    switch (palette, buttonKind, locale) {
+    let svg: (ColorPalette, Locale) -> SVG
 
-    //BUY NOW WITH AFTERPAY
-    case (.whiteOnBlack, .buyNow, Locales.greatBritain),
-      (.mintOnBlack, .buyNow, Locales.greatBritain):
-      return .clearpayBuyNowWhiteOnBlack
-    case (.blackOnMint, .buyNow, Locales.greatBritain),
-      (.blackOnWhite, .buyNow, Locales.greatBritain):
-      return .clearpayBuyNowBlackOnMint
-
-    case (.mintOnBlack, .buyNow, _),
-      (.whiteOnBlack, .buyNow, _):
-      return .afterpayBuyNowWhiteOnBlack
-    case (.blackOnMint, .buyNow, _),
-      (.blackOnWhite, .buyNow, _):
-      return .afterpayBuyNowBlackOnMint
-
-    //CHECKOUT WITH AFTERPAY
-    case (.whiteOnBlack, .checkout, Locales.greatBritain),
-      (.mintOnBlack, .checkout, Locales.greatBritain):
-      return .clearpayCheckoutWhiteOnBlack
-    case (.blackOnMint, .checkout, Locales.greatBritain),
-      (.blackOnWhite, .checkout, Locales.greatBritain):
-      return .clearpayCheckoutBlackOnMint
-
-    case (.mintOnBlack, .checkout, _),
-      (.whiteOnBlack, .checkout, _):
-      return .afterpayCheckoutWhiteOnBlack
-    case (.blackOnMint, .checkout, _),
-      (.blackOnWhite, .checkout, _):
-      return .afterpayCheckoutBlackOnMint
-
-    //PAY NOW WITH AFTERPAY
-    case (.blackOnMint, .payNow, Locales.greatBritain):
-      return .clearpayPayNowBlackOnMint
-    case (.mintOnBlack, .payNow, Locales.greatBritain):
-      return .clearpayPayNowMintOnBlack
-    case (.whiteOnBlack, .payNow, Locales.greatBritain):
-      return .clearpayPayNowWhiteOnBlack
-    case (.blackOnWhite, .payNow, Locales.greatBritain):
-      return .clearpayPayNowBlackOnWhite
-
-    case (.blackOnMint, .payNow, _):
-      return .afterpayPayNowBlackOnMint
-    case (.mintOnBlack, .payNow, _):
-      return .afterpayPayNowMintOnBlack
-    case (.whiteOnBlack, .payNow, _):
-      return .afterpayPayNowWhiteOnBlack
-    case (.blackOnWhite, .payNow, _):
-      return .afterpayPayNowBlackOnWhite
-
-    //PLACE ORDER WITH AFTERPAY
-    case (.whiteOnBlack, .placeOrder, Locales.greatBritain),
-      (.mintOnBlack, .placeOrder, Locales.greatBritain):
-      return .clearpayPlaceOrderWhiteOnBlack
-    case (.blackOnMint, .placeOrder, Locales.greatBritain),
-      (.blackOnWhite, .placeOrder, Locales.greatBritain):
-      return .clearpayPlaceOrderBlackOnMint
-
-    case (.mintOnBlack, .placeOrder, _),
-      (.whiteOnBlack, .placeOrder, _):
-      return .afterpayPlaceOrderWhiteOnBlack
-    case (.blackOnMint, .placeOrder, _),
-      (.blackOnWhite, .placeOrder, _):
-      return .afterpayPlaceOrderBlackOnMint
+    switch buttonKind {
+    case .buyNow:
+      svg = SVG.buyNow(palette:locale:)
+    case .checkout:
+      svg = SVG.checkout(palette:locale:)
+    case .payNow:
+      svg = SVG.payNow(palette:locale:)
+    case .placeOrder:
+      svg = SVG.placeOrder(palette:locale:)
     }
+
+    return svg(palette, locale)
   }
 
   func svg(localizedFor locale: Locale, withTraits traitCollection: UITraitCollection) -> SVG {
